Cutover logging.yml to read the log_level from env vars.
I don't think the other log settings need to be configurable. If we ever decide to
switch to a new type of logger, let's change the config file.

Note: I had to change the spot that loads the config to treat it as an ERB
(b/c thats the format we're using to load env vars) instead of pure YAML.

TESTING:
- restart, make sure the app starts and that it logs at the level specified.