Update grade export tables to match current export format.
We have a rake task that exports the results of "Export Grades Breakdown"
into a table for Periscope on a course by course basis. This semester we
edited some rubrics and added rows (and perhaps rows too). Also, we added
new assignments, like all the Impact Surveys. So the columns in the .csv
no longer match the columns in the table to hold their values.

This commit updates the tables for Course71ProjectScore and Course73ProjectScore
to match the current Export::GradeDownload format. The tables are just meant to
hold a moment in time snapshot of the data, so it's ok to blow it away and repopulate.

This also adds some validation and throws exceptions if the format changes and
the data can no longer be stored in the table accurately. Hopefully Sentry
will pick these exceptions up and alert us b/c before it was still dumping
data into the tables, but silently and the wrong data.

Note: the tables for old courses: 56, 57, and 58 haven't been touched, so
the rake task will now fail if we try to export into them.

TESTING:
- refreshed the dev DB from production
- ran: bundle exec rake db:migrate
- ran: bundle exec rake data:export:project_scores[73] &>> log/courseScoreExportCron.log
- Saw that the task succeeded.
- Went into the database and selected a column value that was wrong before and
  verified that it matches whats in the grade export now.

RELEASE PLAN:
- Comment out the cronjob on production to stop populating course 56, 57, and 58.
- Update the cronjob for course 71 and 73 to properly redirect stderr to the log
  using "&>> log/courseScoreExportCron.log" instead of "2>&1 >> log/courseScoreExportCron.log"
  so that the exception shows in there too.
- Notify data team to refresh the Periscope cache with the newly created tables.
- Notify the data team that the columns in the table have changed so they can update any
  SQL queries they need to.