Create dev and test only UI tool to reset Magic Fields.
... aka retained data (but we always talk about them as magic fields
with the non-engineers, so stick with that language in the UI.

This change adds a new 'Reset Magic Fields' button on the user's
profile settings page. Clicking this clears out all data in these fields
for that user. This will allow us to more easily test and develop with
users in a fresh state without being required to create a new user each time.

I put this option behind a new dev_only_tools configuration option. We can put
any future tools or features that should not be avaible in the production
environment behind this flag. The reason for this feature being dev only
is because it's irreversible and ripe for abuse by regional folks when troubleshooting.
I'd rather they loop engineers in if they ever need to reset all the work anyone
has done in the course. We shouldn't really ever do that...

TESTING:
- Used the UI button at http://canvasweb/profile/settings and reloaded a module
  I had done work on to see that it was cleared.
- Did the same when masquerading as a user to make sure the correct user id was used
  (as opposed to the admin)
- Ran the feature by URL hacking it to reset a different user than the one logged in
  (note: i'm not exposing this in the UI, but I want to let devs be able to do it if they need to)
- Set the config to false and ensured none of the above worked.
- Removed the config entirely and ensured none of the above worked.

RELEASE PLAN:
- Configure this to true on the staging server.
- We can leave the prod server alone since the absence of the setting is interpretted as false.