Don't build docker-compose dir into the image.
This was causing local changes to docker-compose/config
files to not take effect and require a rebuild of the docker
image. These files are for the dev env only, so don't build them
into the image. Instead just mount the dir as a volume in the dev
env so we can copy over the configs. Note that this is less of a problem
now that configs are checked in and read from the ENV, but we still need
the ability to update the docker-compose/.env file and have it take effect
without needing a rebuild, hence the change.

TESTING:
- rebuilt the image, changed a file in the docker-compose dir,
  (e.g. added a config value in docker-compose/config/beyondz.yml)
  restarted the service and verified that the config change took effect
- Ran the image not using docker-compose.yml
  (e.g. docker run -it <image_name> /bin/bash) and verified that the
  docker-compose dir wasn't there