Check in configs that now use ENV vars.
This one is a biggie. It's in preparation for moving to heroku.
Now the various configs that need to be changed read their config values
from the ENV variables. Note that for those we don't need to change per
env and aren't secret are just hardcoded so its less to manage.

The big change here is that we no longer deploy the configs that are
stored locally on the staging and prod server, but instead
rely on a .env.local file stored in the shared directory on the server.

Note that this commit fixes up sentry and outgoing_mail config load
to ignore values set to nil so they work in the dev env (or other envs)
where the file exists but the key isn't set.

A nice side effect of this change is that in the Docker dev env,
we no longer have to deal with the cognitive load of dealing with
dev only configs in the docker-compose/config dir vs those in the
root app/config dir. It's all the same now and just the .env controls the
behavior. This should allow us to streamline working with the dev env
when we change config values by not needing to rebuild.

TESTING:
- Rebuilt in the dev env and made sure things generally worked.
- A big part of this can't be tested locally and we'll only find out if it
  works as part of the staging/prod deploy. I anticipate followup PRs to
  fix anything I got wrong there (in the config/deploy.rb file)

RELEASE PLAN:
- Snapshot production.
- Create a .env.local file with staging and production values and drop them
  in the shared deploy directory on those servers with the correct permissions.
- Move all existing configs on the staging and prod servers to a backup directory.
- Deploy to staging and fix any issues with Capistrano related to the changes in
  config/deploy.rb
- Sanity test that the site loads login and logout work and generally you can navigate
  the courses. Also sanity test features that depend on the configs.
  E.g. logging, sanitized HTML, files in S3 buckets, outgoing mail, sentry, help tab, etc.
- Deploy to prod and sanity test google analytics which is prod only.