Speed up Portal Docker rebuilds, by A LOT! (#544)
* Cutover beyondz.yml to read from ENV vars.

This is in the docker-compose/config file. When everything is done
we'll check it all into <root>/config and get rid of that file.

TESTING: you can login and the help tab points to http://fake/help/url

* Don't re-run compile assets on every Docker rebuild.

Docker caches each step in the Dockerfile. Since we were adding
the files into the image before running canvas:compile_assets,
any change to anything was causing canvas:compile_assets on a rebuild.
This change explicitly adds the files and dirs needed by canvas:compile_assets,
then runs it, then adds the files for the entire app at the end
so that rebuilds are fast, unless you change a file that requires
canvas:compile_assets to run.

TESTING:
- Rebuilt the image and it worked. Changed a file not required by
  canvas:compile_assets and rebuilt. Compile assets didn't run.