Make it easier to add a gem and create a new Gemfile.lock in the dev env.
Mount Gemfile.d and Gemfile.lock as volumes so that changes inside the container
are reflected outside. On restart, if bundle install needs to run, run it.
With this flow, when you add a gem, just add it to the proper Gemfile.d file
and do a devrs to generate the Gemfile.lock. Everything will show up in a git
diff on the host machine so you can check it in once you verify that it works.

TESTING:
- Added the rails12factor gem to Gemfile.d/heroku.rb and did a devrs.
  The app restarted and stayed up and the Gemfile.lock was updated.