Reduce the docker manifest size when pushing to Heroku.
When pushing code and letting Heroku build the Docker image,
I was getting "State changed from starting to crashed" when it
tried to bring it up, with no logs or nothing. The manifest size is

   latest: digest: sha256:XXXXXXX size: 7655

This thread seems to indicate that could be the issue:
https://github.com/moby/moby/issues/24444

This is a test to see if that's the case. If it works, I need to
see if I can refactor the Docker build to use a multi-stage build
in order to reduce the manifest size instead of maintaining two separate
Dockerfile's