Whitelist redis 'auth' command since Heroku's redis add-on requires a password.
I feel like ther is a better solution but I was having trouble understanding what's
going on here. From what I gather, I think there is the option to setup a Twemproxy
server to act as a proxy to the actual redis servers. But maybe the Twemproxy
stuff is used locally as a way to deal with the pooling and connection retries?
Either way, since we're not doing anything fancy with redis and multiple servers
this should be safe for now. If we get to that point and this fails, they we can
fix it more properly.

Here was the top of the failing stacktrace without this change:
app[web.1]: Original exception: Redis method `auth` is potentially dangerous, and should only be called from console, and only if you fully understand the consequences. If you're sure, retry after running Shackles.activate!(:deploy)
app[web.1]: Trace: /app/lib/canvas/redis.rb:106:in `write'
...