Make avatar initials match display name on People Page
Closes USERS-1

The documentation says that the user avatar initials should
come from the user's full name, so this makes that happen
instead of the sortable name (which was weird because usually
the initials in the avatar are then reversed).

After talking with Matt Meservey, we'll use the display name
here instead of the full name, in order to make the source of
the avatar initials more consistent everywhere. This does mean
that the documentation will need a tweak.

Test plan:
* Have a user without an avatar picture
  - the sortable name, full name, and display name should
    all be different (they usually are)
* Go to Admin -> People
* Find that user
* The avatar icon initials should come from the
  display name, not the sortable name

Change-Id: Ie55457d98a436e7a849845f612a740f3a098d031
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/215892
Reviewed-by: Clay Diffrient <cdiffrient@instructure.com>
Tested-by: Jenkins
QA-Review: Steve Shepherd <sshepherd@instructure.com>
Product-Review: Steve Shepherd <sshepherd@instructure.com>