Another try to get structured JSON logging in STDOUT.
Something was still logging to the file, so explicitly intiliaze
the rails_stdout_logging stuff. Also, go back to just making the
logjam agent use the Rails.logger.level setting for what it writes
as structured data. The previous commit to read from the ENV was
unecessary and less maintainable.