First pass at configuring DB connection pool properly.
We default to 2 worker processes. Each process has it's own connection pool.
When the master process forks the workers, we need to disconnect the
connections in the master and re-establish in the workers. If we were
on Rails 4.1 this would be easier and not duplicative by using database.yml.
Note: with 2 processes and 5 threads per process with a connection pool of
MAX_THREADS per process (aka 5 if not set), we only have DB 10 connections.
If we ever see the connection count go above that, something is wrong and
we're leaking DB connections.

Final note: we need to tune this stuff based on the memory available
(which impact the # of processes) and whether or not we can actually
run a multi-threaded server. We just need to load test
to see if it's multi-threaded.