Fix AR scoping issues in user_search by last_login
Fixes USERS-70
flag=none

The scopes in user_search didn't bring in all the
right columns out of the users table in order to
be able to properly include the last login info
from the corresponding pseudonyms. As a result, when
on the account people page, one would both have an
active search AND sort by the last_login field, the
last_login column would be nil, and all that data
would vanish out of the UI.

So we've got to explicitly bring in all the columns
(via users.*) for it to work. I also added a test in
case someone breaks this later... it's awfully subtle.

Test plan:
* Rerun the steps to reproduce in the ticket
* When limiting the users list via a search, you
  should be able to sort on the "last login" field
  and have that column not disappear and also sort
  properly (both ascending and descending)

Change-Id: Icd31bfab19e4e074f42ac1105904fcb11a0db10a
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/217466
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Tested-by: Jenkins
Reviewed-by: Landon Gilbert-Bland <lbland@instructure.com>
Reviewed-by: Cody Cutrer <cody@instructure.com>
Product-Review: Cody Cutrer <cody@instructure.com>
QA-Review: Steve Shepherd <sshepherd@instructure.com>