The employer information was being exported wrong in the LinkedIn export.
Something must have changed with the way V2 of the LinkedIn API
returns the list of employers / positions or maybe it was just always
broken if someone has mulitple current positions lists.
It was arbitrarily choosing one of the jobs to consider their current one.
This change makes is so that the most recently added job is considered
current if they have multiple.
See: https://app.asana.com/0/1133543009734854/1149776609697206

TESTING:
- chose a Fellow who had multiple jobs listed as current and downloaded
  their raw LinkedIn Data. Loaded that and ran this code against it,
  cross checking what their profile looked like on the website to ensure
  that the correct title and employer were picked from the list of active
  jobs.

Note: there is still a bug with getting someone's company when doing a full
export of the data. It's marked with a TODO and a link to the Sentry error.
Will do a followup commit to fix that.