Fix LinkedIn data export for recent school information.
The educations listed may or may not have dates associated with them,
however, the most recent one should be listed first in the list so
just grab that. I think all the fancy logic there before was to handle
V1 of the API or maybe it was just an oversight that this was an option.