Fix logic to get most recent school in LinkedIn export.
For real this time. Through testing on course 73, 71, and 57
I've realised there are a lot of different possible ways the schools
can be listed. At the end of the day, we want the school listed
at the top of the LinkedIn UI in the web browser, but there is no
way to reliably get that through the API. So the algo returns the
school with the latest date set (considering that a school may have a
start date but no end date). If there are no schools with dates set,
it returns the first item returned in the JSON b/c hopefully that's
most likely to be at the top of the LinkedIn UI.

TESTING:
- Did an export on courses: 57, 71, 73 and spot checked all school names
  and graduation dates to make sure they weren't high school or prior schools
  to Rutgers and San Jose.