Run delayed_jobs in the foreground on a Heroku worker dyno.
If we run it as a daemon, like we do on AWS, the Heroku dyno crashes
immediately and we get no logs. Digging into the guts of the inst-jobs
gem, I discovered that we can run it using "run" instead of "start" as
the command and that will cause it to run in the foreground and tail the logs
to stdout.

See this for details:
vendor/bundle/ruby/2.1.0/gems/inst-jobs-0.11.2/lib/delayed/cli.rb

TESTING:
- This is a rebase / squash of 3 commits I deploy to Heroku until I finally
  got this working. Now the worker dyno stays up and the jobs run.
- Run heroku ps:resize worker=standard-2x the first time the worker is deployed.
- Note: that during testing I configured only 2 workers with max memory of 1GB
  using DELAYED_JOB_WORKERS and DELAYED_JOB_WORKER_MAX_MEMORY env vars.
  We'll want to tune that. I suspect 1GB won't be enough for some jobs since
  the app takes up about 900MB once it's fully started and without accounting
  for the work it has to actually do. With a non-performance dynos, we only get
  1GB of memory and can go up to 2GB before the dyno will crash. Anything between
  1 and 2 GB will just be slow since it will swap to disk. To avoid having to pay
  for a performance dyno and to stay fast, we'll need to optimize the memory that's
  used.