Add rake task to restart dynos.
This is intended to be used from the Scheduler add-on
(or some other add-on) to restart the dyno in the middle of
the night daily so that we're not at the whims of the dyno
manager cycling it.

Once we optimize the app boot and amount of memory it uses
so that we can run multiple dynos instead of a single
performance-l dynoe, we won't really need this anymore.

TESTING:
- After creating an oauth token using
   heroku authorizations:create -d "Portal Admin API token" -s write
 I ran:
   HEROKU_OAUTH_TOKEN=my-token HEROKU_APP_NAME=my-app bundle exec rake heroku:app:restart_dyno[web]
 and verified that the dyno restarted