Remove some unneeded React.memos
Having these here means that react is first going to shallow diff any
props for referential equality (meaning `() => “hi”` is not the same as 
`() => “hi”` since they are different functions in memory even though
they return the same thing) and then if that shallow diff fails, it
is still going to re-render all children. In most cases that means these
are actually making react do MORE work.

So if there are any particular ones of these that we know are preventing
us from having to re-render stuff that is more costly than the extra
prop-diffing is saving us, and we know that is a bottleneck, then lets
comment that we did that on purpose but if not, we can remove these

Test plan:
* the a2 student page should work and look the same but it should
  actually render a little faster than it did before
  
flag=assignments_2_student

Change-Id: I6693799c10e009d41f33fa9d3559175029c76463
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/218664
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Tested-by: Jenkins
Reviewed-by: Steven Burnett <sburnett@instructure.com>
QA-Review: Ryan Shaw <ryan@instructure.com>
Product-Review: Ryan Shaw <ryan@instructure.com>