Add ability to turn on Rack::Lock and change Puma queue_requests.
The hunch is that the Portal is not threadsafe (since it was never
run ina multithreaded web server before). We can't just spawn a
ton of worker processes to make it handle concurrency (leaving it
single threaded) b/c that would take up too much memory on a
Heroku dyno. However, when stress testing the server with 5 worker processes
I was getting Net::ReadTimeout on the TCP socket. I think it's related to the
backlog of requests in the request queue and it dead-locking itself:
https://github.com/puma/puma/issues/612

I'm going to tinker around with allowing threads, but using Rack::Lock
to make it functionally single threads according to this:
https://devcenter.heroku.com/articles/deploying-rails-applications-with-the-puma-web-server#thread-safety

I'm also going to try out setting queue_requests to false to verify if
that's what's up.

Some useful reading:
- https://github.com/teamcapybara/capybara/issues/2227
- https://github.com/puma/puma/blob/master/docs/architecture.md
- https://github.com/puma/puma/issues/876
- https://github.com/puma/puma#known-bugs (though we're on Ruby 2.1, which isn't in the list.)