Fix infinite redirects and redirects to incorrect locations when loading files.
This is a hack and leaves in place a ton of code we don't use related to
configuring canvas to serve files across different servers / shards that this
change could break. In our world, we only have the app hosted on one domain
and don't have to isolate things (files) based on the requesting domain so it's not worth
figuring out the proper way to fix this since we're rewriting the whole codebase
anyway.

Context:
- When loading a page like /courses/57/pages/onboard-to-braven?module_item_id=5322
  in the dev env or on a Heroku server, HTML elements like
  <img src="/courses/1/files/27771/preview" ... > were either redirecting infinitely or
  redirecting to the wrong domain. I think it's actually related to the port used to access
  the app.

TESTING:
- I put in a bunch of logs in the files_controller and loaded a wiki page with the
  HTML similar to above that was failing and instead of trying to redirect to a
  safe domain (like it was doing prior) it just redirected to S3 like its supposed to
  and loaded the inline/preview image.
- When pushed to staging or prod, we need to test that files can be uploaded both on a course
  and when submitting assignments. We also need to check that TAs can access uploaded file
  submissions for Students.