Speed up production builds (heroku was timing out).
There is a 15 min timeout and you have to contact heroku support
if you want to increase it. I set the JS_BUILD_NO_UGLIFY env var
to skip uglifying the JS (which takes a long time):
https://github.com/instructure/canvas-lms/issues/1177

Not sure this is a long term solution, so if we can speed of the build enough
in other ways, roll this back!