Try to fix compile assets build error.
"Undumpable Exception -- #<Psych::SyntaxError: (<unknown>): mapping values are not allowed in this context"

Just commenting out the stuff related to getting Review Apps going. Will
have to revisit that once we actually do it. The syntax seems to trip up something in the
compile assets build that must read this yaml file.