add a prefix to the cache migrations.
inside the dockerCacheLoad and dockerCacheStore methods, a prefix
is automatically added in the method if there is none that is the
name of the build. that caused the calls to try and pull different
images when in different builds. add this prefix will cause
the same image to be pulled across builds.

fixes: CCI-150

flag = none

Test-Plan:
look in the subbuild and make sure its pulling the same cached image

Change-Id: I78852d21e50a14fc7804ccd55aaea74bb51bacb0
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/220199
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Tested-by: Jenkins
Reviewed-by: James Butters <jbutters@instructure.com>
Reviewed-by: Robert Lamb <rlamb@instructure.com>
QA-Review: Rex Fleischer <rfleischer@instructure.com>
Product-Review: Rex Fleischer <rfleischer@instructure.com>