Better short-circuit for Scout APM.
Using the Heroku add-on, the SCOUT_KEY is always there and
if we set SCOUT_MONITOR == false we still initialize the gem.
Don't do that by checking for this disable override ENV var.

NOTE: When the ENV var is set in Heroku to false (as recommended in the
Scout docs), it comes through into Rails as a string which evaluates
to true if we just check ENV['SCOUT_MONITOR']. This could be implemented
better to deal with various types of "false" such as 0, but this code will
go away over the next year and this is unlikely to cause issues b/c I plan
on documenting what configs are available in the readme and will use the
example of just putting in the string "false" into the ENV var to disable
it.