remove Location field from child calendar event edit dialog
When clicking on a calendar event, you are given an edit
dialog with a small selection of items, with the option to
open a new page which has all the edit options. This dialog
contains a field for Location, even though the Location
attrbute cannot be updated on a child event, since it belongs to
the parent. This change removes the field from child events.

Test Plan:
- Create a normal calendar event
- Open the edit modal for that event and see the Location field
- Create a child event (such as a course event that uses different dates
     for each section
- Open the edit modal for one of the child events and see that the
     Location field is not shown

closes KNO-33
flag = none

Change-Id: If9cda305c2a2359983f2565225a336fa1718942a
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/221454
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Matthew Lemon <mlemon@instructure.com>
QA-Review: Matthew Lemon <mlemon@instructure.com>
Product-Review: Ben Nelson <bnelson@instructure.com>