Update submission/grades page similarity icons
flag=new_gradebook_plagiarism_indicator
closes TALLY-246

Test plan:

SETUP
Enable Turnitin for your setup:
  - Set some fake settings for your root account in a console:
    > account = <your root account>
    > account.turnitin_account_id = 1
    > account.turnitin_shared_secret = "fred"
    > account.settings[:enable_turnitin] = true
    > account.save!
  - At this point, account.turnitin_settings should return an array of
    values if the above is set up properly (otherwise it returns nil)
Enable Vericite for your setup:
  - Open the plugin at http://<your canvas>/plugins/vericite
    - Create a configuration for "All Accounts" and un-disable it
    - The rest of the settings don't matter

CREATION
- Have a course using New Gradebook with 5 or more students
- Create several assignments:
  - A Turnitin-based assignment with text entry submisisons
  - A Turnitin-based assignment with attachment submisisons
    - For both the above, enable Turnitin on the assignment edit page
  - A VeriCite-based assignment with text entry submisisons
  - A VeriCite-based assignment with attachment submisisons
    - For both the above, enable VeriCite on the assignment edit page
  - An assignment with text entry submisisons
  - An assignment with attachment submisisons
    - These two assignments will use originality reports, so no need to
      enable either of the above
- For each assignment
  - Have each student submit a text entry or attachment as appropriate
  - Set up plagiarism data in the following states (one per student):
    - "Error" state
    - "Pending" state
    - Scored, 70% similarity
    - Scored, 40% similarity
    - Scored, 10% similarity

ACTUALLY CHECKING THE ICONS

The existing behavior of the icons varies dramatically depending on the
page in question and the state/source of the similarity report. For
instance, on the student grades page, "pending" submissions only show an
icon if the submission uses an originality report (Turnitin and
VeriCite-based submissions do not show an icon). This patchset makes no
attempt to regularize any of the existing logic; instead, for a given
page, you should check that a given submission consistently displays an
icon (or does not) with the New Gradebook Plagiarism Indicator feature
flag enabled or disabled.

With that in mind, check the following pages:
  - Grade summary page
  - Submission details page
    - Beneath the grade input in the upper right for text entry
      submissions
    - In the main submission frame for attachment submissions
  - SpeedGrader submission view (not the right-hand side, just the
    preview frame)

In all cases, in a plagiarism icon is shown (or hidden) with the new
flag OFF, a corresponding new-style icon should be shown (or hidden)
when the flag is ON.

Change-Id: Ic9ada3fe8f1e983429d1baa05e649adb07db396a
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/225155
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
QA-Review: Adrian Packel <apackel@instructure.com>
Product-Review: Jonathan Fenton <jfenton@instructure.com>
Reviewed-by: Spencer Olson <solson@instructure.com>
Reviewed-by: Gary Mei <gmei@instructure.com>