fix message students option in appointment groups
When an instructor used the "Message Students" option shown in the
Calendar for an Appointment Group that has multiple courses associated
to it, Canvas would only send the message to the students who were
enrolled in the first listed course under the "context_codes" parameter
in the Appointment Group's API (Tends to be the course with the oldest
ID). Now we send the messages by adding the course context codes as tags
to the POST request to conversations.

Test Plan
- Create 2 different courses and add a different student to each one
- As the teacher navigate to the calendar and create an appointment
  group associated to both courses
- As the students reserve the appointment group
- As the teacher click on the appointment group and select the "Message
  Students" option at the bottom of the dialog and send the students a
  message
- Login as both students and check that both have recieved the message

fixes KNO-212
flag=none

Change-Id: I55112390d2b9bdb831dcfff63658f27e9d04dbc1
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/223912
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
QA-Review: Ben Nelson <bnelson@instructure.com>
Product-Review: Matthew Lemon <mlemon@instructure.com>
Reviewed-by: Steven Burnett <sburnett@instructure.com>