Change rce's upload placeholder
The previous placeholder was an <img src="data:image/gif...">. When
autosave saved it, it got saved as an <img src="blob:...">, which
cannot be restored. By changing it to a <div>, the problem goes away.
The only side-effect is that if the upload fails and the user wants
to delete the placeholder, it used to take one click of the BS key.
Now they have to delete the "Loading..." text and the <div> will
eventually go with it.

closes LA-475
flag=rce_enhancements

test plan:
  - with notorious not running
  - in the RCE, upload media
  > expect the "Loading..." grey placeholder box
  - refresh the page
  > expect to be prompted with auto-saved content
  - say Yes
  > expect the grey box in the RCE
  - iwth notorious running
  - in the RCE, upload media
  > expect the "Loading..." grey box
  > expect the grey box to go away once the media appears

Change-Id: If16037b4b3310eb8bdc8212bc1ed1d8dc83b657d
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/224433
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Jon Willesen <jonw+gerrit@instructure.com>
Reviewed-by: Carl Kibler <ckibler@instructure.com>
QA-Review: Anju Reddy <areddy@instructure.com>
Product-Review: Ed Schiebel <eschiebel@instructure.com>