Ignore Redis "OOM command not allowed" errors.
A properly configured redis server should recover on it's own from this
and the user impact is that the key appears not to be cached, which
would just slow things down while we're in this state.

Note: see comments in code for how I've worked around the issues
leading to this error in our environment.

TESTING:
- Configured the maxmemory to be small in my dev env and confirmed
  that the OOM errors were being logged to the error report but not
  thrown. Note: login is completely broken in this state.

NOT TESTED:
- I'm 99% positive that these errors will still be sent to Sentry which
  will help with debugging. Look in ./config/initializers/sentry.rb
  and notice how it hooks into all Canvas errors logged and sends them over.
  Also, at this point, "see a blank white page, think redis"
  is instituional knowledge so if it starts happening and we don't get
  Sentry errors, we can figure that part out then.