Fix dbrefresh for real.
The new Heroku based nightly backups don't have DROP blah commands
so it wasn't actually working like it used to. It's faster to just
remove the DB volume which blows away the DB and then pull the
backup and restore from it.

This also cuts over SSO from the old box to the new one.

TESTING:
- Did a refresh and actually got an up to date DB instead of all the
  unique constraint violations that used to happen.