spec: Fix flaky spec announcement_permission_spec.rb
closes: KNO-116

Looks like the alerts are working in Jenkins again. This
was failing because of an alert bug that was fixed.
Unskipping the test and retriggering to make sure it is fine,
then will put it up for review.

Change-Id: Ifc4345e228d20492d9b012243b277ec61764a903
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/226346
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Cody Cutrer <cody@instructure.com>
QA-Review: Cody Cutrer <cody@instructure.com>
Product-Review: Cody Cutrer <cody@instructure.com>