allow teachers to send messages in unpublished courses
We are allowing Teachers, TAs, and Designers to be able to send messages
in unpublished courses. Additionally we are allowing them to also be
able to continue a conversation after a course has concluded. They will
not be able to create new conversations using the concluded course as a
context but they will be able to continue whatever conversation they
were already having. Note that students will still not be able to send
messages or reply to existing conversations after the course has
concluded.

/ ---- ---- \
| Test Plan |
\ ---- ---- /

- Create an unpublished course
- Enroll at least 2 teachers
- As one of the teachers attempt to send a message to the other
- Note that the message should send successfully
- Conclude the course
- As the teacher note that you are no longer able to create new messages
  using the concluded course as context
- As the teacher note that you are able to continue existing
  conversations

fixes KNO-290
flag=none

Change-Id: I8abe105948802c6b6096b652fa7dc058a8c5be77
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/228638
Reviewed-by: Rob Orton <rob@instructure.com>
QA-Review: Rob Orton <rob@instructure.com>
Product-Review: Rob Orton <rob@instructure.com>
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>