Preserve RoleOverrides of granulars when group is unlocked
Refs USERS-246

When operating on a group permission, which doesn't really exist,
you're operating on all the granulars in the group. A little
edge case is that unlocking the group would revert all the
granulars to the default state, because the RoleOverride model
will delete the RoleOverride if a permission is unlocked but has
a nil value for the override.

So we need to retrieve the existing override state of all the
granulars of a group, and if the update operation is :explicit but
no :override value is given, then respect the existing override.
To delete a RoleOverride (return the permission to the default
state), set :explicit to false.

Test plan:
* tests pass

Change-Id: I92e82c0b31ab6063a305d06ef1a6de24c626f424
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/229997
QA-Review: Ahmad Amireh <ahmad@instructure.com>
Product-Review: Ahmad Amireh <ahmad@instructure.com>
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Ahmad Amireh <ahmad@instructure.com>