Modify Canvadocs launch params - add all user_ids from group assignments
Refs DIG-1749

Canvadocs Notifications has a bug where if a document is part of a group
assignment, the notification will only go to the group member with the
lowest (numerically) canvas user id.  This is because Canvas was only
sending that id.  To fix this, we need Canvas to send all user ids for
the group. so we can notify them all.

Test plan:
- Create a group assignment
- Submit a doc for the assignment
- Confirm that the docviewer launch params include an array of user ids

Change-Id: I4991f051e4e7f357be40bdf9b5770e292b8780a1
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/231560
Reviewed-by: Davis Hyer <dhyer@instructure.com>
Reviewed-by: Robert Hanna <rhanna@instructure.com>
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
QA-Review: Caleb Guanzon <cguanzon@instructure.com>
Product-Review: Daniel Lee <dlee@instructure.com>