Refactor UploadButton
This change refactors UploadButton into UploadForm, which is
just the hidden form + processing the files selected by the
user and UploadButton which is the visible button the user clicks
plus UploadForm.  UploadButton provides the functionality it
always has, while UploadForm can be used where we don't want
the button.

refs UXS-54
flag=none

no flag because this is a refactor of existing code. If it's
determined we need to flag this, I can make that happen, but it
will likely mean having all the original and new files in place
and switch between the two. As it currently is though, there
will be a better git history.

test plan:
  - the Upload button on the files page should work
  - you should be able to upload multiple files
  - if you upload the same file(s) twice, you should be
    prompted whether you should replace or rename
  - if you upload a zip, you should be prompted
    to upload or expand

Change-Id: Iaee03da4f7e7da0e0284724ab33ce209c5dd9bef
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/232360
Reviewed-by: Augusto Callejas <acallejas@instructure.com>
Reviewed-by: Mark McDermott <mmcdermott@instructure.com>
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
QA-Review: Mark McDermott <mmcdermott@instructure.com>
Product-Review: Ed Schiebel <eschiebel@instructure.com>