Adjust required role for admin tourpoints
Some users were requesting that the admin tour appear
to a more narrow scope of permissions
(https://community.canvaslms.com/docs/DOC-26624-ready-release-notes-2020-04-11).
This commit makes it so the admin tour only appears for users with the "Account Admin" role.

It does modify the application controller to send the current users's
account roles to the JS Environment. This is required to know
whether the user has the correct role. It needs to be in the
application controller because the admin tour should be able to
appear on any page the user visits.

Test Plan:
Sign in as a user with the 'Account Admin' role.
Clear the "canvas-tourpoints-shown-admin "localStorage key
The admin tour should appear when the page is reloaded

Sign in as a user with some other account role.
First create a new role in the Account Permissions page.
You can create this role by going to Account Settings >
Admins tab and adding a new admin account.
Clear the "canvas-tourpoints-shown-admin "localStorage key
The admin tour should NOT appear when the page is reloaded

Refs UXS-105

flag = product_tours

Change-Id: I474ade584d743d7f0fa190fa283997737cf8891f
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/233550
Product-Review: Katrina Hess <khess@instructure.com>
QA-Review: Daniel Sasaki <dsasaki@instructure.com>
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Augusto Callejas <acallejas@instructure.com>
Reviewed-by: Aaron Shafovaloff <ashafovaloff@instructure.com>