Fix layout issues on course Files setting page
Rearranges the markup so the breadcrumbs are
inline with the rest of the header content.

Notice that this page does not look exactly like the
other settings pages. Such is the struggle of rendering
React markup next to Rails markup.

Anyway, the point is the files Tutorial Tray arrow should
live on the right side of the screen (which it now does)

Test Plan:
Make sure Tutorial Trays feature flag is on
Open up the files settings page of a course
See that the tutorial tray arrow is on the right
side of the screen.

Closes CNVS-48850

flag = none

Change-Id: I97791f0dd88afeec13ba0e7888dbb8a3a22f65db
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/233886
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Product-Review: Jody Sailor
Reviewed-by: Ed Schiebel <eschiebel@instructure.com>
QA-Review: Ed Schiebel <eschiebel@instructure.com>