add gergich check to enusre yarn.lock is checked in
fixes: CCI-303
flag = none

Test-Plan:
this is likely going to break on first PS with this
gergich check. so if it does, then we update yarn.lock
in this PS to show it working.

if not, we'll need to add a change in package.json to require
a yarn.lock check failure.

Change-Id: I7d30e7953142dde71dbbe746e089725a3cb8d473
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/234179
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
QA-Review: Rex Fleischer <rfleischer@instructure.com>
Reviewed-by: Jacob Powell <spowell@instructure.com>
Reviewed-by: James Butters <jbutters@instructure.com>
Product-Review: Rex Fleischer <rfleischer@instructure.com>