parse pagination links more safely in cheat depaginator
This is a sanity check for a situation which does not presently occur in
Canvas. There was previously no safe checking on the 'Link' header in a
response, which would result in an uncaught error for users.
Server-side pagination is good enough that this does not happen. But it
is better to be safe than sorry.

Making this change has the immediate effect of making testing easier.
But it prevents the sort of regression described above.

closes TALLY-865
flag = gradebook_dataloader_improvements

test plan:
 * Verify Jenkins passes

Change-Id: I0c92e701bc99aa19851cf404588de49a4a189b15
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/234942
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Adrian Packel <apackel@instructure.com>
Reviewed-by: Spencer Olson <solson@instructure.com>
Product-Review: Spencer Olson <solson@instructure.com>
QA-Review: Jeremy Neander <jneander@instructure.com>