limit gradebook control to data updates in data loader
Despite the volume of code here (mostly specs), all this does is make
the initial rendering of the UI behind promise resolution, effectively
making it more reactive to state changes. The state changes are then
performed by the Gradebook methods called by the DataLoader when the
various essential data finish loading.

There is tremendous spec coverage around data loading, especially in
Seleniums. With as little as is changed in this commit (mostly
additions), the existing and added coverage is plenty to assure that
everything here is working correctly.

As a note, the existing path to finishing the initial render of the UI
remains untouched so that the Old DataLoader used in production is in no
way impacted.

closes TALLY-769
flag = gradebook_dataloader_improvements

test plan:
 * Verify Jenkins passes

Change-Id: I6149a3875b15cfa9e15d1ef6797bbcc14b0ec3c4
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/235434
Reviewed-by: Adrian Packel <apackel@instructure.com>
Reviewed-by: Spencer Olson <solson@instructure.com>
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Product-Review: Spencer Olson <solson@instructure.com>
QA-Review: Robin Kuss <rkuss@instructure.com>