use author_name from data on teacher notifications
the data element was not passed into the presenters which is handling
the logic to use the author name or the anonymous user info.
Summary Message was trying to load submission comment which is also not
present in that presenter since we just pass a submission over. I think
the email presenter was working through some magic, cause I was not able
to find an error on it, but this is more explicit on how it is passed in
to the presenter. Also they are the same content, no need to have a
separate presenter like other submission notifications. Other
notifications are different partially cause you can reply to them, but
these are notifications for stuff outside of canvas and we will not
process replies.

test plan
 - specs should pass
 - summary message should still work

fixes KNO-441


Change-Id: I07a8e73253b80f1f244ad1cb830fef6f05610237
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/235894
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Davis Hyer <dhyer@instructure.com>
QA-Review: Davis Hyer <dhyer@instructure.com>
Product-Review: Davis Hyer <dhyer@instructure.com>