change from using WORKSPACE to current directory for builds
while migrating to a build repo, we need to be able
to run these build scripts within the dir(){} on the checked
out root of this project. that will allow us to run these
scripts.

refs: CCI-393
flag = none

Test-Plan:
this build run and we test it with force-failure.
execute this with the canvas-builds while referencing this
PS to ensure that it will fix the WORKSPACE problem.

Change-Id: Idfe656771e2e4208e3b0980c942186a516f3bac1
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/237905
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Jacob Powell <spowell@instructure.com>
Reviewed-by: Derek Bender <djbender@instructure.com>
QA-Review: Rex Fleischer <rfleischer@instructure.com>
Product-Review: Rex Fleischer <rfleischer@instructure.com>