Exclude account JavaScript from media iframe
The Media Iframe page which the new RCE uses can be interfered with
by third-party JavaScript which has been loaded into the theme of an
account. In this case, the problem is specifically with Firefox, but other
JavaScript could affect other browsers in different ways.

Since there isn't really a good reason to load account JavaScript within
the media iframe, we should probably just disallow it altogether. This
patchset adds a flag to the iframe page which excludes the account
JavaScript only on that page; other pages are unaffected.

Test plan:
Before checking out this patchset, do the following to replicate the
issue.

Make sure you have some extra JavaScript included in your theme -
this one from Ashford makes Firefox not work properly:
https://instructure-uploads.s3.amazonaws.com/account_75490000000000001/attachments/12690750/ashford_js_20200420_edited.js

Navigate to a page that has the media iframe on it. This can be done
by uploading a video through the new RCE. On Chrome, it should load
fine; on Firefox, it perpetually say "Loading..."

Now checkout this patchset. That same media iframe should load
correctly on Firefox.

Verify that the JavaScript is still loading on other pages - it should be
loaded at the bottom of the #application element in the DOM.

flag = rce_enhancements

Closes LA-1088

Change-Id: Id6ad064ef7c07657c3dff55a9ee8f00282e1b891
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/240037
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Ed Schiebel <eschiebel@instructure.com>
QA-Review: Robin Kuss <rkuss@instructure.com>
Product-Review: Alex Anderson <raanderson@instructure.com>