accept the last page view send beacon request
Page view interaction is capped to a minimum of 5 seconds to a
maximum of 10 minutes or 600 seconds.

If self.interaction_seconds(or an instantiated PageView object) is nil
we will default to 5 seconds

If seconds > 5 we'll return the interaction_seconds to the UI
else we'll return the string '--' in place of interaction_seconds

In application controller we do a before_action `set_page_view`
if it's a user request and the request method is GET. We do an
after_action `log_page_view` if it's not a user request, but Javascript.
We explicitly look for `is request.xhr?` before logging the page view
and interaction seconds. For us to log the last page view without it
getting cancelled on page unload is by firing off a `sendBeacon()`
HTTP request. For this request to make it through the `log_page_view`
method we need to identify it by the proper response format and make the
strict XHR check be an OR.

fixes USERS-14
flag = none

Test plan:
* turn on page views
* go to a page and sit there for more than 30 seconds
* clear your rails log and your browser log
* click to go to a new page
* you should see a request in your browser Network log to
  page_views/xxxx-xxxx-xxx-xx-xxx?page_view_token=xxxx
* look at your rails server logs, you should see something like:
  Processing by PageViewsController#update as */*
  Parameters: {"interaction_seconds"=>"1005”, …
  ”id"=>"36bcbd62-3a11-44bc-8dd2-69410050ff74"}
* as an Admin check that user's page views and verify in the
  Time column an integer representing seconds of page interaction
  is logged for the proper route used to test

Change-Id: I972011058201997661a83c1b9365f5971e0ca72f
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/240128
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
QA-Review: Ahmad Amireh <ahmad@instructure.com>
Product-Review: Ahmad Amireh <ahmad@instructure.com>
Reviewed-by: Ahmad Amireh <ahmad@instructure.com>