message creator refactor part 5
this particular code was added to message creator when processing
messages was moved from notifcation.rb. When it was brought over it was
broken because it was assigning the message to daily instead of checking
for daily. This was later fixed because in a commit that was for
reducing memory usage.

delayed_messages is delayed_messages for all users passed into the
message_creator at that time, and this should not be used to determine
if a fallback message should or should not be sent.

Test plan
 - specs should pass

flag=none
refs VICE-414

Change-Id: I10f12d4d8a84984ddf46679b793f8e7db7cfb1f5
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/240337
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
QA-Review: Matthew Lemon <mlemon@instructure.com>
Product-Review: Matthew Lemon <mlemon@instructure.com>
Reviewed-by: Matthew Lemon <mlemon@instructure.com>