RCE respects image when creating a link
If an image is selected when the user creates a link, link the
image and don't replace it with the link text.
I tried creating a unit test for the change, but it became more about
testing the mock tinymce than it did the actual code.

closes LS-1140
flag = rce_enhanements

test plan:
  - add an image to the RCE
  - click it, then create a link (external or course, doesn't matter)
  > expect the image to be linked and not replaced with text
    (you can confirm there's a link there by looking at the HTML view)

  - try all manner of other link creating and editing scenarios and
    make sure they still work. This is the umteenth time I've tweaked
    link editing. Seems I fix 1 thing and break another.

Change-Id: I25316807c03e675186743c8422ee64a94c6cbcef
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/240491
QA-Review: Robin Kuss <rkuss@instructure.com>
Tested-by: Ed Schiebel <eschiebel@instructure.com>
Tested-by: Keith Garner <kgarner@instructure.com>
Product-Review: Ed Schiebel <eschiebel@instructure.com>
Reviewed-by: Alex Anderson <raanderson@instructure.com>