Use smaller line spacing for code blocks
This addresses a use case with courses that have significant code
components on quizzes, in particular if the questions are sensitive to
whitespace with how lines line up with each other (e.g. on format string
questions), which might be difficult to use for students with dyslexia
or similar issues.

Test Plan:
- Confirm that <pre> code blocks have smaller line spacing than the
  default 1.5.