Handle RCE switching to/from Links w/o error
The problem was that in switching from Course Links to
Files, we default to All Files, which has to be in the
user context, not course so we can retrieve the media.
Since we are switching from Links to All Files, it also
didn't make sense for the top drop-down in the tray to have
Course Files and User Files, so that menu is now
just Links and Files. If you've gone from Links -> Files
then change All -> something else (say Images), since we're already
in the user context, we change to User Files / Images.
From there you can change to Course Files if you want.

closes LS-1168
flag=rce_enhancements

test plan:
  - in the new rce
  - click Links > Course Links from the toolbar
  > expect the tray to open
  - open the Select where "Links" is currently selected
  > expect the options to be "Links" and "Files"
  - select "Files"
  > expect to see All Files
  - change "All" to Documents
  > expect to see User Files / Documents
  - change to "Links"
  > expect to see course links
  - try any other transition you can think of.

Change-Id: Ifa695abb7529ab5349429d27ec9835ed4973883a
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/241717
Reviewed-by: Alex Anderson <raanderson@instructure.com>
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
QA-Review: Robin Kuss <rkuss@instructure.com>
Product-Review: Ed Schiebel <eschiebel@instructure.com>