Use granular checkboxes instead of PermissionButton in tray
Fixes FOO-647
flag=none

If the "Tray" is slid out for a granular permission by clicking
on the name of the granular permission in the main permission
table, the controls there displayed the full PermissionButton
component, which allowed the full set of actions via the popup
menu. But granulars can only be enabled and disabled, not operated
on in any other way, so trying to change them in the tray resulted
in a 400 from the back end.

Now the tray just shows the same black and white checkmarks as in
the main permission table, to reflect the same functionality.

Test plan:
* Go to your account's permissions settings
* Find the "Manage Pages" group row
* Click on the "Manage Pages" link on the left to pull up the tray
* It should look like it always did, with red or green permission
  buttons for each role
* Now expand the Manage Pages group permission to expose the three
  granulars
* Click on "Delete Pages" (or one of the other two) to pull up the tray
* It should only show checkboxes instead of the red and green circles
* Clicking on the checkboxes should enable or disable the granular as
  appropriate
* Changes made in the tray get reflected in the main table as well

Change-Id: Ia20f7f8284432382d56eadd4cc485170e4667de3
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/242061
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Ethan Vizitei <evizitei@instructure.com>
QA-Review: Ethan Vizitei <evizitei@instructure.com>
Product-Review: Ethan Vizitei <evizitei@instructure.com>