Add score statistics to the assignments API as an 'include' option
This is designed to match the grade distribution graphs
shown in the web UI on the 'Grades' screen to students.

The permissions model is identical: students can view grade
grade distribution (min, max, mean) only if all of the following:
  - The 'hide_distribution_graphs' setting is NOT enabled
  by the instructor in the course's settings
  - There are at least 5 graded submissions
  - The student's submission is graded, and the grade is visible
  to them.

Instructors (or anyone with the read_admin permission) can
always view distributions when they are looking at a student
with a graded assignment.

Test Plan:
  - Make API requests to the GET endpoints for individual assignments,
  assignment groups, and for listing all assignments for a course
  - Test these API endpoints with and without the
  include[]=score_statistics flag.
  - Verify that the statistics match those shown in the grade
  distribution graphs on in the web UI
  - Verify that the statistics are only available in the API
  when there are 5 graded submissions and when not disabled by
  the instructor (this behavior should match the distribution
  graphs in the web UI)