Fix infinite loading spinner on RCE LTI tools
TL;DR: Removing the spinner URL from the LTI tool
iframe fixes a race condition causing the tool to not
open.

Sometimes, when opening an RCE LTI tool, a loading
spinner appears and the tool doesn't load.

The code around this is really messy and difficult to tell
what's going on. By my best guess, here's what's going on.

The LTI tool loads in an iFrame, but before that happens,
the source of the iFrame is the loading spinner gif.
Somewhere, the contents of the iFrame are getting
replaced with the LTI tool, but I can't tell where that
is happening.

However, when the spinner appears and spins forever,
it seems like the LTI tool doesn't even try to load. There
are no network requests, other than to fetch the spinner
graphic.

When I removed the spinner URL from the iFrame src attribute,
the LTI tool loaded correctly every time. This leads me to
believe that there is some kind of race condition happening
which is why the bug is so difficult to reproduce.

I replaced the spinner gif with the InstUI spinner.

Test Plan:
As a teacher, navigate to an RCE
Make sure you have some LTI tool added to your account
In the RCE, launch the LTI tool
An InstUI spinner should appear briefly in the modal.
The LTI tool should launch correctly, and the LTI tool should appear.

flag = rce_enhancements

Fixes LS-1150

Change-Id: Iac36760e407cd469c0140e2a09ac4518f8bad7bc
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/242463
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Ed Schiebel <eschiebel@instructure.com>
QA-Review: Robin Kuss <rkuss@instructure.com>
Product-Review: Peyton Craighill <pcraighill@instructure.com>