Don't send error object, just string, to mobile login page
Fixes FOO-572
flag=none

During the Covid Strikeforce effort, we altered the error message
sent back for a failed login so that the flash would persist for
more than the default few seconds. Unfortunately this did not take
into account the mobile login screen which was still expecting a
string to display, and thus was displaying the JSON instead. So
let's not render the more complex JSON if the user is using a
mobile browser.

Test plan:
* Try to reproduce mobile login error message problem as specified
  in the ticket
* Mobile login error should be plain text with a URL, not a JSON

Change-Id: I1426d6f63dcc17c487b7f5133928786fbe4daa52
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/242750
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
QA-Review: Simon Williams <simon@instructure.com>
Reviewed-by: Simon Williams <simon@instructure.com>
Product-Review: Simon Williams <simon@instructure.com>