Don't paginate in such tiny slices when getting subaccounts
Fixes FOO-728
Refs GH-1673

GitHub issue for canvas-lms #1673 pointed out that if an account
a great deal of subaccounts, when creating a new course the list
of subaccounts becomes ready only very slowly because the page
size on the API call is defaulting to 10. This results in many,
many trips back to the well before the subaccount list is usable.

We set it to 200 here.

Test plan:
* Creating a new course via the modal still works
* Use your network tab to verify that the API call includes a
  page size i.e. /api/v1/accounts/1/sub_accounts?per_page=200

Change-Id: I73d402324116060bc5c6e0edbb3a739610b8e380
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/243229
Reviewed-by: Cody Cutrer <cody@instructure.com>
Product-Review: Cody Cutrer <cody@instructure.com>
QA-Review: August Thornton <august@instructure.com>
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>