Include same-origin permission on iframes
The iframe sandboxing made it so media uploads
didn't properly render after being uploaded. This
is obviously a problem, so this patchset adds
the allow-same-origin sandbox permission to
all iframes.

This effectively reverts LS-743, so that ticket
was reopened so it can be addressed at a
later date.

This will need to be warm fixed.

Test Plan:
Upload media with the new RCE.
Save it.
The media should load correctly on the page.

Refs LS-743
Fixes LS-1260

Change-Id: I130af41a369d19ec02abf18682569a88d0a18f77
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/243902
QA-Review: Robin Kuss <rkuss@instructure.com>
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Jackson Howe <jackson.howe@instructure.com>
Product-Review: Alex Anderson <raanderson@instructure.com>