Show RCE auto-save modal only when RCE is visible
On the Account Settings Announcement, the autosave modal
can appear, even if the RCE is invisible on a separate tab. This
is less than ideal.

This makes it so the autosave modal only appears if the RCE is
actually visible on the page and even work if it is inside a
'display:none' parent.

Test Plan:
With RCE Enhancements and RCE autosave on
On any page, make some changes to the RCE. Then click cancel
Go back to the RCE for that page. The Autosave modal should
appear.

On the Account Settings page, go to the Announcements tab
Click 'New Announcement'
Type-y Type in the RCE. Click cancel.
Reload the page.
No Autosave modal should appear.

flag=rce_enhancements

Fixes LS-1229

Change-Id: I1ba51dfbe016d828ce9ea44176890f4d325a2cc6
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/243969
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Ed Schiebel <eschiebel@instructure.com>
QA-Review: Robin Kuss <rkuss@instructure.com>
Product-Review: Alex Anderson <raanderson@instructure.com>