Add Lato Extended font
Note: This reverts a previous revert, reintroducing the fix
without the regression.

The Lato Google Font is missing a bunch of glyphs which are
necessary for displaying certain languages. This adds the
extended character set fonts to Canvas to self-host the fonts.

It also changes the `font-family` CSS property to use the new
"Lato Extended" font, falling back to regular "Lato". It continues
to load the Google Fonts font, in the hopes that the Google
Fonts font will load faster (because it is about 10x smaller, due
to fewer character glyphs), and then the extended Lato font will
load and override the Google Font.

Test Plan:
Open the following page in the RCE and copy the raw HTML:
https://osmith.instructure.com/courses/1093/pages/reo-a-iwi-a-e-i-o-o-u-u-m-n-e-s-c-r-z-u
Paste that HTML into a page RCE raw HTML editor on your
testing Canvas instance and save the page. Also copy and
paste the page title to test bold characters.
ON WINDOWS, notice that the font renders basically the same
for extended characters as it does for latin characters.

Also, make sure the italic version works correctly
Within a Canvas course, create a page.
Enter text, highlight the text and select the italics icon to italicize the text. Notice that the text is italicized.
Save the page and notice that the text should be italicized and the correct font weight.

Fixes LS-317

flag=none

Change-Id: I2a44fd43d5433f3cecc9a1c43589643284536245
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/244837
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
QA-Review: Robin Kuss <rkuss@instructure.com>
Product-Review: Alex Anderson <raanderson@instructure.com>
Reviewed-by: Jackson Howe <jackson.howe@instructure.com>