Add spinners to show when a permission change is in flight
Closes FOO-873
flag=none

Since the buttons and checkboxes on the permission page are
controlled, changing their state in the UI does not result in
an immediate change in how they're displayed... it's only when
the API call to do the change completes that the icon or the
checkbox is updated to reflect it. This can be confusing because
users expect _something_ to happen when they click on a button.

So let's add little spinners to everything on the permissions
page, and in the slideout tray too. They appear whenever an API
call is in-flight for whatever permission/role this is, and
go away when the API call completes (or fails).

Test plan:
* Setting permissions all still works, including granulars
* Works in the tray too
* The spinners appear while the API call is outstanding, everywhere.

Change-Id: I6ce0343e5fb20eb562ed909c5a064b33277e8762
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/245612
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Simon Williams <simon@instructure.com>
QA-Review: Simon Williams <simon@instructure.com>
Product-Review: Charley Kline <ckline@instructure.com>