Fix broken video resizing
With the tinymce upgrade to v5, what tinymce hands back as the selected
node when clicking on an iframe changed, and our logic for dealing with
that didn't.

This change does most of the work up front to find the actual embedded
player's iframe, then downstream functions deal with that. While not perfect
this will help isolate tinymce changes from our manipulations.

With PS 2, allow a video that's smaller than the requested size to expand
and fill the available space. Before it would snap back to it's natural
size if that was bigger than the user asked for size.

closes LS-1351
flag=rce_enhancements

test plan:
  - upload a video
  - click on it, then Video Options to open the tray
  > expect to see the current video size in the Size section
  - change the size (try predefined and custom) and click Done
  > expect to see the video change size
  - wouldn't hurt to try updating other video attributes

  - repeat for a video pulled from the media tray

  - save and re-edit
  > expect you can still update the video's size and other attributes

  - add a small (in size, not length) video to the RCE
  - using the Video Options custom size, make it larger than it really is
  > expect it to expand to the specified size.

  > expect fullscreen to look right for landscape and portrait videos

  - using the tray, change the video from embed to a link
  > expect it to work
  - save
  > expect the link to work and look right
  - shrink the browser
  > expect the player to shrink as it becomes  necessary

Change-Id: Icc0ae7d5bf5800390b830275f6fb95915676c2ba
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/245589
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Alex Anderson <raanderson@instructure.com>
QA-Review: Robin Kuss <rkuss@instructure.com>
Product-Review: Ed Schiebel <eschiebel@instructure.com>