Refine heuristics for showing the autosave modal
The Autosave modal is trigger happy. It just looooves
to open up, even when it's not supposed to.

Specifically it opens up when there are images or videos
in the RCE editor, because Canvas enjoys ~making life hard~
swizzling the content, so it will never match up with
the autosave storage.

This patchset makes it so the autosave only compares
text content (instead of HTML content) between the
saved content and the autosaved content.

It also adds a hook which the page can call to tell the
RCE that it is closing legitimately, indicating that
autosave content should be deleted.

Pros:
It works 80% of the time, every time
No more obnoxious false positive popups

Cons:
It won't work when the user has only uploaded videos
or images anymore, it only works when the user has
made changes to text.

That's just the way the cookie crumbles, I guess.

This patchset only adds the close/save hook to the
pages RCE; it doesn't add it to the discussions,
announcements, assignments, or any other RCE.

Test Plan:
Create a new page
Put in some content and save
Edit that page.
Add an image.
Reload without saving.
The modal should not appear
Upload another image, and put in some text content.
Reload the page. The RCE modal should appear.

flag=rce_enhancements

Fixes LS-1322

Change-Id: I23d87847d384d4ea9ac84ef1dd43aa720c68027c
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/246334
Reviewed-by: Ed Schiebel <eschiebel@instructure.com>
QA-Review: Robin Kuss <rkuss@instructure.com>
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Product-Review: Peyton Craighill <pcraighill@instructure.com>