use partition_by_shard for user_observers on merge
because users can live cross shard, it is more simple to activate the
observation_links shard to make sure the proper ids are used and updated
this also splits observer side and student side up a bit, using a
similar pattern.

There are good number of tests in this area, mostly on the split users
side, but there was one case that was failing prior to this commit which
the new spec was added inside of split_users.

test plan
 - specs should pass
 - have user_observers and observees from various configs and shards
 - merge the users
 - the ids should all be valid and point to correct users

fixes VICE-765
flag=none

Change-Id: I83750844f5bb50b17dbf70661fd8a739b307020c
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/246878
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: James Williams <jamesw@instructure.com>
QA-Review: Rob Orton <rob@instructure.com>
Product-Review: Rob Orton <rob@instructure.com>