Show a spinner while images or media are uploading in the rce
While images or media are uploading, the RCE displays a grey box as a
placeholder for where the image/media will go. This didn't provide much
in the way of feedback that anything was happening.  This change adds
a spinner. Since we can't render instui components w/in tinymce, the
new-improved placeholder copied the instui Spinner's svg and css to
mostly replicate it from scratch.

As an added bonus tweak: before, if the user didn't like the newly
uploaded image and did "undo" do get rid of it, the first undo removed
the new image/media and put the placeholder back. It took a 2nd undo
to remove the placeholder. The placeholder is no longer part of the undo
stack so that doesn't happen.

closes LS-495
flag=rce_enhancemnts

test plan:
  - in the RCE, upload an image
  > expect the placeholder to have a spinner
  - undo (cmd-z)
  > expect the image to disappear and put you back to where
    you started w/o ever seeing the placeholder again.

  - repeat for a video
  > expect the same result.

Change-Id: I93a121638882bd1bc07745e40a9e6132d0563321
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/247156
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Jackson Howe <jackson.howe@instructure.com>
QA-Review: Robin Kuss <rkuss@instructure.com>
Product-Review: Ed Schiebel <eschiebel@instructure.com>