Make file link URLs in the RCE uniform
- linked files have /download at the end of their path
  and wrap=1 in the query_string. This should cause a click to get the
  access logged by canvas and then preview it
- embedded files have /preview and no wrap=1, which should get it
  downloaded, not logged and not previewed which is needed for the
  <img> src attribute

There is some URL manipulation happening in the RCE that should be taken
care of in the RCS, but with the work happening here, we're not dependent
on an RCS update. That will likely come later (I have a commit in waiting)

closes LS-972
flag=rce_enhancements

test plan:
  - link files from all over the place in the RCE
  > expect them to have the prescribed URL shape
  - embed an image
  > expect it to show up

don't bother clicking on the links and expect the file to always be
previewed as expected. That needs to happen in another ticket.

Change-Id: I9eb587ad8af132257c8b371fcac1c7a8d0305405
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/247225
Reviewed-by: Jeremy Stanley <jeremy@instructure.com>
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
QA-Review: Robin Kuss <rkuss@instructure.com>
Product-Review: Ed Schiebel <eschiebel@instructure.com>