Update math equation rendering
This commit changes nothing in the RCE, but changes how canvas renders
equations in the resulting page. Rather than
adding mathml in a hidden span adjacent to the equation image for
MathJax to process, this change replaces the image with a span containing
the LaTex source for MathJax to format.  This is better for a couple
reasons.

MathJax is not intended for formatting equations as they are
being edited and dealing with MathJax processed equations in the RCE when you
may want to edit an existing equation is never going to work well.

The visible MathJax-ified equations in the resulting page provides the
accessiblity we require.

This approach will update all existing content with math images,
so old content gets the benefit too.

Contrary to what some believe, you will not be able to select, copy
and paste parts of an equation.

closes: LS-1401
flag=new_math_equation_handling

test plan:
  - with new_math_equation_handling flag on
  - insert a math equation in the RCE
  > notice that when you click on the equation's image, you do not
    get the "Image Options" popup button
  - click on the equation
  > expect the "Edit Equation" popup button, not "Image Options"
  - edit the equation and save
  > expect it to be updated
  - save the page
  > expect the equation to fade into view after being processed by MathJax.
  > expect the MathJax menu when you right click in it
  > expect screenreaders read it nicely
  - edit the page again
  > expect the equation as an image again.
  > expect to be able to edit the equation
  - switch your user to a different language
  - open a page with an equation and right-click in the eq.
  > expect the mathjax context menu to be in the user's language
    (assuming mathjax supports it)

Change-Id: Ieac6785d51c0cab475b1176712f46fc2c964ff71
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/247471
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Jackson Howe <jackson.howe@instructure.com>
QA-Review: Daniel Sasaki <dsasaki@instructure.com>
Product-Review: Peyton Craighill <pcraighill@instructure.com>