prevent Safari from using no-cache to cache with validate
This update is to assist the Safari CORS issues with Canvas uploaded
images that are served from an AWS S3 bucket. Safari errors out with
CORS error. This is only when Safari makes a CORS OPTIONS validate
request on the cached image, that other browsers do not make.

> FetchEvent.respondWith received an error: TypeError: Cross-origin
> redirection to https://instructure-uploads.s3.amazonaws.com/[the bucket
> link and params] denied by Cross-Origin Resource Sharing policy: Origin
> [the Canvas LMS DNS] is not allowed by Access-Control-Allow-Origin.

Only Safari is caching and making the validate request. Chrome and
Firefox do not.

According to the Mozilla Cache-Control post, the "no-cache,no-store" is
contradictory:

> GOOD: "Cache-Control: no-store"
> BAD: "Cache-Control: no-cache,no-store"
> https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Cache-Control

> no-cache: The response may be stored by any cache, even if the response
> is normally non-cacheable. However, the stored response MUST always go
> through validation with the origin server first before using it,
> therefore, you cannot use no-cache in-conjunction with immutable. If you
> mean to not store the response in any cache, use no-store instead. This
> directive is not effective in preventing caches from storing your
> response.

Unfortunately Rails will add a private cache directive by default in the
cache control response header if "no-cache" is omitted. However, if the
"no-store" response directive is provided it indicates that any cache of
any kind (private or shared) should not store this response, which is
what we want.

Because of the difficulty of testing and reproducing locally, we are
deploying this with a quick-revert Setting. If this is causing problems,
set `Setting.set("legacy_cache_control", "true")` to revert to the old
behavior

closes #1700
closes FOO-2235
flag = legacy_cache_control (Setting)

test plan:
- ensure you see the updated Cache-Control header response
- smoke test file downloads

Change-Id: I92e9da47f39ff865b38bbefb7e1ae678e2f7290f
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/275721
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Jacob Burroughs <jburroughs@instructure.com>
QA-Review: August Thornton <august@instructure.com>
Product-Review: August Thornton <august@instructure.com>