avoid modal close on document click
prior to this commit the GroupModal was requesting to be closed
when the outlying document received any mousedown or mouseup
events. this can be annoying, as one could lose their form input
in the shuffle. we'll include the optional prop and set it to
explicitly false to avoid this particular scenario.

refs FOO-907
flag = none

test plan:
 - this functionality is tested at the root <Modal /> component
   layer within InstUI

Change-Id: I71f4d7105d7950b10113985e6999ff21fac409ba
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/248578
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Charley Kline <ckline@instructure.com>
QA-Review: Charley Kline <ckline@instructure.com>
Product-Review: Charley Kline <ckline@instructure.com>