Disable Apply button in ColorPicker if hex is invalid
Previously, if the user tried to save an invalid hexcode in the
ColorPicker, an alert would tell them the color was invalid. This
made it difficult for SR users to know ahead of time if the hexcode
was valid. Now, the Apply button is disabled until a valid color is
entered, and a screenreader message about the field being invalid is
announced when the user leaves the input field to supplement
the invalid icon that is displayed to sighted users. Tooltip also
added to the icon.

fixes LS-1437
flag=none

Test plan:
 - Open color picker (there's an instance on the dashboard and in
   calendar)
 - Type in an invalid hex code, expect apply button to be disabled
   and to see an icon with tooltip
 - Enter valid hex code, or click one of the predefined colors,
   expect apply button to now be enabled
 - Expect functionality to still work
 - Turn on SR - when hexcode is invalid, tab out of input field,
    expect to hear a message saying color is invalid

Change-Id: Ife58de18489bb698084738e886ab8245f1fa94f9
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/249181
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Jeremy Stanley <jeremy@instructure.com>
QA-Review: Jeremy Stanley <jeremy@instructure.com>
Product-Review: Jackson Howe <jackson.howe@instructure.com>