add timeout to cleanup function callback executed by protectednode
in some cases if the node removal happened after the code checks for a
node removal event the cleanup steps are executed and they can cause
extremely long running build. A fix to that is the introduction of a
timeout in the cleanup function.

refs DE-306
flag=none

test-plan:
- Canvas-lms build passess

Change-Id: Ia739bbdc1b72eac8ae7e95d5d36ebd698ce6d56e
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/249252
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Aaron Ogata <aogata@instructure.com>
Product-Review: James Butters <jbutters@instructure.com>
QA-Review: Andrea Cirulli <andrea.cirulli@instructure.com>