Don't raise error when tool launched with duplicates in context
Fixes INTEROP-6260

Right now LTI 1.3 assignment launches will fail in the
following scenario:

There are duplicate LTI 1.3 tools in the context: Tools A and B

- An assignment was created with Tool A
- The assignment was modified to use Tool B

This scenario is a legitimate situation that could occur
when an institution is upgrading an installed tool and want
to test that things work before uinstalling the old tool.

The solution here was to remove the check that required the
tool on the resource link record to match the currently
configured tool (See g/170818).

Also note the message of that commit correctly indicated that
this check was "...almost certainly incorrect
long-term behavior" (although it made sense at the time
it was added)

Test Plan:
- Install an LTI 1.3 tool and configure it with an assignment
- Install a duplicate LTI 1.3 tool
- Edit the assignment to point to the duplicate tool
- Validate that the assignment still launches and that
  AGS can be used from the duplicate tool

Change-Id: If1be3894b26a7cb845a8f70ca5e030c917467f0a
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/250017
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Mysti Lilla <mysti@instructure.com>
QA-Review: Xander Moffatt <xmoffatt@instructure.com>
Product-Review: Weston Dransfield <wdransfield@instructure.com>