Show little placeholder when uploading image as link
When uploading an image, the user can choose to embed it in the rce,
or display the filename as a link which opens the image in a new
tab. When displaying as a link, the placeholder that appears while
uploading should be small since it will be replaced by a link, not a
large image.

flag=rce_enhancements
fixes LS-1557

Test plan:
 - Click 'Upload Image' in RCE and select an image from computer
 - Under attributes, select 'Display Text Link'
 - Click 'Submit'
 - While uploading (set network speed to slow if needed), a little
   placeholder (same size as the link text) should appear
 - Uploading images to embed should still show full size placeholder

Change-Id: I786a0c7d1d67a4142080efdefc1823743f22e566
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/251668
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Ed Schiebel <eschiebel@instructure.com>
QA-Review: Ed Schiebel <eschiebel@instructure.com>
QA-Review: Robin Kuss <rkuss@instructure.com>
Product-Review: Jackson Howe <jackson.howe@instructure.com>