Apply MathJax to equations in Discussion replies
closes LS-1595
flag=new_math_equation_handling

Most RCE content is processed by UserContent.escape in the
canvas backend on its way to the browser, and that is where
most equation images are replaced with their LaTex for MathJax
processing.  The flow is different for Discussions whose replies
follow a different path. This change replaces the image with LaTex
in the same place in the browser as where the hidden mathml was
previously injected into the DOM.

test plan:
  - in a discusion, add a reply
  - use the Equation toolbar button to insert some math
  > expect the equation image
  - save by clicking Done
  > expect the image to be replaced with MaxJax typeset content
    (right click and you'll get the mathjax menu)
  - edit the reply
  > expect the equation image in the RCE
  - edit and add some inline math (e.g. $$y = x^2$$)
  - save
  > expect both the original image and the inline equation to
    be typeset by MathJax

Change-Id: Icedcbbf3b2f084b1a9c254ec2a070b31c35ffed9
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/251979
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Jackson Howe <jackson.howe@instructure.com>
QA-Review: Robin Kuss <rkuss@instructure.com>
Product-Review: Ed Schiebel <eschiebel@instructure.com>