Fix CORS request for equation image
closes LS-1599
flag=none

A recent update does a preemptive fetch of the equation image to ensure
it can be returned. We did not have the necessary incantation to
properly make the CORS request. While chrome and ff were OK with
the request headers, safari was not and rejected the request with
the message "Cannot use wildcard in Access-Control-Allow-Origin
when credentials flag is true."

test plan:
  - in safari, clear your cache (or create a brand new equation)
  - in the RCE, use the equation editor to create an equation
  - click "Insert Equation"
  > expect to see the equation image in the RCE

Change-Id: If8d1f6b72741b2943faf251be5f04320157b337f
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/252115
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Robin Kuss <rkuss@instructure.com>
QA-Review: Robin Kuss <rkuss@instructure.com>
Product-Review: Ed Schiebel <eschiebel@instructure.com>