Sort the permission list properly
Fixes FOO-1192
flag=none

The permission list is supposed to be sorted, and in fact the
code tries to sort it before launching the permissions interface,
but it's using a compare function that returns a Boolean instead
of the expected negative, zero, or positive numeric value. Pretty
sure the behavior of `Array.sort()` is undefined in that case,
so no surprise that the list isn't getting sorted right.

Test plan:
* Go to the permissions page
* "Manage Pages" should be in the right alphabetical spot on
  the list and not at the end

Change-Id: I7f2b1fdc28f2677de3834e6c193a65fa3dc082b5
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/252410
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Ahmad Amireh <ahmad@instructure.com>
QA-Review: Charley Kline <ckline@instructure.com>
Product-Review: Charley Kline <ckline@instructure.com>