Allow importer to recognize override score changes
Allow the gradebook importer to recognize changes override scores and
record them as part of a gradebook upload. (The changes are not yet
shown to the user on the confirmation page or applied in any way; they
are only stored in the temporary upload object created on the backend.)

closes EVAL-1293
flag=import_override_scores_in_gradebook

Test plan:

Run the following tests with two CSV files: one with grades for the
course as a whole, and the other with grades for a specific grading period.
For the latter file, you can either ensure the separate "export grading
periods in column headers" commit is merged and the associated feature
flag turned on when you export, or you can fake it by taking an existing
file and renaming the "Override Score" column to include the grading
period (e.g., "Override Score (Fall 2020)").

- Have a course with override grades enabled and some grading periods
  active
  - Give out some override grades for the course and for individual
    grading periods
  - Export a CSV
  - After exporting, make some changes to override scores in either the
    gradebook or the CSV file
- With the feature flag ENABLED:
  - Open the Network tab or equivalent in your browser's developer tools
  - In Gradebook, import the CSV file from earlier
  - On the confirmation page, check the network requests for a request
    to gradebook_uploads/<id>/data
    - Examine the response JSON, and look at the "students" hash
    - Assuming there was at least one student with an override score
      change, each student hash should contain an "override_scores"
      entry with override score data (even if unchanged)
    - If there were no override score changes at all, none of the
      "override_scores" arrays should contain any elements
  - The actual "review changes" page will not recognize the changes yet,
    so if it claims that there are no changes, that's fine
- With the feature flag DISABLED:
  - Upload a modified CSV and examine the response JSON as above
  - It should not contain any override grade changes for any students

Change-Id: I6f4d391e8c639b1894306e325c5d1c1736dd58e5
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/252696
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Spencer Olson <solson@instructure.com>
Reviewed-by: Gary Mei <gmei@instructure.com>
QA-Review: Kai Bjorkman <kbjorkman@instructure.com>
Product-Review: Syed Hussain <shussain@instructure.com>