Targeted MathJax processing
closes LS-1640
flag=new_math_equation_handling

Currently, the mutation observer watches the body and if any of
the mutationList items' addedNodes list is an element and is not
in the ignore list, then we ask MathJax to reprocess the whole body.
The call back into mathml is debounced to keep from doing this too
often when chunks of the page are updating.

With this change, we walk the addedNodes nodes list and fire the
process-new-math event for each one then mathml then looks
to see if there's actually any math in the element, and if it's
not ignored, then has mathjax process it. This speeds up prcessing
because MathJax doesn't have to look at the whole doc and potentially
reprocess a bunch of math it's already processed and hasn't changed,
it's not having to wait for the debounce timeout.

test plan:
  - open pages with math
  > expect it to be typeset by mathjax
  > expect it to see the typeset math faster than before
    (good luck with that)

Change-Id: I9d3f8a0ab1efb6ed88c34d1bcf6caeec215ba140
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/253512
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Charley Kline <ckline@instructure.com>
QA-Review: Robin Kuss <rkuss@instructure.com>
Product-Review: Ed Schiebel <eschiebel@instructure.com>