Fix invalid prop issue for No Outcomes Billboard
After talking with Ryan, we decided to utilize the
message and heading property to display the two line
no outcomes message and instructions to the user.
This switches from passing an HTMLElement to the
heading property, which is not allowed, to using the
heading and message property.

closes OUT-4120
flag=improved_outcomes_management

test plan:
- Go to Account > Settings > Feature Options
- Enable Improved Outcomes Management FF
- With Improved Outcomes Management FF Enabled
- Go to Account > Outcomes
- You should see the no outcomes billboard that displays the
correct context in the message
- Go to Course > Outcomes
- You should see the no outcomes billboard that displays the
correct context in the message
- Disable Improved Outcomes Management FF
- Without Improved Outcomes Management FF disabled
- Go to Account > Outcomes
- You should see previous Outcomes manager
- Go to Course > Outcomes
- You should see previous Outcomes manager

Change-Id: I9cdee0d09b8cc2bcfebfbb7fc04515f1a2c49a11
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/254837
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Michael Brewer-Davis <mbd@instructure.com>
Reviewed-by: Pat Renner <prenner@instructure.com>
QA-Review: Martin Yosifov <martin.yosifov@instructure.com>
Product-Review: Jody Sailor