spec: fix failing plugin settings spec
In a not so distant past we added some changes to override
a few class methods on the PluginSetting class. If the
particular plugin is present in gems/plugins, which it is,
in Jenkins and production, we want to account for that.

fixes FOO-1284

test plan:
 • specs pass with 100% success rate

Change-Id: I84553174aea7f5553bb446032efe0bb963bcb13a
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/256342
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Cody Cutrer <cody@instructure.com>
QA-Review: Cody Cutrer <cody@instructure.com>
Product-Review: Cody Cutrer <cody@instructure.com>