Fix the new RCE overflowing in Discussions
closes LS-1797
flag=none

The fix is to set min-width:0 on the fieldset wrapping the RCE. This
is necesary because the browsers have an implicit min-width:min-content
that prevents it from shrinking.  The _real_ solution is to add this
to our global CSS, but if this is going to get hotfixed I don't want
to change any global stylesheet.

Fix #2: in the text entry assignment submission page the RCE is in a
table. The fix to keep it from overflowing here is to use
table-layout:fixed.  I can to this w/o specifying column widths because
there's only 1 column and they it'll be 100%.

test plan:
  - on a real phone or with the browser set to emulate a mobile device
    navigate to a new Announcement or new Discussion page
  > expect the RCE's toolbar to be constrained to fit on screen
    and drag-scroll
  - as a student, submit a text-entry assignemnt
  > expect the RCE to behave

Change-Id: I79a3a7f0e4580f14f7143ae56bd43d91fe3a0bd9
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/256923
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Robin Kuss <rkuss@instructure.com>
Reviewed-by: Jeff Largent <jeff.largent@instructure.com>
QA-Review: Robin Kuss <rkuss@instructure.com>
Product-Review: Peyton Craighill <pcraighill@instructure.com>