fix render loop in Navigation component
The reason the error occurs is because we're trying to update
state on the initial render, which in turn triggers a re-rendering,
which tries to set state again. So on and so forth.

UnreadCounts has the ability to store counts in the user's session
storage, which can help alleviate polling the API. This can also
cause some edge cases in the way our code is setup. If the
Navigation class component was written as a functional component
using hooks we'd have a better ability to handle instant prop
changes from child props updating a parent's state. In this case
we're seeing our `onUpdate` prop set to a callback function,
which gets fired almost immediately on initial render because
the value is instantly retrieved from local storage, causing the
infinite re-render loop due to the inherent nature of setState.
This of course doesn't _always_ happen, because sometimes that
value is stale or not present.

The solution in this case is to optionally bypass local storage
in the UnreadCounts child component and receive the update from
the API polling response. This is a more fluid update and allows
for the initial render to happen before setState can re-render
with the updated props.

closes FOO-1541
flag = none

test plan:
 • As a teacher load /courses/:course_id/grades/:student_id
 • In dev tools JS console you shouldn't see a Warning for:
   "Cannot update during an existing state transition"
 • As a student send a message to the prior teacher to update
   their unread count to > 0
 • As a teacher using responsive web view load your Dashboard
 • Verify you see an indicator circle to the top-right of
   the left hamburger menu
 • Select the hamburger menu and verify you see an indicator to
   the right on Inbox for how many unread messages you have

Change-Id: I9e17aeb294ffb85125f3d61da6dfd3339c4d70a7
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/258276
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Charley Kline <ckline@instructure.com>
QA-Review: Charley Kline <ckline@instructure.com>
Product-Review: Charley Kline <ckline@instructure.com>