pre-load correct image for cache calculation
In a previous PS, we cached the ruby-passenger image on starlord and update it via a cron job to prevent cache disruption during the working day. Unfortunately, we forgot to update the image pulled for the hash calculations, causing it calculate the wrong hash if the dockerhub image has changed and the sync job hasn’t run yet.

Change-Id: I17bb91f7f08ca4352fc3f0ae88563867f3ece6f6
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/258833
Reviewed-by: Kyle Rosenbaum <krosenbaum@instructure.com>
Reviewed-by: Andrea Cirulli <andrea.cirulli@instructure.com>
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
QA-Review: Aaron Ogata <aogata@instructure.com>
Product-Review: Aaron Ogata <aogata@instructure.com>