Show year on course availability inputs
A bug is preventing the year from being set to something other than
the current year in the course availability options. This change
fixes that issue by showing the year in the input itself, which is
probably a good thing anyway since tons of courses span 2 calendar
years.

A side affect of this change is that we need to increase the size of
the CanvasDateInput inputs to allow for the extra details in the
input.

fixes LS-1962
flag=new_course_availability_ui

Test plan:
 - Enable the flag at siteadmin
 - In a course's settings, change the course start or end date to
   something next year
 - Save the page and expect the year change to persist.

Change-Id: I468e2a9bad832062e42dbb206da1b55bea7564db
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/260203
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Nate Armstrong <narmstrong@instructure.com>
QA-Review: Nate Armstrong <narmstrong@instructure.com>
Product-Review: Jackson Howe <jackson.howe@instructure.com>