Add a user option to hide global announcements
Some users have devices where global announcements clog up space on the
dashboard and have to be closed every time there is a new announcement.
This change has an option for those users who want to not have them
appear at the top of the page but rather in the Global Announcements
section of Canvas, allowing this change to be enabled under Feature
Options in Canvas.

Test Plan:
- Login as a site admin
- Create an extremely long global announcement
- Go to the Dashboard tab
- Notice the announcement appears
- Go to your Account > Settings
- Under Feature Options, enable "Hide Global Announcements on the
  Dashboard".
- Go to the Dashboard tab
- Notice the announcement no longer appears, but is still visible in
  Account > Global Announcements.