Remove (vacuous) use of InstUI <Select> in RoleTray
Closes FOO-1573
flag = none

The Permissions RoleTray includes a call to a function which
renders the old <Select> component. This is a pretty trivial
invocation which was easy enough to swap out with the new
<SimpleSelect>.

This is just kind of a formality though, to get rid of the
old InstUI 6 deps... because the function that renders the
Select is short-circuited (line 447).

Test plan:
* none (component is currently short-circuited, the next
  person who enables that select box, if anyone ever does,
  can worry about any issues that there may be with this)

Change-Id: I830434c6f3c6fb0dfc8fd67a8c2db5e6bfd93689
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/261030
Reviewed-by: Ahmad Amireh <ahmad@instructure.com>
QA-Review: Ahmad Amireh <ahmad@instructure.com>
Product-Review: Ahmad Amireh <ahmad@instructure.com>
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>