require appropriate shape prop type
refs VICE-1343
flag=none

The shape function returns a function, not an object. Defining a prop
type that is an instance of a "shape" is actually defining the type
to be an instance of a function.

test plan:
  - run `yarn test:jest DiscussionsFullPage`
    - there should be no console error regarding the prop
      being provided as an object instead of `bound checkType`

qa risk: low

Change-Id: Ib60ccd852b83f34b2dcde77ecade939a329c5e4d
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/263527
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Omar Soto-Fortuño <omar.soto@instructure.com>
QA-Review: Omar Soto-Fortuño <omar.soto@instructure.com>
Product-Review: Omar Soto-Fortuño <omar.soto@instructure.com>