fix flakey calendar specs
refs DE-665

The calendar has view rendering logic that depends on setTimeout() to complete in order for the view to be in its final form. This causes selenium test flakiness because it isn't aware of how to wait for these timeouts.

Test Plan
1. FSC fails without this change
2. FSC passes with this change

Change-Id: Ic323153ab685d1ae8a2efec43d50ebe0582b9809
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/264325
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: James Butters <jbutters@instructure.com>
Reviewed-by: Jeremy Stanley <jeremy@instructure.com>
QA-Review: Aaron Ogata <aogata@instructure.com>
Product-Review: Aaron Ogata <aogata@instructure.com>