Validate submission type on non-API submissions
When a student submits an assignment via the web interface, make sure
that the assignment actually accepts the submitted type. In most cases
this isn't an issue because you're only allowed to select from the types
the assignment accept, but it could become one if the assignment changed
types while a student was midway through submitting, or if someone was
trying to do something sneaky.

Note that this specifically applies to the current student view--not the
enhanced view--and does not affect submissions made via API (for which
we already checked the type).

fixes EVAL-1749
flag=none

Test plan:
- Have a course that is *not* using the enhanced student view
- As a teacher, create an assignment that only accepts a certain
  submission type (e.g., text entries)
- As a student in a different browser, open the assignment and begin a
  submission, but do not actually submit
- As a teacher, update the assignment and change the allowed submission
  type to someting else (e.g., URLs)
- As a student, try submitting the assignment with the original type
- You should get an error
- Smoke test submitting to assignments with the *correct* type and
  ensure there are no errors for the following types:
  - Text entries
  - File uploads
  - An LTI submission (e.g., Studio)

Change-Id: I50ba59afd94996d60cc8275e2ff7f73ddeb03c3c
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/267717
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
QA-Review: Aaron Shafovaloff <ashafovaloff@instructure.com>
QA-Review: Syed Hussain <shussain@instructure.com>
Product-Review: Syed Hussain <shussain@instructure.com>
Reviewed-by: Syed Hussain <shussain@instructure.com>
Reviewed-by: Kai Bjorkman <kbjorkman@instructure.com>