Show announcements only to correct sections in k5
If an announcement is only posted to certain sections, be more
specific about who sees those announcements in k5. Users with
:read_as_admin, :manage_grades, :manage_assignments, or
:manage_content can always see all announcements in a course, but
otherwise an enrollment in a section the announcment is posted to is
required.

Also only fetch announcements for the JS_ENV if we're in a subject -
no point in doing the work for classic courses.

fixes LS-2335
flag = none

Test plan:
 - Create a second section in a k5 course, but don't enroll any
   test students in that section
 - Create an announcement posted to just the new section
 - As a student enrolled in the course (but not the new section),
   visit the course home
 - Expect to not see that announcement (though any other
   announcements created in the past 2 weeks should show)

Change-Id: I0e51a21f93d70ed106c81b8b0afc680183ce5ecd
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/267580
Reviewed-by: Robin Kuss <rkuss@instructure.com>
QA-Review: Robin Kuss <rkuss@instructure.com>
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Product-Review: Jackson Howe <jackson.howe@instructure.com>