Hide the grades tab in Dashboard If all subjects have it hidden
A new flag is passed to the JS ENV variables from the user
controller to hide the Grade tab for students when all subjects
except the homeroom course have this tab hidden or if the student
does not have active courses, this will happen only if the user is a
student and does not have the teacher role.

fixes LS-2215
flag=none

Test plan:
Scenario 1:
1. Enroll a user to 2 or more courses as student
2. For each course go to course settings and then navigation to
hide the Grades tab for all of the courses where the student is
enrolled.
3.As teacher go to the homeroom and expect to see the Grades tab as
usual
4.As student go to the homeroom and expect not to see the Grades Tab.

Scenario 2:
1.Enable the Grades tab in all the courses used in the previous
scenario
2.In each course go to the settings tab and set the participation to
course and override the dates to make the course inactive
3.As teacher go to the homeroom and expect to see the Grades tab
as usual
4.As student go to the homeroom and expect not to see the Grades
Tab.

Change-Id: Ia7f730a7951f4108571d0b241d87201e9a66ed86
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/268016
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Jackson Howe <jackson.howe@instructure.com>
QA-Review: Jackson Howe <jackson.howe@instructure.com>
Product-Review: Peyton Craighill <pcraighill@instructure.com>