MSFT sync better metrics for expected non-200 resp
Currently all graph service responses that are not 200s or 404s will be
counted in the metrics under "error". This commit separates out the
logic into a passed-in block, similar to the "ignored" block for
run_batch. Not only does this make it easier to understand what's going
and how many true errors we have for other endpoints, if we don't catch
a special case like we should (e.g.  the text we look for changes, we'll
be able to see that in the metrics -- for instance, before, when looking
at the metrics we could blindly ignore all 409 errors for POST /teams,
but we wouldn't know from that if the text we look for ("group is
already provisioned") changed and we stopped handling the 409 as an
expected error.

Also two other small changes
* change "::" in tag values to "__", these are awkward to deal with in
  datadog
* in GraphServiceHttp,  for the error statsd metric, add error name
  instead of just "unknown" when no status code received

refs INTEROP-6750
flag=microsoft_group_enrollments_syncing

Test plan:
- decrease DELAY_BEFORE_CREATE_TEAM in syncer_steps.rb. this will cause
  an error when trying to create the team, causing an "expected" error
  response
- instrument statsd by adding to syncer_steps.rb, or in a console you
  intend to run synchronous syncs in:
    class << InstStatsd::Statsd
      def increment(*args)
        puts "DEBUG STATSD increment: #{args.to_json}"
      end
    end
- run a synchronous sync job for a course without a group on the
  microsoft side, e.g.
    microsoft_sync_group.syncer_job.run_synchronously
  there should be two "expected" statsd incremented:
  - one for msft_endpoint="get_teams" (checking whether the team exists)
  - one for msft_endpoint="post_teams" (trying to create a team too
    early)
- add a user to the group in the microsoft admin. add the user
  on the canvas side. Run a partial sync manually:
    microsoft_sync_group.syncer_job.run_synchronously(:partial)
  there should be an "expected" statsd incremented for
  msft_endpoint="patch_groups" (adding a duplicate user) and it
  will fall back to using the post_$batch endpoint to add the users.

Change-Id: I8ce7c4244187228a46a06bcff161bf25461c7019
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/268514
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Mysti Lilla <mysti@instructure.com>
Reviewed-by: Ryan Hawkins <ryan.hawkins@instructure.com>
QA-Review: Ryan Hawkins <ryan.hawkins@instructure.com>
Product-Review: Evan Battaglia <ebattaglia@instructure.com>