MSFT Sync: I18n of errors part 1
Errors happen in a job, so to localize the errors to whatever the user
viewing the page sees, we need to store the un-i18n-ized version in the
database. However, if we do that, the messages won't be extracted to the
i18n yml file; so we still need to call I18n.t (we do this in #public_error)
with a constant string so the extractor can pick them out.
I'm not sure if jobs are even run in a non-en locale, but in case they
are, we temporarily set the locale to 'en' before running that I18n.t()
for that error to make sure we store the un-i18n-ized version.

TODO follow up commits:
* change more errors to PublicErrors
* GracefulCancelError can be a subclass of PublicError

Test plan:
- have a course with no owners (teachers/tas/designers) and try to sync.
- check the last_error field on the MicrosoftSync::Group model. It
  should be a JSON hash with appropriate public_error_message and no
  public_interpolated_values.
- check the UI (course integrations settings) to make sure the error
  shows up there (just the public message, not JSON)
- Add the following to es.config/locales/yml:
  microsoft_365_allows_a_maximum_of_max_owners_in_a__b448f5de: "Microsoft 365 permite como máximo %{max} administradores de un Microsoft Team."
- change MembershipDiff::MAX_ENROLLMENT_OWNERS to 2. restart Canvas.
  Add two owners in a course and try to sync.
- check the last_error. It should have a public_message and separate
  public_interpolated_values with "max": 2.
- check the UI to make sure the error message shows up with
  the interpolations.
- change your language to Español
  (/profile/settings -> Edit Settings) and reload the course settings
  integration page. The error message should now show up in Spanish.
  ¡Empieza la fiesta!

refs INTEROP-6739
flag=microsoft_group_enrollments_syncing

Change-Id: Ia723946e2d2f2b58fdaeb2a99cacfd23ed6f3197
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/269064
Reviewed-by: Xander Moffatt <xmoffatt@instructure.com>
QA-Review: Tucker Mcknight <tmcknight@instructure.com>
Product-Review: Evan Battaglia <ebattaglia@instructure.com>
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>