MSFT sync rename UPN -> ULUV (User LookUp Values)
Currently, to find users on the Microsoft side, we take their email,
username, or SIS id and look up the Microsoft user with the UPN
(UserPrincipalName) equal to that value. Thus, we call that
email/username/SIS id a "UPN" in the code.

Now however we are moving to a model with the user can choose what
target attribute on the Microsoft side they wish to match that value
(email/username/SIS id) to. (Additionally, they will be able to add an
arbitrary suffix to the value on the Canvas side before using it to look
up the user on the Microsoft side). Thus, we need a new term for this
value that we are calling a "UPN". I propose the term ULUV, or User
LookUp Value.

The only thing this commit does is rename constants/classes/variables.

refs INTEROP-6739
flag=microsoft_group_enrollments_syncing

Test plan:
- specs. as I add functionality in the next commits we will test that
  the old functionality (looking up by UPN) still works.

Change-Id: I5898e9c0fd53c67c4f24dbf6166e67d68af8d9f3
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/269157
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Ryan Hawkins <ryan.hawkins@instructure.com>
QA-Review: Ryan Hawkins <ryan.hawkins@instructure.com>
Product-Review: Evan Battaglia <ebattaglia@instructure.com>