InstID tokens, part 2: authentication
fixes INTEROP-6911

also fixes some specs in jwts_controller_spec -- the intention was to
use a services JWT to try to generate a services JWT and see it fail.
it did fail, but for the wrong reason -- the spec didn't actually pass a
services JWT in the auth header, it passed something unrecognizable.
the fix here is to make it a proper JWT, and add a more specific
assertion so we're confident that we're exercising the right code path.

test plan:
 - see test plan for commit c44b1844f74 for how to generate RSA keypairs
   and configure InstID to use them
 - get an API access token by following
   https://canvas.instructure.com/doc/api/file.oauth.html#manual-token-generation
 - use that token to generate an unencrypted InstID token like:
   $ curl 'http://localhost:3000/api/v1/inst_ids?unencrypted=1' \
     -X POST \
     -H 'Authorization: Bearer <access_token>'
 - now use the returned token to hit some other endpoint, e.g.:
   $ curl http://localhost:3000/api/v1/users/self \
     -X GET \
     -H 'Authorization: Bearer <InstID_token>'

Change-Id: Ie6c01dd12d98f68c138da1960b87f91e42b8b04d
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/269345
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Ethan Vizitei <evizitei@instructure.com>
QA-Review: Michael Ziwisky <mziwisky@instructure.com>
Product-Review: Michael Ziwisky <mziwisky@instructure.com>