Improve a11y form validation in New Permissions Role
Fixes FOO-950
flag = none

When adding a new role, if the new role name field is left
blank, the "Save" button was simply disabled, without any
indication of what the problem is. Proper accessibility
guidelines are to make the input field `required` so that
screenreaders can reflect this to the user, and also if an
attempt to save the form with an invalid field, an error
should be displayed.

Test plan:
* Go to /accounts/self/permissions and click on "Add Role"
* The form in the tray should have the Save button enabled
  even if the Role Name field is blank
* But clicking on the Save button with the field blank should
  display an error under the role name now

Change-Id: I60b48f7c3aba993c6560241c76054807498128ec
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/269802
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: August Thornton <august@instructure.com>
QA-Review: August Thornton <august@instructure.com>
Product-Review: Charley Kline <ckline@instructure.com>