iterate on dump helper
stop assuming that if nothing failed, this object is the culprit (it's likely nothing
failed because we skipped something self-referential, but we're not actually done
checking all of the other parts of it. instead, just check for the specific
dump problems we know we're looking for.

also clean up the output to be more useful

Change-Id: Id3535ba06807e70e93c7039ca748aed5b0ae6760
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/270715
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Simon Williams <simon@instructure.com>
QA-Review: Cody Cutrer <cody@instructure.com>
Product-Review: Cody Cutrer <cody@instructure.com>