Show observed users picker in subject
Place the same parent picker currently on the k5 dashboard within
k5 subjects as well. Only shows enrollments that exist in the
current subject.

This also moves the logic for calculating the users that appear in
the picker to the server, so that we have instant access to the
picker once the js renders and can preload the correct dashboard
cards. Uses a cookie for saving the selected user (so the server
can get at that too). The query to fetch observed users is cached
and run against secondaries.

Also tweaks the subject header to remove Student View button on
small screens, since the button exists in the mobile nav header.

closes LS-2480
flag = k5_parent_support

Test plan:
 - Open dashboard as a user who is observing a few students and
   who has (at least) one enrollment of their own
 - Expect picker to show their name
 - Open the subject where they're enrolled as student/ teacher
 - Expect picker to still show their name
 - Open a subject where they're observing
 - Expect picker to pick first (alphabetically) observer
 - Expect picker in the subject to be searchable and switchable
   just like on the dashboard
 - Expect selected user to persist when moving between dashboard
   and subjects (as long as the selected user is available in that
   subject)

Change-Id: Iccf5114735e2d600cc70b41934ab8c1cf9647b11
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/271360
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Jeff Largent <jeff.largent@instructure.com>
QA-Review: Jeff Largent <jeff.largent@instructure.com>
Product-Review: Jackson Howe <jackson.howe@instructure.com>