MSFT sync: avoid InternalError when Group restored
If a job goes into a retry state (i.e., a job is scheduled in the
future to start off from an intermediate step), and during this
time the user deletes and restores a Group before the job continues,
when the job starts again, it will be confused because the Group
is not deleted but the job_state has been cleared. (In a continuing
job, the job_state holds the current step, which must match up
with the argument in the job).

The solution is just to add {restored: true} to the job state to signal
that to an in-progress job (of which there can only be one) that it can
be ignored. When the continuing job starts, it will understand that
the group has been deleted and restored since the job was started,
so it can be dropped.

Note that if we didn't clear the job_state at all when
deleting/retrying, we could end up in a similarly confusing situation
such that kicking off a new job after restoring would be confused
because it's not in a fresh (no job_state) state.

closes INTEROP-7019
flag=microsoft_group_enrollments_syncing

Test plan:
- increase DELAY_BEFORE_UPDATE_GROUP in SyncerSteps to 60 seconds or so
  to facilitate testing, so the job will be in a "retrying" state for
  longer.
- run a jobs container
- Clear ms_group_id on the group and kick off a sync job
  (group.syncer_job.run_later)
- while the job is waiting [retrying] (you can double check by checking
  the group workflow_state) delete and restore the group in the UI.
- when the job continues there should not be an InternalError. Instead
  the message "In-progress job starting again but job state record was
  deleted & restored since" should be shown. The job should not
  continue, but rather be dropped.
- check the state of the MicrosoftSync::Group. It should still be in
  "pending".

Change-Id: Iac0e997e4b28a526262153919c562fd3c3be3dde
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/271943
Reviewed-by: Xander Moffatt <xmoffatt@instructure.com>
QA-Review: Xander Moffatt <xmoffatt@instructure.com>
Product-Review: Evan Battaglia <ebattaglia@instructure.com>
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>