Add pair student option in the observee picker
A new option was added to the student picker, it will be shown only
in the dashboard, this option will show up and modal where the user
can type the pairing code to link a student, the same API used in the
/profile/observees page in consumed in this modal

If the user has been create as a normal user(not in the parent
registration way), and the observer role has been provided in a
course manually, the user picker is rendered in the dashboard, but
the observing option is not display in the account navigation, so to
keep consistency the “Add Student” option is only shown if the user
has the observing option in his account navigation menu

For error handling, no empty values are allowed to be sent and if the
pairing code is invalid, a message below the input and a flash alert
will notify the user.

Closes LS-2523
flag = k5_parent_support

Test plan:
- With the parent support feature enabled, create a couple of
students
- Enroll those students to a course
- Make sure the Authentication provider is set to “Observer Accounts
Only” in the root account
- Access as each student and generate a pairing code for each of
them
- Create a canvas account as a parent using one of the codes
- When you login as the parent go to the dashboard and expect to see
the “Add Student” option in the user picker
- Click the “Add Student” option and expect to see the pairing
student modal
- Type the other pairing code you generated, and expect to the
student be paired with the parent user.

Change-Id: Id8e48215f3ba31134d713bc754c91c2efdb88794
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/272273
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Ed Schiebel <eschiebel@instructure.com>
QA-Review: Ed Schiebel <eschiebel@instructure.com>
Product-Review: Peyton Craighill <pcraighill@instructure.com>