Add ability for LTI tools to send toast messages
This adds the ability for LTI tools to send toast alert messages through the
Canvas alert system.  This enables them to get the benefit of having a
fully integrated alerting solution including screen reader
announcements.  It also enables messages to be visible regardless of
scroll location within the external tool's iframe.

To facilitate ease of knowing where the message is coming from, the name
of the external tool is prepended to the message sent by the tool.

Test Plan:
  - Have an LTI tool send a "lti.alertMessage" message
  - The alert should show as "Tool Name: Alert Message Text"