Replace i18n locale fallbacks with something smarter
Refs FOO-2330
flag = none

The default locale fallback behavior in the i18n gem
is not very smart... it just pulls things off the back
of the locale one by one. This isn't necessarily what we
want in Canvas, however, since we rely on the private
extension `x-k12` to signify some of our special locales.
So for instance we would want `sv-FI-x-k12` to fall back
first to `sv-x-k12` and then to `sv`, but that is not what
the gem's fallback would do.

This replaces the call to the gem's `self_and_parents` method
with a call to a smarter fallback list generator. We couldn't
really modify the existing one since it is pretty hard-coded
to implement a linked list of class instances.

Test plan:
* in Rails console, try running I18n.fallbacks[x] for various
  locales x, and see if the returned list of fallbacks makes
  sense. Some things to try:

            de
            de-AT
            sv-FI-x-k12
            en-US-POSIX
            de-u-hc-h24
            zh-Hant-x-k12
            en-Latn-US-vary1-vary1-u-hc-h24-x-k12

Change-Id: I9cc05d4d275032a834c0aba3064b2daaecd0e9b9
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/273169
Reviewed-by: Cody Cutrer <cody@instructure.com>
QA-Review: Cody Cutrer <cody@instructure.com>
Product-Review: Cody Cutrer <cody@instructure.com>
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>