Sync term, start and end dates with homeroom
The course participation section is now synced with the homeroom
participation if the option “Sync enrollments and course start/end
dates from homeroom” is checked. If the homeroom participation option
is set to course, the course start_at and conclude_at dates will be
synced with the ones in the homeroom, and if the participation is
set to term, the same enrollment_term will be set to the course too.

Closes LS-2580
flag=none

Test plan:

- Set a k5 course as a homeroom
- Set the homeroom participation to term
- Create a course with the option “Sync enrollments and course
start/end dates from homeroom” checked
- Expect to the new course has the participation option set to term
and the term be the same as the homeroom
- Go to the homeroom settings and change the participation option to
the course and select the start and end dates
- Go to the settings page of the course you created and click update
- Expect the participation option to be set to course and the dates
match with the ones in the homeroom

Change-Id: Idea7b476ad818f9777fc8ff2ca89c7b024b6b6ce
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/273478
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Ed Schiebel <eschiebel@instructure.com>
QA-Review: Ed Schiebel <eschiebel@instructure.com>
Product-Review: Peyton Craighill <pcraighill@instructure.com>