Fix regression on am/pm input in jquery timepicker
Refs FOO-2347
flag=none

The Intl.DateTimeFormat representations for AM and PM
are in upper case, while tz.format liked to use lower
case for them. The literal value of that is used to
pick the default value for the am/pm selection box in
the time portion of the jquery datepicker, and since
it die not match exactly, defaulted to blank, which for
some reason is interpreted as PM on select.

Also fixes a failing spec because of work done entirely
when the day of the month had two digits, and Oct 1
does not equal Oct 01.

Test plan:
* be in a locale that uses am/pm (en works fine)
* visit /courses/:id/assignments/:id/edit for some
  course and assignment you have
* in the "due date" near the bottom, select some date
  and time that is in the AM, then exit the picker
* the formatted date and time should remain in the
  field and be correct (i.e. in the AM)
* go back into the picker for that field
* Note that "AM" correctly appears in the time picker
  portion at the bottom
* Select a different date and exit the picker
* Now repeat the above for a PM time to verify that
  that still works as well
* Now repeat all of the above for a locale that does
  not use am/pm (suggest fr or de), and verify that
  selecting times on the 24-hour clock still works too

Change-Id: I56f586fceff23df633b89f985479518b687f2cfb
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/274942
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
QA-Review: Charley Kline <ckline@instructure.com>
Product-Review: Charley Kline <ckline@instructure.com>
Reviewed-by: Simon Williams <simon@instructure.com>