Add awaitElement helper
Refs FOO-906

Sometimes a DOM element doesn't quite show up in time for
you to try to use it. This allows one to very efficiently
wait for an element with a given ID to appear.

It returns a Promise which resolves to the new element
itself when it finally appears, or rejects if it did not
show up in the DOM for a configurable timeout period.

Test plan:
* Unit tests for this pass

Change-Id: I3f846ea50dd40df096ec3a2fa8c0aeb718f5680d
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/279756
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Ahmad Amireh <ahmad@instructure.com>
QA-Review: Charley Kline <ckline@instructure.com>
Product-Review: Charley Kline <ckline@instructure.com>