Fix planner translation snafu
This started out confirming that the planner supported Castilian
Spanish, but morphed into something a bit broader.  The code
that loads the translation json files based on the user's locale
had a bug that it wouldn't map locales using '-' to their corresponding
file names uising '_' (e.g es-ES => es_ES.json).

Once that was fixed I noticed that the planner was only partially
translated until it updated/rerendered. This is because importing
the translation json is async and the code loading the planner wasn't
waiting on it or general planner initialization. Now that code
initializing the planner waits for it to complete, the planner is
translated on first render.

This exposed an issue where code calling initializePlanner wasn't
waiting for the returned promise to resolve before doing more work.
That's fixed too.

closes LS-2738
flag=none

test plan:
  prereqs
  1. build with the js with `RAILS_LOAD_ALL_LOCALES=1`
  2. run your canvas server with `RAILS_LOAD_ALl_LOCALES=1`

  - change your user's language to Espanol (espana)
  - open the student planner
  > expect it to be in Spanish.

  - open a course page
  > expect the To Do sidebar to load

  - turn on Canvas for Elementary
  > expect the Schedule tabs to work

Change-Id: I0da78e0c14000496044e37ea287a48218262bb0f
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/279861
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Jackson Howe <jackson.howe@instructure.com>
QA-Review: Jackson Howe <jackson.howe@instructure.com>
Product-Review: Ed Schiebel <eschiebel@instructure.com>