Fix joinining on nil in valid_route?
For a root path of `/` we were calling `path.chomp("/")` to remove
trailing slashes which would leave us with an empty string for the path.
Then when we tried to join in `Rails.root.join` we were attempting to
get the substring of `path[1..]` which resulted in `nil` and threw an
error.

Now we will clean the path using Rails' `normalize_path` method to remove
trailing slashes and get it ready for matching against our routes and
for the join we use the safer `path.delete_prefix("/")` to return an
empty string instead of nil in the case of a single slash making it in.

fixes LS-2903
flag=none

test plan:
- In the rails console, create a CourseLinkValidator object with a course
- Call `course_link_validator.valid_route?('/')` and see that it is true
- Call `course_link_validator.valid_route?('/favicon.ico')` and see that
it is true
- Try other GET verb paths from routes and other public files to verify
they return true
- Try random incorrect strings to verify they return false

Change-Id: If282b234a67f0e26fb31950aa5d9ec5571c2d29c
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/281427
Reviewed-by: Jeremy Stanley <jeremy@instructure.com>
QA-Review: Jeremy Stanley <jeremy@instructure.com>
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Product-Review: Eric Saupe <eric.saupe@instructure.com>