fix remove missing status speedgrader bug
There was an edge case where the missing status was not removed until a
full page refresh. When a teacher gives a grade and the focus does not
leave the grade box before moving to the next student, the missing is
not taken off unless the page is refreshed. If the focus is taken away
from the grading box before moving to the next student, then the missing
status is removed as expected.

closes EVAL-2158
flag=remove_missing_status_when_graded

Test Plan:
    - Have a teacher with some students in a course with an assignment
    that is missing
    - Navigate to that assignment in the speedgrader
    - Give a grade and navigate to the next/prev student before taking
    away focus from the grade input
    - Then navigate back to the graded student and notice that the
    missing status is removed and the grade given is shown
    - Test this with the `edit_submission_status_from_speedgrader` FF
    on and off and ensure the mount point of the status pill isn't
    jumping around when navigating between students
    - smoke test the edit_submission_status_from_speedgrader menu and
    ensure it still works as expected
    - smoke test the remove missing status functionality outside of this
    edge case and ensure it works as expected

Change-Id: I6835c101b45b1a3658f16420ca4613a332016a0e
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/282115
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Dustin Cowles <dustin.cowles@instructure.com>
Reviewed-by: Aaron Shafovaloff <ashafovaloff@instructure.com>
QA-Review: Aaron Shafovaloff <ashafovaloff@instructure.com>
Product-Review: Syed Hussain <shussain@instructure.com>