Microsoft Sync: Cross-shard communication channels
Currently, when we look up email addresses for a user, we just look on
the shard the sync is happening on (the shard the course is on).
CommunicationChannels almost live on the user's home shard, so if there
are cross-shard users, we need to activate their shards before looking
for the CommunicationChannel.

I also modified a spec helper because it was not detecting an error's
call to I18n.t correctly for some reason for my local MRA setup. (I
checked that the new helper also fails when the error class doesn't
actually call I18n.t)

closes INTEROP-7147
flag=none

Test plan:
- Have MRA Canvas set up
- Ensure you have the Microsoft Sync creds set in
  config/dynamic_settings.yml
- One on shard, set up Microsoft Sync in the root account and in a
  course.
  - turn on feature flag
  - in account settings enable sync with our test tenant and use "email"
    and "UPN" as the local and remote lookup fields
  - in Course Settings -> Integrations turn on sync for the course
- Add two users to the course, one in another root account. At least one
  should be a teacher. Both should have emails corresponding to users in
  our Microsoft test tenant. Set the enrollments' workflow_state to
  "active".
- Make communication channels for each users in their respective shards.
  Make sure each shard has only the communication channel for the user
  local to that shard, i.e., there are no communication channels with
  user_id > Shard::IDS_PER_SHARD. Confirm the communication channels:
  run `communication_channbel.confirm` on each, which should set the
  workflow_state to "active".
- Sync the course
    course.microsoft_sync_group.syncer_job.run_synchronously
- The sync should succeed and the users should be visible in the
  Microsoft admin console

Change-Id: Ic87be3d9dfe0578370f2f04e8509361b75269961
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/283233
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
QA-Review: Xander Moffatt <xmoffatt@instructure.com>
Product-Review: Evan Battaglia <ebattaglia@instructure.com>
Reviewed-by: Mysti Lilla <mysti@instructure.com>