Switch dashboard sidebar on observed user change
The sidebar api was previously updated to support an observed_user
param. This change implements the switching in the UI. The currently
selected student's data is prefetched and loaded in the sidebar. When
the observer picker is updated, fetch the new student's data and show it
in the sidebar.

Note: the student todo list is not supported for observers until the
observer planner work is complete.

closes LS-2967
flag = observer_picker

Test plan:
 - Enable observer_picker flag and open dashboard as an observer with
   several linked students
 - Open network requests developer pane; expect to see only one request
   to /dashboard-sidebar
 - Expect to see the correct data in the sidebar for the currently
   selected student
 - Switch the student
 - Expect the sidebar to reload with correct data
 - Expect to see one new /dashboard-sidebar request
 - Switch back to the original student and expect the sidebar to show
   their data once again
 - Disable observer_picker flag and expect the dashboard to work as
   before

Change-Id: Id8a078e0510ef54cd98279992406fce4bff9c9c5
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/283955
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Robin Kuss <rkuss@instructure.com>
QA-Review: Robin Kuss <rkuss@instructure.com>
Product-Review: Jackson Howe <jackson.howe@instructure.com>