Make InstUI render links, not buttons, in the help menu
Fixes FOO-2779
flag=none

InstUI's <Link> component automatically adds `role="button"` whenever
an onClick prop is sent to it, but for screenreader purposes those help
links need to read out as links.

While we investigate a way to control the behavior of <Link> regarding
the element role, for now the solution is to wrap the <Link> in a <span>
and put the click action on the span.

Test Plan:
- See the steps to reproduce in the ticket
- All the links in the help menu should get read out by a
  screenreader as "Link" not "Button"

Change-Id: Ia2493e89f6816696d7a46bebe958377410fa53a3
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/289242
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: August Thornton <august@instructure.com>
Product-Review: Charley Kline <ckline@instructure.com>
QA-Review: Charley Kline <ckline@instructure.com>