Make API paginator and pagination UI smarter
Fixes FOO-2478
flag=none

To improve database performance, Rails no longer allows the
Folio paginator to retrieve the total count of objects in
the collection it is paginating. Thus, it no longer knows
what the last page is, so it tries to guess by asserting
that if the query got back fewer records than the per-page
value, then this must be the last page. However that logic
fails when the last page contains exactly per-page records.
In that case the returned collection shows a next-page but
no last-page; however, trying to retrieve that next page
returns an empty collection. This confused the front-end
paginators quite a bit.

So here we modify the API helper so that if the last page
is unknown (because it wasn't allowed to do a count), then
also retrieve the NEXT page beyond the one just requested.
If that page has fewer than per-page records, then it must
be the last page, and we indicate so.

Also improved the behavior of the User and Course list
pagination UI so that it understands when the last page
has been discovered, so it can be remembered in the display.

Test plan:
* See the steps to reproduce in the ticket
* That behavior should no longer happen
* Also check out accounts/self/users and play around with
  that pagination as it was also modified
* In general the pagination behavior in the user interface
  should be looking and working better

Change-Id: I72ccd4b1e9e90ac6186277068864fa36886700dc
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/289772
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Jacob Burroughs <jburroughs@instructure.com>
Product-Review: Charley Kline <ckline@instructure.com>
QA-Review: Charley Kline <ckline@instructure.com>