Use enrollment created_at if start_at is not present
The enrollment start_at is often not set and can cause confusing
due dates to be created based on the effective_start_at method.
With this change we check that method and the created_at date of the
enrollment and take the more recent of the two. This fits more with
the idea that a student is "enrolled" when either the enrollment was
created or the course/term/section started.

fixes LS-3146
flag=course_paces

test plan:
- Create a course that started in the past with a course pace
- Add students using the people picker
- Verify the assignment overrides created are based on the enrollment
created_at date
- Update the start_at time on one of the enrollments
- Republish the course pace
- Verify the assignment overrides for that enrollment were based on the
start_at date

Change-Id: Id12029ca4ff470a66b8ef41c793c1684a42367f5
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/290497
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Ed Schiebel <eschiebel@instructure.com>
QA-Review: Ed Schiebel <eschiebel@instructure.com>
Product-Review: Eric Saupe <eric.saupe@instructure.com>