Fix mis-recognized PM time indicator in en-CA
Fixes FOO-2919
flag=none

In the en-CA (Canadian English) locale, the Canvas locale
files list the representation for post meridian (i.e. PM)
times as "pm" but the Unicode CLDR standard (used by
browser Intl.DateTimeFormat) has "p.m." This made the two
not compare equal, so PM times entered into the date-
picker when in the en-CA locale got misread and converted
back to AM.

It appears that the Unicode standard is actually correct
so this was updated in the locales.yml file. The code
that compares the two to determine whether the selected
time in the datepicker was also made a little more sturdy
and thus less likely to make this mistake again.

Checking the other locales that use a 12-hour clock (and
there are surprisingly few), this tactic of ignoring a
handful of punctuation characters when comparing a day
period supplied by Intl.DateTimeFormat to the Canvas
locale files will always work. (So this should fix up
Greek and Korean as well, if they were to have the same
issue.)

Test plan:
* See the steps to reproduce in the ticket
* A time in "p.m." should not spuriously flip back to
  "a.m." when the picker is closed

Change-Id: Ia41e84db09dd60fd18b72ee99a9d726ceeb4394a
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/291623
Reviewed-by: Ahmad Amireh <ahmad@instructure.com>
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Product-Review: Charley Kline <ckline@instructure.com>
Product-Review: Ahmad Amireh <ahmad@instructure.com>
QA-Review: Ahmad Amireh <ahmad@instructure.com>