Mark CMPs outdated when updating assignment dates
The module progressions need to be re-evaluated if an assignment's
due date information is updated so that an assignment which is not
assigned to a student won't block their progression within a module.
This change will mark associated progressions as outdated if an update
is made to an assignment's due date info.

closes LS-3196
flag = none

Test plan:
 - Create a course with 2 sections
 - Enroll a student in section 1
 - Create a module; add 2 assignments
 - Set module requirements so that both assignments are required to be
   submitted
 - As the student, submit the first assignment
 - As the student, visit the modules index page
 - Expect the module to be "Started" (icon at top of module)
 - As the teacher, edit the second assignment - assign it to just the
   second section
 - As the student, visit the modules index and expect the module to be
   completed (green checkmark)

Change-Id: I341680f430d94e3a806a5f4451b180c36359cb9a
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/293044
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Luis Oliveira <luis.oliveira@instructure.com>
QA-Review: Luis Oliveira <luis.oliveira@instructure.com>
Product-Review: Jackson Howe <jackson.howe@instructure.com>