AGS: clarify error when file url times out
why:
* when the file upload url given in a content item times out or takes
too long to respond (>30s), the error returned by the file service is
obtuse and not properly parsed and returned by the Scores API
* providing a more clear error can make it much easier for tools to
understand why the request is failing

breaking changes:
before: An AGS Scores API request that includes `content_items` in the
  canvas extension with a file upload url that times out or takes longer
  than 30 seconds to respond would fail with a 500 and
  `internal_server_error` code, and an unhelpful message that contained
  HTML referencing a 502 Bad Gateway
after: An AGS Scores API request that includes `content_items` in the
  canvas extension with a file upload url that times out or takes longer
  than 30 seconds to respond will fail with a 504 status code and a
  messages that makes it clear that the file url timed out.

note:
* InstFS currently returns a 502 Bad Gateway for these errors, though
that is a bug and it means to actually return a 400 Bad Request with a
body that makes it clear the file upload url timed out.
  * basically the problem is that the timeout circuit breaker in InstFS
  (`src/api/router.js:211`) uses `req.setTimeout` instead of `res.
  setTimeout`, which just cuts off the request without sending anything
  to finish the response, resulting in a 502 Bad Gateway. It actually
  means to return a 400 Bad Request.
  * simply switching this out results in some express errors about
  "sending headers after the response has closed", meaning there are
  some control flow bugs to work out to make sure that the rest of the
  request handler doesn't keep going

todo:
* Should we fix the InstFS bug detailed above ourselves, or open a
ticket with the team?
* which HTTP status code is best?
  * 504 Gateway Timeout implies that an upstream server (aka the file
  upload url) timed out and was out of the Canvas server's control.
  * 422 Unprocessable Entity implies that the
  request body (including the file upload url) was formed correctly but
  not able to be processed (since the file upload failed).
  * 400 Bad Request is what InstFS _should_ send back, so it may make
  sense to forward that on to the client.
* what message should show to make sure tools know that the file upload
url timed out and that it's the file upload server's fault, not Canvas's
or InstFS's?
  * currently it says "file url timed out"

closes INTEROP-7339
flag=ags_scores_file_error_improvements

test plan:
* follow the instructions in doc/lti_manual/11_testing.md to set up AGS
Scores API requests, and use the Scores page of the Canvas API Doc to
construct the request body.
* somehow acquire a file url that takes longer than 30 seconds to finish
responding - I have a tiny local express server that sleeps for 30s
before delivering some text.
* set up InstFS locally and configure it to work with Canvas - confirm
that it is working and test it with an AGS Scores request that uses a
file upload url that works
* in InstFS, open `src/api/router.js` and edit line 211 to set a timeout
of 5000ms instead of 30000, to avoid tripping Canvas's timeout breaker
* with the FF off
* send a Scores request that uses your file url that takes 30s
* the request should fail with a 500 Internal Server Error, and the
message should contain HTML from InstFS that shows a 502 Bad Gateway
* with the FF on
* repeat the same request
* it should fail with a 504 Gateway Timeout and the message should be
"file url timed out"
* test the intended InstFS response by changing line 211 from `req` to
`res.setTimeout` (see note section above for explanation). InstFS will
now return a 400 instead of a 502.
* repeat the same request, and it should fail with the same 504

Change-Id: I184fcb862b8946dbe13503d5d786daf32f4843b2
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/293130
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Evan Battaglia <ebattaglia@instructure.com>
QA-Review: Evan Battaglia <ebattaglia@instructure.com>
Product-Review: Xander Moffatt <xmoffatt@instructure.com>