Make calendar event titles better
This PS adds a tooltip if the title does not fit in a line.

closes LS-3191
flag = wrap_calendar_event_titles

Test Plan:
- In the calendar create a calendar event with a long title that
doesn’t fit in the event slot
- Expect the title to be cut at the end of the line and 3 dots to
appear
- Place the mouse over the calendar event and expect a tooltip to
appear with the entire title
- Expand and shrink the window and expect the tooltip to appear
when the title doesn't fit in the event container and not to be
displayed if the title is fully shown.

Change-Id: I828eb1d9feaad4b011d68475dea7a88b8d91fbbc
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/293159
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Product-Review: Allison Howell <allison.howell@instructure.com>
Reviewed-by: Ed Schiebel <eschiebel@instructure.com>
QA-Review: Ed Schiebel <eschiebel@instructure.com>