compute total pages count for UI pagination
This change will help prevent doing an n+1 of an expensive
query to fetch the next page in order to find the last page.
We are computing the count ourselves so we can take advantage of
preloading the records in certain circumstances and so we can
only run the count if the UI calls for it. Folio automatically
counts the number of items in the folio for us if we don't
explicitily set the :total_entries page option to `nil`.

closes FOO-2931
flag = none

Test plan:
• You should *not* see the behavior outlined in FOO-2478
• Exercise both /accounts/1/users & /accounts/self api
  index actions and verify pagination works as expected
  (now /accounts/1/users will be populated with the last page
   since we compute the value and our UI sets it)

Change-Id: I93e01f2ab3708e631be4c698486407b4347279fd
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/293447
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Charley Kline <ckline@instructure.com>
QA-Review: Charley Kline <ckline@instructure.com>
Product-Review: August Thornton <august@instructure.com>