Gradebook Previous Export Time
When user modifies (deletes, renames, or moves) the gradebook export,
the previous export date/time shows when the file was last modified
rather than when it was created. Fixed the issue to where now it
shows the created date/time instead of the updated time.

flag = none
Fixes EVAL-2469

Test Plan:
1. Create a course with at least one student
2. Create an assignment in the course and publish it
3. Go into the gradebook and export gradebook
4. Take note of the export time
5. Go to Account -> Files -> Unfiled and wait a few minutes
6. Rename, delete, or move the exported file.
7. Go back to the gradebook and under Actions,
view that the previous export time was when the file was created
and not when you edited(renamed, deleted, or moved) the file.
8. Repeat steps 3-7 using the Enhanced Gradebook Filters flag set

Change-Id: I8e6d5ab597c835281bfe907c73ce6dd89a8ef035
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/294877
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Product-Review: Jody Sailor
Reviewed-by: Spencer Olson <solson@instructure.com>
Reviewed-by: Kai Bjorkman <kbjorkman@instructure.com>
QA-Review: Eduardo Escobar <eduardo.escobar@instructure.com>