Uprev @instructure/updown
Refs FOO-2801
flag=none

Turns out the updown package used a JS feature that is
not 100% supported by all the supported Canvas browser.
That was fixed in the package and this bumps the dep
version to pull it in to Canvas.

Test plan:
* If you were getting "TypeError: Object.hasOwn is
  not a function" you should not get it any more.

Change-Id: Ibdaf313492c762a0b65143269be0054f06f4b908
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/296686
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Jacob Burroughs <jburroughs@instructure.com>
QA-Review: Charley Kline <ckline@instructure.com>
Product-Review: Charley Kline <ckline@instructure.com>