Mop up InstUI 8 prop deprecations for <List>
Refs FOO-3064
Refs FOO-2744
flag=none

The <List> component used to accept a variant="unstyled" prop
but that was replaced with a Boolean isUnstyled prop instead,
so this has to be fixed before the InstUI 8 upgrade. Fortunately
this is a really easy one to automate because the new prop is
a drop-in functionally-equivalent replacement for the old one.

Test plan:
* Maybe spot-check a couple of the touched UI elements to be
  sure they still look good?

Change-Id: Idb05b6b686ba4f3a0b3e5328d8956082dba66416
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/299135
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
QA-Review: Charley Kline <ckline@instructure.com>
Product-Review: Charley Kline <ckline@instructure.com>
Reviewed-by: Sean Scally <sean.scally@instructure.com>