Correct API docs for DELETE /accounts/:id/admins/:id
Fixes FOO-2995
flag=none

The API docs for deleting an admin role are misleading as they
don't highlight the fact that the role_id of the account admin
being deleted must be specified as a param in the DELETE API
call. The default is the built-in base role AccountAdmin which
probably will work in most cases, but if the admin being deleted
has a custom role that's merely derived from AccountAdmin, it
has to be made explicit.

The easiest way to avoid confusion is simply to make the `role_id`
param on the DELETE call required, so that it must always be
specified for the call to perform as desired.

Test plan:
* none, comment changes only
* you agree with the changes to the API docs

Change-Id: I2802e33968ce317ddb25f7d081a3a3ac15eed378
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/300696
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: August Thornton <august@instructure.com>
Product-Review: Jesse Poulos <jpoulos@instructure.com>
QA-Review: Charley Kline <ckline@instructure.com>