make tii graphql type aware of error/status keys
This tii graphql interface was first implemented 3 years ago and has
sat stale for a few years until recently when we allowed for
orignality reports to be viewed within assignment enhancements. It was
not written to be aware of either 'status' of 'student_error' as
possible keys set by the plagiarism service. (these are only set when
the turnitin product for this account has expired) This was causing us
to try to load an object via an asset string that did not exist.

flag=assignments_2_student
refs EVAL-2610

Test Plan:
    -specs pass

Change-Id: I7c9b622db9c3ea73c472f66bef26def044c9a37e
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/301568
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Christopher Soto <christopher.soto@instructure.com>
Reviewed-by: Eduardo Escobar <eduardo.escobar@instructure.com>
Reviewed-by: Spencer Olson <solson@instructure.com>
QA-Review: Spencer Olson <solson@instructure.com>
Product-Review: Spencer Olson <solson@instructure.com>