Ensure ujsLinks runs before retriggerEarlyClicks
Some of our buttons (many created by link_to helper) have data-method
attributes that allow a button to submit a post/delete request instead
of a get request. A snippet in ui/boot/initializers/ujsLinks is
required to run to make this possible. If the user clicks on a button
while the js is still loading (particularly, if they click before
ujsLinks runs), then a button with a data-attribute other than 'get'
won't be respected, and the user might be directed to a nonexistant
page. We have the _earlyClick/retriggerEarlyClicks functions to detect
clicks that happen before the page loads and replay them later, but
retriggerEarlyClicks (which removes the earlyClick listener) sometimes
runs before ujsLinks. If the user clicks on a button in this window,
then the click won't be intercepted and the user might end up on a
nonexistant page.

This commit moves the ujsLinks initializer earlier in the frontend
setup to guarantee that it runs before the earlyClick listener is
removed.

closes LS-3498
flag = none

Test plan:
 (You might want to try reproducing this without the fix first, since
 the timing can be finnicky)
 - Open a course
 - While the page is still loading, click the student view button
 - Expect student view to load (or nothing to happen at all, just not
   to see a 404 error)
 - In student view, refresh the page and click the leave student view
   button while the page is still loading
 - Expect no error here either

Change-Id: Ib4f6fdcf28f1d7b05887c290d424d55428ddaa8c
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/302732
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
QA-Review: Davis Hyer <dhyer@instructure.com>
Product-Review: Davis Hyer <dhyer@instructure.com>
Reviewed-by: Davis Hyer <dhyer@instructure.com>