Fix contrast issues with CondensedButton in some of shared
Refs FOO-3129
flag=none

The use of InstUI <CondensedButton> to replace the
deprecated <Button variant="link"> created some issues
with customers who decided to invert the contrast of
their "Primary Button" style in the theme editor. To
wit, if the Primary Button is changed from something
light-on-dark to something dark-on-light, then an instance
of a CondensedButton would have very poor contrast, as
its text is rendered in the color of the background of a
regular Button.

Test plan:
* hand-check the touched components in the Gradebook,
  Rubrics, and FlashAlert to be sure they look okay

Change-Id: I057729e2641e030d13382c8fdeb995cb5e75ed86
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/303343
QA-Review: Charley Kline <ckline@instructure.com>
Product-Review: Charley Kline <ckline@instructure.com>
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Sean Scally <sean.scally@instructure.com>