Fix contrast issues with CondensedButton in discussion_topics_post
Refs FOO-3129
flag=none

The use of InstUI <CondensedButton> to replace the
deprecated <Button variant="link"> created some issues
with customers who decided to invert the contrast of
their "Primary Button" style in the theme editor. To
wit, if the Primary Button is changed from something
light-on-dark to something dark-on-light, then an instance
of a CondensedButton would have very poor contrast, as
its text is rendered in the color of the background of a
regular Button.

Test plan:
* Most of the touched components have Storybooks that can
  easily be checked to verify that the UI is preserved;
  for any that aren't, hand-checking might be warranted

Change-Id: I1943f233b3431a0dbd95855eeea91e71218f6707
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/303540
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: August Thornton <august@instructure.com>
QA-Review: Charley Kline <ckline@instructure.com>
Product-Review: Charley Kline <ckline@instructure.com>