Fix contrast issues with CondensedButton in ui/features
Refs FOO-3129
flag=none

The use of InstUI <CondensedButton> to replace the
deprecated <Button variant="link"> created some issues
with customers who decided to invert the contrast of
their "Primary Button" style in the theme editor. To
wit, if the Primary Button is changed from something
light-on-dark to something dark-on-light, then an instance
of a CondensedButton would have very poor contrast, as
its text is rendered in the color of the background of a
regular Button.

Test plan:
* Maybe hand-check the affected components to make sure they
  look okay still? These are pretty straightforward changes
  that have been done a lot already, so there's pretty high
  confidence

Change-Id: I7d8d9b2f0af4dbb4870cb37271d395083d799309
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/303636
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: August Thornton <august@instructure.com>
QA-Review: Charley Kline <ckline@instructure.com>
Product-Review: Charley Kline <ckline@instructure.com>