Add heap analytics snippet
This depends upon a DynamicSettings setting that is not yet defined,
so expectation at this point is that this is a no-op.

I expect that after we define a setting, this will use that as the heap
APP_ID and then load the snippet into environments where it is defined.

refs FOO-3172
flags=none
test plan=
1. canvas loads with no changes without this setting defined

Change-Id: Idcc0980271587a84652f204babca78bc283c1972
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/304129
Reviewed-by: Jacob Burroughs <jburroughs@instructure.com>
QA-Review: Sean Scally <sean.scally@instructure.com>
Product-Review: Sean Scally <sean.scally@instructure.com>
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>