keep response formats consistent when nesting respond_to
Rails determines the desired response format from the
HTTP Accept header submitted by the client. when that header
is set to accept all formats i.e. "*/*" the default behavior
for respond_to is to always choose the first response format.
that gets tricky when you have nested respond_to, because
rails will only match on a single format per action.

The bug in this case was that we called the first respond_to in
ApplicationController#require_context_with_permission with the
default format of :html, and then the logic path for users without
a session was to follow to AuthenticationMethods#redirect_to_login
where the first format within respond_to was changed to :json.

*boom*

We now specify a common block for different formats that we want
to respond to in a similar fashion using .any; and as the first
format, it gives us some redundancy in nested respond_to blocks.

fixes FOO-3023
flag = none

Test plan:
 • specs

Change-Id: I56650acc1633f8839f24cbc7135faf96976ed763
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/304239
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Jacob Burroughs <jburroughs@instructure.com>
QA-Review: August Thornton <august@instructure.com>
Product-Review: August Thornton <august@instructure.com>