remove prop requirement for ObserverOptions
refs LS-3571
flag=none

The only instance of the prop being used in the component already has
null safety baked in. We have one instance of this component that
doesn't provide said prop so we are opting to match the expectation
of the component and usage of it by removing the requirement.

test plan:
  - tests pass

qa risk: low

Change-Id: I8c3b0e02ed93b48f75595057b37db2df3d5fae09
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/304578
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Robin Kuss <rkuss@instructure.com>
QA-Review: Robin Kuss <rkuss@instructure.com>
Product-Review: Davis Hyer <dhyer@instructure.com>