Add an enhanceUserContent option to open canvas links in a new window
Adds the new canvasLinksTarget value to the opts argument to enhanceUserContent
When set, links to canvas resources will have their target attribute set
to this value if (1) there is currently no target or (2) the current target
is "_blank".  Some fun values are
 _blank: links will open in their own new tab
 _self: don't do this, it defeats the purpose
 anything_else: clicks on links will open in the same new tab

The purpose of this is so NQ can set a value so links don't open
canvas inside the NQ lti's iframe.

Because file links have onclick handlers that override the default
behavior and open the file preview in an overlay or inline, this
option has no effect on them.  This is mostly for links to
other stuff, like assignments, and pages, etc.

closes MAT-1099
flag=none

test plan:
 Without sending the new option
 > expect links to course resources (assignments, pages, etc) to open
   in the same browser tab

 With sending the new option (best tested in NQ.)
 This is going to be awkward since it requires sending an argument
 to enhanceUserContent. What I would do is:
 - set a conditional breakpoint at the first line of enhanceUserContent
   that looks something like
   opts.canvasLinksTarget = 'xyzzy', 0
   (the ,0 keeps it from actually stopping)
 - create RCE content with a canvas file link and a canvas resource link
 - save
 - in the saved page, click on the links
 > expect the file to be previewed
 > expect the canvas resource to open in a new tab
 - if you have >1 canvas resource link, and set canvasLinkTarget to
   something other than _blank, expect clicking on them to open
   in the same tab.

Change-Id: I9895036c61e4b7eba3d4b409a8932356a3dfa2da
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/305404
Reviewed-by: Jon Scheiding <jon.scheiding@instructure.com>
QA-Review: Jon Scheiding <jon.scheiding@instructure.com>
Product-Review: Ed Schiebel <eschiebel@instructure.com>
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>