Always use native fullscreen
This looks big, but it's not as bad as you might think.
By default all INSTUI popups (Tray, Modal, Select, ...) mount
to a portal added to the bottom of the document's body. When
the RCE is in fullscreen, that mount node isn't visible, so
we have to tell INSTUI where to mount popups using the mountNode
property. There are a lot of popups in the RCE.

Similarly for tinymce, but it was easy to simply move its
mount point when in fullscreen.

The LTI modal always does a window.confirm when closing. This
kicks you out of fullscreen. We should create a blocking modal
to replace window.confirm w/in canvas and that would resolve
this annoyance.

The change to package.json is to add a missing dependency in
src/rce/plugins/instructure_icon_maker/components/CreateIconMakerForm/ImageSection/ModeSelect.js

Added: disable fullscreen mode on iPhones and iPads since
       support there is sketchy and we don't look good

closes MAT-941
flag=none

test plan:
  To test that the a11y checker works in fullscreen
  you'll have to pull the latest master and copy it
  into canvas' node_modules, something like
  $ cp  -R ../tinymce-a11y-checker/lib node_modules/tinymce-a11y-checker
  $ yarn build:js
  (I could not get yarn link to work)

  - open the rce
  - click the fullscreen button in the status bar
  > expect the rce to go fullscreen
  - click the fullscreen button
  > expect the rce to return and be the same size as before

 - repeat but be in html edit view to start
 - repeat but switch views while in fullscreen
 > always expect the current view to be the right size after
   restoring out of fullscreen

 - while in fullscreen open various trays and modals
   - content tray, a11y checker, external link, etc.
 > expect them to show up
 note: if you choose to pick a file from the finder, opening
 the finder will kick you out of fullscreen

 - with lti tools installed, click on the Apps toolbar button
   and open an lti tool
 > expect the tool to initialize and be visible
 note: when you exit the tool you'll get a window.confirm
 popup which kicks you out of fullscreen.

 *** MOBILE ***
 - load a page with the RCE on an iPHone
 > expect no fullscreen button in the status bar

 - load a page with the RCE on a mobile browser
   (android phone and an iPad)
 - tap the fullscreen icon
 > expect the RCE to fill the screen
 - make the keyboard come and go, rotate the device
 > expect to be able to see the status bar
 - click the fullscreen button
 > expect the RCE to return to normal size

 EXCEPTION:
 - load the RCE on an iphone (simulator works)
 > expect the fullscreen button to be omitted
   from the status bar (because fullscreen is
   not supported on the iphone)

 - load the RCE on a real iPad (the simulator
   will not do)
 > expect the fullscreen button to be omitted

Change-Id: I4c25dc4f06b2d0c19f44c33fa324274fd47235ed
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/305819
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Jacob DeWar <jacob.dewar@instructure.com>
QA-Review: Jacob DeWar <jacob.dewar@instructure.com>
Product-Review: David Lyons <lyons@instructure.com>