fix firefox date input position
when adding an available and until date to an
assignment in firefox the cursor will be at the end
of the textbox. this is causes users not to be able to
see the date without some manual steps. this fix will
set the cursor to the start of the date so the user can
see the whole date

fixes EVAL-2737
flag=none

test plan:
- create an assignment in firefox
- add a date to the Available From date picker
- verify after clicking off that you can see the
  beginning of the date in the textbox
- do the same prev 2 steps with the Until date picker

Change-Id: I62435b046a12a511918176fe2c56e06b99d82f49
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/306265
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Eduardo Escobar <eduardo.escobar@instructure.com>
Reviewed-by: Kai Bjorkman <kbjorkman@instructure.com>
QA-Review: Eduardo Escobar <eduardo.escobar@instructure.com>
Product-Review: Deborah Kwak <deborah.kwak@instructure.com>