Fix pace modal navigation flow for Chrome VO
For some reason, the Modal.Body component has a tabindex=-1 that
prevents the body from being accessible with the keyboard. It looks
like this is already on the instui team's radar and will be fixed in
future releases. But for now, this workaround makes the body be part
of the navigation flow

closes LS-3640
flag=course_paces_redesign

Test plan:
- Using Chrome go to the course pace landing page
- Open the default course pace modal
- Navigate with control-option-right arrow
- Expect to be able to navigate through the modal body

Change-Id: I99d6bf81fc46e55a51766d58e926deb9a29ce7c8
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/306588
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Luis Oliveira <luis.oliveira@instructure.com>
QA-Review: Luis Oliveira <luis.oliveira@instructure.com>
Product-Review: Jonathan Guardado <jonathan.guardado@instructure.com>