Filter authenticity_token field when disabling
The authenticity_token is changed from a document event. The disabling
of all input fields causes the form value to be locked in to a possibly
invalid authenticty_token when submitted because the disabling runs
before the document event is able to change it. By leaving this field
available it accepts the change from the document event.

See ui/boot/initializers/injectAuthTokenIntoForms.js

fixes LS-3661
flag=none

test plan:
- Go to the dashboard
- Click the Start a New Course button
- Open the browser developer console
- Find the authenticity_token field in the #new_course_form
- Submit the form
- Verify the field was never disabled
- Verify the course was created

test plan in production:
This is not causing an error locally but is happening in production. To
verify this fix works in production go to the link provided on the ticket
where the problem is seen by impersonating the user.
- Click the Start a New Course button.
- Open the browser developer console.
- Place a breakpoint after the elements are disabled around
jquery.disableWhileLoading.js:77.
- Submit the form.
- Find the authenticty_token field.
- Remove the disabled attribute.
- Allow the flow to continue.
- Verify the course was created.
- Delete the course.

Change-Id: I371a96ceb68fe9dbd19f05d7b5538def5b1b0510
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/307805
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Robin Kuss <rkuss@instructure.com>
QA-Review: Robin Kuss <rkuss@instructure.com>
Product-Review: Eric Saupe <eric.saupe@instructure.com>