Correct mangled try/catch block
This mangled try/catch block (no exception object specified to catch) causes the embedded chat to fail to load (as well as cause a number of issues for those of us who jump on the coat-tails of its iframe-resizing code in instutional learning content!). The simplest of fixes is proposed. Also see issue 2131.