add guard to activate_role for unique constraint (account_id, name)
we had a case where a role for a specific account was created and
later deactivated. a new role with identical name was created in
its place, and the old role was attempted to be re-activated.
this violated a uniqueness constraint on [:account_id, :name]
for active roles.

we now guard against re-activating roles when a currently active
role exists in the same context with the same name.

fixes FOO-3315
flag = none

test plan:
 • within an account create a custom role
 • de-activate that custom role
 • create a new role with that same name
 • attempt to re-activate the former role
 • verify helpful error reponse as to why that action will not
   be permitted with an http code 400

Change-Id: I03143696d94210e6379195cb6d42ea8e48891c13
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/309457
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Charley Kline <ckline@instructure.com>
Product-Review: Charley Kline <ckline@instructure.com>
QA-Review: August Thornton <august@instructure.com>