Refresh the en.json fallback translations
Refs FOO-3069
flag=none

We've got a new locale-sensitive URL which added a line
to the static en.json file, so it gets dirty every time
translations are built. The "correct" version needs to
be on master to avoid dirtying local dev repos.

To make this easier to review, the actual change in en.json
was the addition of the line:

  "community.basics_find_url" : "https://community.canvaslms.com/t5/Canvas-Basics-Guide/Where-do-I-find-my-institution-s-URL-to-access-Canvas/ta-p/82",

That can be a little hard to pick out from the smashed-
together one line version.

Test plan:
* run script/canvas_update
* your master branch should still be clean

Change-Id: I233b91432e8a7e89c3b3a118472fc2907306071e
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/310065
Reviewed-by: August Thornton <august@instructure.com>
QA-Review: Charley Kline <ckline@instructure.com>
Product-Review: Charley Kline <ckline@instructure.com>
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>