LtiDeepLinkingResponse can indicate that link open in new tab
closes gh-2157
closes INTEROP-7994

Canvas defaults to using an iframe (new_tab: 0) for ltiResourceLink
content items. This requires that a tool provider use the Canvas HTTP
API to update each module item or assignment to specify that it open the
LTI link in a new tab.

Since Canvas doesn't currently take the `window.targetName` into account
when handling ltiResourceLink content items, it seemed OK to handle a
value of "_blank" as a special case for indicating that the content
should open in a new tab.

This would not affect any existing content and it's unlikely (?) to
surprise any users who have set a targetName of "_blank". In fact,
it may now do what they've intended it to do.

Test Plan:
  - Configure a course to launch an LTI Advantage Deep-Linking workflow
  - The tool should return a Deep-Linking response with at least one
    ltiResourceLink content item whose window.targetName value is set to
    "_blank"
  - Once the item populates (module item or assignment), ensure it opens
    in a new tab (note: Canvas will require you to click the "open in
    new tab" button manually to launch the tool, this seems strange but
    is how it's behaved for many years, see
    522abc26d42d843d3d4fda5cd58866e48a386361)

Change-Id: Ie966cd54d622c6b46138a3ddf3f0ebf1a10f1576
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/312816
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Tucker Mcknight <tmcknight@instructure.com>
QA-Review: Xander Moffatt <xmoffatt@instructure.com>
Product-Review: Alexis Nast <alexis.nast@instructure.com>