LtiDeepLinkingResponse can indicate that link open in new tab
Canvas defaults to using an iframe (new_tab: 0) for ltiResourceLink content items.
This requires that a tool provider use the Canvas HTTP API to update
each module item or assignment to specify that it open the LTI link in a new tab.

Since Canvas doesn't currently take the `window.targetName` into account
when handling ltiResourceLink content items, it seemed OK to handle a
value of "_blank" as a special case for indicating that the content
should open in a new tab.

This would not affect any existing content and it's unlikely (?) to
surprise any users who have set a targetName of "_blank". In fact,
it may now do what they've intended it to do.

Test Plan:
  - Configure a course to launch an LTI Advantage Deep-Linking workflow
  - The tool should return a Deep-Linking response with at least one ltiResourceLink
    content item whose window.targetName value is set to "_blank"
  - Once the item populates (module item or assignment), ensure it opens in a new tab