block logins for pseudonyms that are in a suspended state
we offer JIT provisioning for a multitude of auth providers. I'm
specifically targeting SAML as that's where most of our reports
are coming from. If in the future we want to update more, we can
in regards to recognizing and dealing with pseudonyms that have
been suspended

this change gives us the flexibility to include pseudonyms
involved in our SSO authentication flow that have been suspended,
bypass the logic for JIT provisioning (thus not duplicating users
and setting federated attributes that don't pertain to the user
in question) and block the login as no user found with pseudonym;
our fallback when pseudonym == nil

fixes FOO-2717
flag = none

test plan:
 • set up a SAML IDP to interface with Canvas as a SP
   and enabled JIT provisioning with the sis_user_id
   attribute being collected
 • ensure your IDP is sending the federated attribute
   that you have set in the auth provider config for
   sis_user_id
 • have a user associated with the SAML auth provider
 • suspend the user
 • attempt to log in as that user via /login/saml and
   authenticate
 • verify the login is rejected with an error stating
   a user was not found with provided unique id
 • verify no additional user was provisioned through
   JIT

Change-Id: I129ba60a86f90c98416e60b5c76325561e2bbad3
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/310836
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Jeremy Stanley <jeremy@instructure.com>
QA-Review: August Thornton <august@instructure.com>
Product-Review: Jesse Poulos <jpoulos@instructure.com>