remove unnecessary branding fetchMock assertion
refs LS-3831
flag=none

The fetchMock library provides a lot of valuable inspection methods
when mocking network requests. The 'done' method will warn the user via
the console if the mocked route was not called. Given the intention of
this test - as well as the applicable assertion the line after - I opted
to simply remove the reference to 'done' as it wasn't providing us with
anything that 'called' isn't already.

If we opted to keep both assertion in, we would either need to look at
keeping the global supression in (not ideal) or mocking the console warn
method (doesn't seem relevant to the nature of the test).

test plan:
  - tests pass

qa risk: low

Change-Id: Ifeaff92fcb770c9ed85a9ee5d49e3d61e997d718
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/312012
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Charley Kline <ckline@instructure.com>
QA-Review: Davis Hyer <dhyer@instructure.com>
Product-Review: Davis Hyer <dhyer@instructure.com>