provide correct default prop for isCompletelyEmpty
flag=none

The previous default referred to the prop type of bool rather than a
bool value. We are setting the default to true as a function is a truthy
value so this technically is a no-op for the value itself.

Other changes on the patchset are related to linters needing to update
small things. I opted to keep them in the same patch as they are very
minor changes.

test plan:
  - tests pass

qa risk: low

Change-Id: I1172f556eb536077886ca17ff705f0ea31716fb1
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/312780
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Jackson Howe <jackson.howe@instructure.com>
QA-Review: Jackson Howe <jackson.howe@instructure.com>
Product-Review: Davis Hyer <dhyer@instructure.com>