Don't cache autograde permission
If a tool tries to grade an unpublished assignment, it will get an
error, as intended. However, the :autograde permission is cached, so
then the assignment is published, the permission will linger for an
hour.

I tested whether this happens in the UI with teachers grading, and it
turns out it doesn't, because when an assignment is published, it
appears that the teachers' updated_at gets bumped when an assignment is
published. As a result, the user's `cache_key` will be different and the
:grade permission will not be cached.

Caching the permission doesn't really save much anyway, as it is only
checked when the grade changes and grader_id < 0 (graded by tool)

Another effect of this code is that the full error message will always
be shown, unlike before, where if the permission was cached to false,
the error message would just show "Cannot be changed at this time:"
and leave out the reason ("This assignment is still unpublished"), as
`@grading_error_message` is set in can_autograde?, which is not run if
the permission is cached.

fixes INTEROP-7943
flag=none

Test plan:
- Create an unpublished LTI 1.3 assignment in a course with at least one
  student
- Try to update the score with AGS. Example:
    tok post canvas.docker/api/lti/courses/62/line_items/291/score \
      @local_grade_update.json
  where local_grade_update.json is (e.g.):
    {
      "activityProgress": "Submitted",
      "gradingProgress": "FullyGraded",
      "scoreGiven": 0.99,
      "scoreMaximum": 1.0,
      "timestamp": "2023-03-09T16:47:04.640Z",
      "userId": 5,
      "https://canvas.instructure.com/lti/submission": {
        "new_submission": false,
        "timestamp": "2023-03-09T15:20:25.091Z"
      }
    }
- This should fail with an error about the course not being published
- Make the AGS request again. It should fail with the same error message
  (saying the request is not published)
- Publish the assignment
- Make the AGS request again. This time it should succeed.

Change-Id: I107b3755f75dfa8ca40fe22923a929f8ce5559b9
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/313283
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Tucker Mcknight <tmcknight@instructure.com>
QA-Review: Tucker Mcknight <tmcknight@instructure.com>
Product-Review: Alexis Nast <alexis.nast@instructure.com>