User Observer process should restore inactive observer enrollments
It is common for schools to run sis imports in batch mode using the batch_mode_enrollment_drop_status parameter set to inactive. This changes the default behavior of marking data previously imported via sis that is not represented in the latest sis import as deleted to inactive. The User Observer process is currently unable to handle restoring observer enrollments that have been marked inactive.

 Test Plan:
   - Add a user as an observer of a student with active enrollments
   - Mark one or more of the observer enrollments associated to the student as inactive
   - Process a user_observer sis import tying the observer to the student
   - Confirm the observer enrollments marked as inactive are restored