Add use_classic_font? helper
Adds a use_classic_font? helper that we'll use in a later commit to
determine which font k5 users should see. This returns true for any
user who's associated with a k5 account with classic font selection
(and any subaccount of those accounts) as long as the k5_font_selection
flag is enabled. "Associated with" means that the user has an
enrollment in a course in the account or an account_user in the
account. If an observer has selected a student in the picker, then
only the student's courses where the observer is observing the student
are considered for the purposes of association.

closes LS-3830
flag = k5_font_selection
[pin-commit-multiple_root_accounts=8fb9989a007a61d5f8ad660ed73e4f3cd1c96640]

Test plan:
Part 1: regression test k5_user?
 - Login as a k5 student and ensure that k5 mode still appears on the
   dashboard and in k5 courses
 - Login as an observer of both k5 and classic students and ensure that
   the correct dashboard style appears depending on the selected
   student

Part 2: test use_classic_font?
 - Enable k5 - balsamiq in an account
 - Login as a student enrolled in a course in that account
 - In a rails console, expect
   `K5::UserService.new(user, root_account, nil).use_classic_font?`
   to be false
 - Enable k5 - lato in that account
 - Run the command again and expect it to return true
 - Remove the user's enrollments, enroll them in a subaccount of
   the k5 - lato account
 - Expect the command to return true

Change-Id: I15f7562c2e069918e79cd1cade0120fdeab8fa99
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/313995
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Robin Kuss <rkuss@instructure.com>
QA-Review: Robin Kuss <rkuss@instructure.com>
Product-Review: Jackson Howe <jackson.howe@instructure.com>