Remove added space between LTI icon and label
Rendering external_tools/helpers/icon includes the new lines and spaces
that are present in the partial. Calling #squish on the output of that
partial in this helper will remove all the new lines and spaces but this
removes the html_safe flag on the out. Wrapping that entire thing in
the icon correctly and without spaces.

fixes LS-3926
flag=none

test plan:
- Create an LTI in a course with an icon
- Go to the course settings and verify the LTI is rendered with the icon
- Hover over the button and verify there isn't an extra space before
the label
- Inspect the HTML and verify there isn't an extra space or new line
between the icon and label

Change-Id: I954b65d5a6e2d2e2a6354235679d12023f15240c
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/316073
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Jackson Howe <jackson.howe@instructure.com>
QA-Review: Jackson Howe <jackson.howe@instructure.com>
Product-Review: Eric Saupe <eric.saupe@instructure.com>