Don't cache translated string in course settings
Constants in an AR model get loaded when the model is loaded, including
translated strings. If the user switches languages, the cached string
might be stale. Thus, use a lambda in the constant that calculates the
translated setting name instead of storing the translated value. An
alternative would have been to make this entire constant a method
instead of a constant, but this is a less intrusive solution since it
has a number of usages.

flag = none
closes LS-3947

Test plan:
 - Change user's language to something else (in the UI)
 - Go to course settings
 - Scroll down to visibility
 - Check "customize"
 - Expect "syllabus" and "files" to appear in the currently selected
   language

Change-Id: Ifcfb95476f1e8d4f97007bfad1ccea9b0e55aa07
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/316412
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Robin Kuss <rkuss@instructure.com>
QA-Review: Robin Kuss <rkuss@instructure.com>
Product-Review: Jackson Howe <jackson.howe@instructure.com>