Remove page refresh when job is stopped
This removes the window.location.reload() that was being called when
the progress is stopped as it is no longer needed with the state fetch
that happens when the job ends for whatever reason.

This patchset also includes a bit of rearranging of the
ContextModulesPublishMenu.tsx that the TypeScript linter was really mad
about and does not affect the functionality.

fixes LS-3959
flag=module_publish_menu

test plan:
- It's easiest to add a sleep(10) at the bottom of the loop in
app/models/course.rb#batch_update_context_modules to stop the progress
- Have a course with a few modules and module items in them
- Start a Publish All job for the course
- Once some progress has been made click Stop
- Verify the modal closes and statuses are updated for what was completed
and the page does not refresh

Change-Id: I7121fb1355af752836e7c2a5c7771265cb9da55e
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/316670
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Jackson Howe <jackson.howe@instructure.com>
QA-Review: Jackson Howe <jackson.howe@instructure.com>
Product-Review: Jackson Howe <jackson.howe@instructure.com>