add ConnectedFriendlyDatetimes component
An accessibility concern popped up in voiceover for availability dates
when rendered as two components. The space that was part of the prefix
of the "until" clause was ignored, leading to the output of
"11:00amuntil" being read. To work around this, a new component is
created in this patchset that is a single rendered sentence that should
make screenreaders do the right thing.

flag=none

fixes EVAL-3079

test plan:
 - Ensure Assignment Enhancements feature is on
 - Create an assignment that has both an unlock/available dates and a
   lock/until date.
 - On a mac, turn on voiceover
 - As a student on the assignment page, keyboard navigate to the
   availability date.
 - Ensure the date is read correctly with a seperation between the first
   time and the "until" clause.

Change-Id: I8adf7e9a429c1b63d413b96213c7a8829b931537
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/319750
Reviewed-by: Cameron Ray <cameron.ray@instructure.com>
Reviewed-by: Kai Bjorkman <kbjorkman@instructure.com>
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
QA-Review: Kai Bjorkman <kbjorkman@instructure.com>
Product-Review: Sam Garza <sam.garza@instructure.com>