Keep expanded accts expanded after applying changes
This PS looks bigger than it is.
When an account is first rendered expanded, which happens if you
Apply Changes while it's open, we have to detect that we haven't
fetched its sub-accounts. This is the subtantive change.
It also requires the mock data be correct
and it wasn't. That accounts for most of the spec changes.

I removed @ts-nocheck from a few files and fixed typescript errors
(though AccountCalendarItemToggleGroup.tsx turned out to be
more than I wanted to tackle).

closes LF-302
flag=none

test plan:
  - have a bunch of accounts and sub-accounts
  - go to /accounts/:account_id/calendar_settings
  > expect the top account to be expanded, others collapsed
  - click the twisty to expand a sub-account
  - make a change and click Apply Changes
  > expect the sub-account(s) to remain expanded
  - make a change to the root account
  - click the twisty to close the root account
  - click Apply Changes
  > expect the root accoutn to re-open after the update

Change-Id: Id056b6cb12442cf1ad20a5589f1b916c0ffc1607
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/320006
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Robin Kuss <rkuss@instructure.com>
QA-Review: Robin Kuss <rkuss@instructure.com>
Product-Review: Ed Schiebel <eschiebel@instructure.com>