Use :active_media_object_by_media_id scope for adding media files
This fixes false positives on instituitions that have a deleted
MediaObject with a media_id of "maybe". This allows attachments to
process only if there is an active MediaObject with the same id.

This patchset will not resolve the stuck attachments but will fix
attachments going forward. A separate patchset with a datafix is required.

refs LF-393
flag=none

test plan:
- Create a MediaObject that is deleted and has the media_id of "maybe"
- Create an Attachment and make sure it has the media_entry_id of "maybe"
- Call #ensure_media_object on the attachment
- Verify the attachment data is updated successfully

Change-Id: Ibffa0cf7e6df7fb48388a78bee88037b3ffae377
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/321488
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Luis Oliveira <luis.oliveira@instructure.com>
QA-Review: Luis Oliveira <luis.oliveira@instructure.com>
Product-Review: Eric Saupe <eric.saupe@instructure.com>