add position to sis_pseudonym for a more consistent sort
Provide an additional sort key/order for the case where multiple
pseudonyms for a given user have a matching unique_id, and
happen to be in the same workflow state with an included
sis_user_id. This will add a level of uniqueness to the sort
whether by Ruby or ActiveRecord in hopes for consistency and
deterministic behavior.

fixes FOO-3539
flag = none

Test plan --
Prerequisites:
1. An account with two authentication providers.
2. A course in the account

Step 1:
Use a users.csv and a logins.csv SIS import to create a user with two
logins that have the same login id (one for each auth provider)
and different SIS ids. This can also be done in Rails console.

Step 2:
Enroll the user in the course via the UI or API
(don’t add them using a SIS import)

Step 3:
Check the value of sis_user_id at /api/v1/users/:user_id for the user
created in step 1. If it matches the SIS id on the login which appears
first on the user page (the one with lesser id), make a change to the
second login and save then repeat this step. Note: you may have to
juggle the positions of the pseudonyms. If it matches the SIS id on
the login which appears second on the user page. Proceed to step 4

Step 4:
Run a provisioning report for enrollments in the term the course is in.
Note that the sis_user_id in the provisioning report should match the
sis_user_id for the user on the People page in the course or at
/api/v1/course/:course_id/enrollments

Change-Id: Id339d4751da03f2b1a32902fec9ff29159f0909a
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/322234
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Jeremy Stanley <jeremy@instructure.com>
Reviewed-by: Michael Hulse <michael.hulse@instructure.com>
QA-Review: August Thornton <august@instructure.com>
Product-Review: August Thornton <august@instructure.com>