Hack to prefer TII 1.1 tool in turnitin_report
When launching a turnitin report from the `report_url` in the
`turnitin_data` in a submission, we only have the URL to go from, which
(since it doesn't match a tool exactly) really means looking up a tool
by domain. Unfortunately, TII has multiple tools with that domain,
including a 1.3 tool. Since we prefer 1.3 tools, we will find the 1.3
tool, but in this plagiarism flow, TII only uses a 1.1 tool. So instead,
we need to look for a 1.1 tool in this particular case.

Hopefully we'll eventually deprecate 1.1 and move all plagiarism stuff
to 1.3, and do away with all this code...

closes INTEROP-8135
refs INTEROP-8072

Test plan:
- Create an LTI 1.3 tool and LTI 1.1 tool with the same domain. I
  actually made them have the same URL too (I copied the LTI 1.1 tool's
  URL and domain) and made that the URL used below in the turnitin_data
  too.
- Create an assignment that can handle a file upload submission and make
  a submission on a student, which will make an attachment.
- On the submission, update `turnitin_data` to be an object with a key
  of 'attachment_1234' (where 1234 is the Attachment id) and an object
  with status and report_url. Example:
    submission.turnitin_data = {
      my_attachment.asset_string => {
        status: 'scored',
        report_url: my_lti_11_tool.url,
        similarity_score: 73,
        state: :problem
      }
    }
    submission.save
- Check that before this commit, going to speedgrader and clicking the
  plagiarism score icon launches the LTI 1.3 tool
- Check that after this commit, it launches the LTI 1.1 tool.
- Chcek that after this commit, launching the URL with the retrieve
  endpoint (without the 'prefer_1_1') launches the LTI 1.3 tool.

Change-Id: If9f4d53f01936c15f35b6c51b0dba062c14b27b3
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/322730
Reviewed-by: Tucker Mcknight <tmcknight@instructure.com>
QA-Review: Tucker Mcknight <tmcknight@instructure.com>
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Product-Review: Alexis Nast <alexis.nast@instructure.com>