Disables the expand button for single line descriptions
closes OUT-4513
flag=none

This patchset will disable the expand
button when the description is only a
single line for html-text and text type
descriptions. Html descriptions with
extra formatting will always be expandable
however.

This patchset accounts for when
the description is resized by the window
or by the center divider or even when the
description is changed because of the
addition of the TruncateText component and
its onUpdate property. A custom handler is
passed to the onUpdate parameter, and that handler
checks whether the description is truncated or not.

So if the description is truncated, the expand button
is disabled. Else, the expand button can be clicked
on.

Test Plan
- Create a new outcome and edit the description from
the html editor. Make sure this description is only
a few words so the description is not truncated by
ellipses (...).
    - The description is classified as 'html-text'.
    - Verify the description is NOT expandable if
      truncated.
    - Shrink the size of the window until the description
      is truncated and then verify that the description
      is expandable.

- Create a new outcome and edit the description from
the html editor. Make this description multiple lines
at least.
    - The description is classified as 'html-text'.
    - Verify the description is expandable.

- Create a new outcome using an imported standard that
has a single line description.
    - The description is classified as 'text'.
    - Verify the description is NOT expandable
      if truncated.
    - Shrink the size of the window until the
      description is truncated and then verify
      the description is expandable.

- Create a new outcome using an imported standard that
has a longer description.
    - The description is classified as 'text'.
    - Verify the description is expandable.

- Create a new outcome and edit the description from
the html editor. Make sure the description has extra
html formatting features such as a table.
    - The description is classified as 'html'.
    - Verify that the description is expandable.
    - Note that extra html formatting is not rendered
      until the description is expanded.

Change-Id: Ia2612ec66943d8e75b47977c7fc4d328300ede82
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/323033
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Jason Anderson <jason.anderson@instructure.com>
QA-Review: Jason Anderson <jason.anderson@instructure.com>
Product-Review: Kyle Rosenbaum <krosenbaum@instructure.com>