add temp enrollment pairing to enrollments api create
This allows us to create a new temporary enrollment / association for a
known user that is modeled after an existing enrollment. We track the
temporary_enrollment_source_user_id field to know which user currently
has a temporary enrollment and what user the enrollment is based on.

This isn't a novel concept, as it just uses existing logic to create
a new base type enrollment given an existing role. The idea is that we
can update enrollment state with the start_at and end_at dates to
start and conclude the temporary enrollment.

closes FOO-3681
flag=temporary_enrollments

Test plan:
 • Enable the temporary enrollments feature flag
 • Exercise the enrollments API endpoint:
    • /doc/api/enrollments.html#method.enrollments_api.create
 • See associated JIRA for more details regarding request
   specifications and assumptions

Change-Id: Idc34947b43e95ab10f5b00fe8d47bfc7f5233eea
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/324211
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Maya Tyner <maya.tyner@instructure.com>
QA-Review: August Thornton <august@instructure.com>
Product-Review: August Thornton <august@instructure.com>