Guarantee freq. picker is updated when event date is changed
This commit fixes the bug described in the ticket, plus
resurrects code that disables the freq picker if it is
not given a date. This code was once there but was lost
somewhere along the way and is needed to handle bad
date input.

closes LF-558
flag=calendar_series

test plan:
  - create an event and click more options
  - open the freq picker. notice it is using the correct date
  - change the date by editing the text (not by using the calendar popup)
  - open the freq picker.
  > expect it to be using the new date
  - repeat using the calendar popup
  - repeat selecting a new date from the browser's popup list of
    autocompletions

  - delete all text from the date input
  > expect the freq picker to be "Does not repeat" and disabled

Change-Id: Ibb85ae6068ba9f95a99065d840ef787985394bc4
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/324415
Reviewed-by: Robin Kuss <rkuss@instructure.com>
QA-Review: Robin Kuss <rkuss@instructure.com>
Product-Review: Ed Schiebel <eschiebel@instructure.com>
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>