Remove filterUselessConsoleMessages from canvas-rce
The RCE had a dependency on @instructure/filter-console-messages, but
this package is not published to NPM. It works in Canvas since the
package is in canvas-lms:packages/, but not when used as a
dependency in other apps. Since there's only a few console warnings
that this package was filtering, and the RCE already has a console
filter mechanism set up, we'll remove the RCE's dependence on this
package and move those suppressions into jest-setup.js to group all
warning suppressions together. The console noise should remain the
same.

closes LF-704
flag = none

Test plan:
 - specs pass

Change-Id: If3541a88d141736a1383d600c39259f49adb2350
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/326560
Reviewed-by: Jake Oeding <jake.oeding@instructure.com>
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
QA-Review: Jackson Howe <jackson.howe@instructure.com>
Product-Review: Jackson Howe <jackson.howe@instructure.com>