fix double scrolling issue on assign to tab
InstUI does not provide a way to handle the scrollbar behavior
on the Select component, this PS fixes the problem by finding the
scrollable wrapper and adding it the "overscroll-behavior-y: contain"
attribute. To do this, the way of showing/hiding options needs to be
managed by the selector since the listRef is removed every time the
Select hides the options.

fixes LF-799
flag= differentiated_modules

Test plan:

1-In a course with several sections, students and modules
2- Go to the course modules page.
3- Open the module menu.
4- Click on the “Assign to” option.
5- Click on the “Custom Access” option.
6- Scroll over the dropdown options.
7- Expect the scroll to only affect the "Assign to" dropdown
scrollbar and the external scrollbar not to move

Change-Id: Ia8c5313748e684b5a7d9704b7523f9d395bd765a
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/328754
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Product-Review: Jonathan Guardado <jonathan.guardado@instructure.com>
Reviewed-by: Robin Kuss <rkuss@instructure.com>
QA-Review: Robin Kuss <rkuss@instructure.com>