error display for infinite loop observers issue
this is a a change to this
https://gerrit.instructure.com/c/canvas-lms/+/331612
https://instructure.atlassian.net/browse/FOO-3764
this change is to make sure the user knows what error they are
into when they are observing an observer observing themselves
as an alert. prior to this warmfix, they would just get an
error that request did not complete.

closes FOO-3764
flag=none

test plan:
- create two observer enrollments for two students
- link the second observer to the first one
 - now the first user is observing the second
- try to link the first observer to the second one
 - you should get an alert that says "Cannot observe user with
   another user that is being observed by the current user."

Change-Id: Ide4da5d953c01695f2161d5bb67177a02552a2b4
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/332525
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: August Thornton <august@instructure.com>
Product-Review: Jesse Poulos <jpoulos@instructure.com>
QA-Review: Ajmal Esa <ajmal.esa@instructure.com>