Cancel orphaned Progresses
When calling progress.process_job with a singleton and on_conflict:
:overwrite, inst-jobs may not queue a new job and instead update an
existing queued job with the same singleton. If this happens, the
job's handler will now point to the new Progress object and the old
Progress will no longer be updated when the job completes. It will be
left as "queued" indefinitely, and enough of these queued Progresses
can be created on a context to cause performance issues. So, when
calling process_job, detect if a job was updated (not inserted), and
then cancel any other queued Progress objects with the same job id.

Note: if a job is enqueued with a singleton and on_conflict:
:use_earliest, inst-jobs may update an existing job if it encounters
a conflict (and job.enqueue_result will be :updated), but it will
not update the job's handler, so the job will still point at the old
Progress. This commit does not address that use case.

closes LF-1026
flag = none

Test plan:
 - In a rails console, find a course with a course pace
 - Stop delayed_job
 - Run `progresses = []`
 - Run `progresses << pace.create_publish_progress` three times in a
   row
 - Expect the first 2 progresses to be canceled, and the third to be
   queued

Change-Id: I571397e93771fa469aa79f03f8710f66e607b453
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/333550
Reviewed-by: Robin Kuss <rkuss@instructure.com>
Reviewed-by: Jacob Burroughs <jburroughs@instructure.com>
QA-Review: Robin Kuss <rkuss@instructure.com>
Product-Review: Jackson Howe <jackson.howe@instructure.com>
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>