include delete_me_frd param for delete from all accounts
When the “Delete from ALL accounts” button is used in the UI on the
user’s details page, we issue a AJAX “DELETE” request to the users’s
show API endpoint. We’ll now pass delete_me_frd: true along for the
ride so we do the needful in regards to including this user on the
scrubber cleanup list.

Also updated verbiage in the UI to be more clear about what this
button does.

closes FOO-4174
flag = none

Test plan:
* Create a user
* Delete the user from all accounts via the “Delete from ALL accounts”
  button on the user's details page
* Verify that the user is not only soft deleted, but also has
  delete_me_frd: true set on their user record

Change-Id: Icefbc314be7f29f78e07efe5f96693716cfbfe68
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/336609
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Jeremy Stanley <jeremy@instructure.com>
QA-Review: August Thornton <august@instructure.com>
Product-Review: August Thornton <august@instructure.com>