Sort sectionNames var expansion by section id
The variable expansion for com.instructure.User.sectionNames
has no explicit order. While this is fine when the variable is
used in isolation, it becomes problematic when an LTI tool needs
to construct full knowledge of a course's sections using a
combination of com.instructure.User.sectionNames and
Canvas.course.sectionIds. Since the IDs come back in ascending
order, it makes sense to sort the names by the same property,
guaranteeing that names can be mapped accurately onto IDs,
and enabling the tool to accurately reconstruct the sections
for a given course using only the NRPS membership response.