don't unnecessarily copy files before uploading
Rack::Test::UploadedFile, in addition to looking like something
that isn't intended to be used in production code, creates a copy
of the file that is passed in. This slows down things like
reports and exports, and isn't needed for our use cases,
which all involve synchronous uploads (and do not involve #append).

a note about unlinking: in a few places, the code calls #unlink
on the Rack::Test::UploadedFile instance, which (1) did not
delete the file passed into that thing, only the copy it made;
and (2) was unnecessary since that class deleted its copy itself

also note that Rack::Test::UploadedFile is still being used
to simulate uploaded files in tests... as is intended.
but no longer using it just to wrap a `content_type` around
a file we are uploading from the backend

test plan:
 - smoke test the following:
  - content export
  - account report
  - sis import
  - submission export

flag=none
closes FOO-4215

Change-Id: Ife83373be5920958eb7bdf14c2da1a56708ca11e
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/338855
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: August Thornton <august@instructure.com>
QA-Review: Jeremy Stanley <jeremy@instructure.com>
Product-Review: Jeremy Stanley <jeremy@instructure.com>