Only look for help link when page fully loads
Fixes FOO-4247
flag=none

We had a bit of a flake because the front-end boot code in
ui/index.ts would almost immediately look for the existence
of a link classed `help_trigger_dialog` which appears only
on the login screen. But sometimes when that code executes,
nothing with that class is yet in the DOM. It's kind of weird
because that page is rendered by an ERB so it should be there
statically. But there may be some race condition between the
browser's page painting and the execution of that code.

Solution is to just look for that trigger after the document
comes fully ready.

Test plan:
* Load the login page
* The "Help" link should ALWAYS bring up the login page's
  help dialog. Previously, sometimes it would but sometimes
  it wouldn't.

Change-Id: I67f25ddc2d6ab8984df66400bc55438e9eed93c5
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/339151
Reviewed-by: Aaron Shafovaloff <ashafovaloff@instructure.com>
QA-Review: Charley Kline <ckline@instructure.com>
Product-Review: Charley Kline <ckline@instructure.com>
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>