fix rubric builder footer layout
this commit fixes the layout of the rubric builder footer. The footer
was previously stuck to the bottom of the screen with position absolute
and there was some overlapping between the criteria section. This commit
fixes the layout and the footer should always be stuck to the bottom of
the page while the criteria section should be scrollable.

fixes EVAL-3956
flag=enhanced_rubrics

test plan:
- open the rubric builder form page.
- verify that the footer is stuck to the bottom of the page
- verify that the criteria section is scrollable when there are enough
  criteria to fill the page
- verify that there is no overlapping between the footer and the
  criteria section

Change-Id: I00d6612ed94fa7875669881537fc02c6b477e577
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/340193
Reviewed-by: Derek Williams <derek.williams@instructure.com>
Reviewed-by: Kai Bjorkman <kbjorkman@instructure.com>
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
QA-Review: Kai Bjorkman <kbjorkman@instructure.com>
Product-Review: Christopher Soto <christopher.soto@instructure.com>