add rubric rating points reorder logic
this adds existing logic for rubrics when adding points or modifying
points of ratings. when adding a new rating, we will add the points of
that rating to be in between the points of the previous and next rating.
when modifying the points of a rating, we will reorder the points of the
ratings of the ratings to be in descending order while not modifying the
name or description of the rating.

closes EVAL-3962
flag=enhanced_rubrics

test plan:
- navigate to the rubric builder page for a rubric.
- open the criterion modal.
- add a rating at position 0
- verify that the points of the rating is +1 the previous top rating
- add a rating to the bottom of the list
- verify that the points of the rating is -1 the previous bottom rating
  with a minimum of 0
- add a rating in between two ratings.
- verify that the points of the rating is between the points of the
  previous and next rating. if there is only a difference of 1 between
  the previous and next rating, the points of the new rating should be
  set to the top rating it is set in between.
- modify the points of a rating to be higher than the previous rating
- verify that the ratings are reshuffled but the name and description
  of the ratings is not modified.

Change-Id: I50e1c60da3a3d790d995ed8f54136d8a4213ecd6
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/340624
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Derek Williams <derek.williams@instructure.com>
Reviewed-by: Kai Bjorkman <kbjorkman@instructure.com>
QA-Review: Christopher Soto <christopher.soto@instructure.com>
Product-Review: Ravi Koll <ravi.koll@instructure.com>