Fix issue where date cannot be parsed
In the date picker, the date is not being parsed correctly in Japanese
leading to an error where the date is cleared when the user tries to
save. Adding a space to the date format allows the date to be parsed.

Fixes FOO-4257
flag=none

Test plan:
- Create a quiz and set the show/hide dates. See that you can save the
  dates and when you edit again, the date is shown with out an error.
- When you save the quiz without changing the date, the date should not
  be cleared.

Change-Id: Ie403ad84c32bd884b0da66112533868981905bc2
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/342355
Reviewed-by: August Thornton <august@instructure.com>
Product-Review: Jason Perry <jason.perry@instructure.com>
QA-Review: Jason Perry <jason.perry@instructure.com>
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>