send stats for CAS specific auth timeouts
This change adds a new metric to track the number of times a CAS
specific auth timeout error/cutoff occurs. This is useful for tracking
the frequency of these timeouts and for identifying any patterns that
may be causing them.

We include CAS specific tags to help identify the service host that is
timing out, as well as the service ticket involved or any related
error counts as it relates to a TimeoutCutoff

We currently have an existing metric for tracking general auth timeouts,
but it is not specific to CAS. This metric gets rolled up via:
Canvas::Errors.capture_exception(:service_timeout)

refs FOO-4325
flag = none

test plan:
 - auth.timeout_error should be incremented
   when a service times out under Canvas.timeout_protection
 - auth.timeout_cutoff should be incremented
   when a service has timed out repeatedly and is not being
   invoked for a time period

Change-Id: Ie50c8c51d5c37b0fd6eb095026ff0e084af60723
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/342682
Reviewed-by: Alex Slaughter <aslaughter@instructure.com>
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
QA-Review: Alex Slaughter <aslaughter@instructure.com>
Product-Review: Alex Slaughter <aslaughter@instructure.com>