Allow non relative video src in ImageEmbedOptions
A bug occured in prod where instructors were not able to
upload captions to a video they had not uploaded themselves.
This was due to incorrect parsing of the src attribute when
the src was not a relative url, this made it so the request
for media tracks did not use the attachment id. We decided to
allow for parsing of non-relative urls to amend this.

fixes LF-1341
flag=none

test plan:
- have a course with 2 teachers in it
- with teacher one, go to an RCE and upload a
  video. Save the video.
- act as teacher 2 and go to edit the page
- In local, the video should save as a relative url. Go back
  into the html of the video and make it an absolute url
- ex. http://canvas-web.inseng.test/media_attachments_iframe/38?....
- continue editing, and attempt to add captions to the video
- save the captions and expect to see no errors and the captions
  save as normal

Change-Id: I1b9e90f14db737f28842b91a17c740bfac29cbdf
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/343683
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Robin Kuss <rkuss@instructure.com>
QA-Review: Robin Kuss <rkuss@instructure.com>
Product-Review: Sarah Gerard <sarah.gerard@instructure.com>