Fix matery paths support for continuous grade changing
At the time of override assigning and unnassign students due
to mastery path progression we were checking agains the latest
assignment actions to only attemp unnassigning from things the
student was indeed assigned to. The criteria for "latest" was
faulty and would give false negatives after multiple regrades.
That was changed and now we fetch latest based on updated_at
instead of created_at, and if an old set action is used we touch
it so it will be seen as one of the "latest".

fixes LF-1351
flag=none

Test plan:
- Create a mastery path with 2 diverging paths
- Have a student and a submission to test path progression
- Regrade the student grade the student back and forth between
  paths at least 4 times
- Check he is only assigned to the proper path and not multiple

Change-Id: I777354ec597f0483b0d1af2d54c54fa5cb2004ec
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/344003
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Jacob DeWar <jacob.dewar@instructure.com>
QA-Review: Jacob DeWar <jacob.dewar@instructure.com>
Product-Review: Luis Oliveira <luis.oliveira@instructure.com>