Update learning_update_result updated & created live event data
user_uuid, associated_asset, and artifact information is not
available on the live event.  In order for the rubric assessment
audit to be able to successfully audit LOR and AR this information
needs to be added.

title is also being removed as this data contains the student's
first and last name which is considered PII and should not be
passed in live event data. A ticket has been created to remove
the student's name from this column.  See OUT-6302.

closes OUT-6307

flag=none

Test plan:
- Tests are sufficient and Jenkins passes

Change-Id: Ic7f18f9cb3292234b8215e1babf329453be6f1e8
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/344151
Reviewed-by: Angela Gomba <angela.gomba@instructure.com>
QA-Review: Kyle Rosenbaum <krosenbaum@instructure.com>
Product-Review: Kyle Rosenbaum <krosenbaum@instructure.com>
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>