Check for section visibilities in date_details api
Because ungraded discussion topics were previously
able to be differentiated by section, we need to ensure
that those sections are present in the assign to tray.
Note that the test plan uses the 'Post to' section
of the discussion edit page, which will be eventually
removed when the flag is on. This functionality is so
far only for the tray outside of the discussions edit
page.

closes LF-1493
flag=differentiated_modules

test plan:
- create an ungraded discussion and use the 'Post to'
  drop down to assign it to a certain section.
- save, and then open the assign to tray for that
  discussion
- expect to see the section assigned in the tray. Do not save
  the tray yet
- add the ungraded discussion to a module
- assign that module to the same section that the you assigned to
  the discussion
- open the tray again for the discussion, expect to still only see
  the section override (should not say inherited from module)
- save the tray, try making edits to the overrides, expect
  the tray to behave as normal.

Change-Id: I8d3967ffc18237b78821c541ee09cd3958830ab5
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/345626
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Jackson Howe <jackson.howe@instructure.com>
QA-Review: Jason Gillett <jason.gillett@instructure.com>
Product-Review: Sarah Gerard <sarah.gerard@instructure.com>