edit flash message for observers
This changes the flash message to be more informative when an observer
is currently not observing any students. It was incorrectly directing
the user back to the dashboard even though in this state they are not
linked to any students yet.

closes EVAL-4075
flag=assignments_2_student

Test Plan:
- Create a new observer
- enroll them in a course, but do not link them
to any students
- visit the any assignment page
- verify the edited flash message is displayed

Change-Id: Ie3dc7354886a35b745afc772e9cba056851785e9
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/345607
Reviewed-by: Derek Williams <derek.williams@instructure.com>
Reviewed-by: Spencer Olson <solson@instructure.com>
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
QA-Review: Spencer Olson <solson@instructure.com>
Product-Review: Ravi Koll <ravi.koll@instructure.com>