fix classic quizzes group expand/collapse
Due to the jQuery upgrade, we overlooked the testing
of the expand and collapse functionality on classic
quizzes.

From what I could determine, there was a change in
jQuery 3.0 relating to .next(); in older jQuery versions,
.next() would ignore hidden elements. In 3.0, .next()
stopped filtering out hidden elements. In this case, the
.insufficient_count_warning element was being selected
and this broke the logic surrounding the expand/collapse.

This commit updates the logic to select siblings up to
the footer of the question group. Not that, in the DOM, the
group elements are a flat/linear, non-nested, structure;
because they are all siblings, I had to use the footer to
act as a “bumper” for the nextUntil() selector function.

closes FOO-4429
flag=none

test plan:
- login to Canvas as a teacher
- navigate to a course
- click on “Quizzes” in the sidebar nav
- click on the “Questions” tab
- create several groups with children questions
- create a group without a question (this will then dispaly
  the insufficient_count_warning)
- for each group, toggle the vertically-pointing arrow
  found in the groups’s header bar
- toggle all groups on and off
- drag/drop groups and toggle the expand/collapse
- drag/drop questions withing groups and toggle the
  expand/collapse
- expect expand/collapse to show/hide questions in groups

Change-Id: I064ab55e86fd1479728b0ed3afe2a6aeef3f06e5
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/345642
QA-Review: Michael Hulse <michael.hulse@instructure.com>
Product-Review: Michael Hulse <michael.hulse@instructure.com>
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Charley Kline <ckline@instructure.com>