remove max visible course templates limit
This change removes the Setting and the limit on the query for the max
visible course templates in the settings page. We will now show all the
course templates in the settings page as a stop-gap measure until we
have a better solution such as a search / fuzzy matcher on course
template name.

closes FOO-4425
flag = none

test plan:
 * configure one or 200+ courses as course templates
 * in the root account, you can choose the template, or no template,
   and save
 * in a sub account, you can choose a template, no template, or inherit
 * in the settings page, you should see all the course templates
   regardless of the number

Change-Id: I4f0d3185b1a26d938f2485d813cc261818aa2839
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/345722
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Ajmal Esa <ajmal.esa@instructure.com>
QA-Review: August Thornton <august@instructure.com>
Product-Review: August Thornton <august@instructure.com>