Remove the mastery paths checkbox
When differentiated_modules is enabled, don't show the mastery paths
checkbox on the page edit page, and disable the API which creates an
assignment when that param is included in wiki_pages_api#create/update.

Instead, if the user includes the "Mastery Paths" pill on the tray,
then create the associated assignment if not present, which makes the
page show up in the conditional release editor. The API expects all
existing overrides to be included in the request, so other overrides
will be created on the assignment at the same time. The assignment will
NOT be removed if the Mastery Paths override is deleted - thus we don't
need to copy overrides from the assignment back to the page.

Page renames/ publish & unpublish will be synced to the associated
assignment via after_save callback on WikiPage.

flag = differentiated_modules
closes LX-1515

Test plan:
 - Disable mastery paths in a course
 - Go to a page; expect neither the checkbox to appear on the edit
   page nor the Mastery Paths option to appear in the tray
 - Add some overrides, save; expect an assignment to NOT be created on
   the page

 - Enable mastery paths in a course
 - Go to a page, add some overrides, save; expect an assignment to NOT
   be created
 - Add the Mastery Paths override with some other overrides, save;
   expect an assignment to get created, and for the assignment overrides
   to be saved on the assignment (i.e., persist in the tray)
 - Go to an assignment, edit, click the "Mastery Paths" tab; expect
   the page to appear in the add modal
 - Update the page's name, check the mastery paths modal again, expect
   the updated name to appear

 - Disable differentiated_modules
 - Regression check that the checkbox still appears on the page edit
   page and that checking it creates the associated assignment

Change-Id: I875c1eca71f336386130dbd5dc210b141da82155
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/347692
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Sarah Gerard <sarah.gerard@instructure.com>
QA-Review: Sarah Gerard <sarah.gerard@instructure.com>
Product-Review: Jackson Howe <jackson.howe@instructure.com>