Change Assign To link to button on show pages
On the assignment and quiz show pages, the "Assign To" button was an
`a` element instead of a button, which is not appropriate for buttons
that cause in-page actions as opposed to link to another page.

The button on other show pages and on all edit pages was already a
button element.

closes LX-1683
flag = differentiated_modules

Test plan (automated):
 - Visit assignment show page
 - Inspect the Assign To button
 - Expect it to be a button element
 - Repeat for quiz show page

Change-Id: Ie3a98f064fd534b160eec7888691890654b14f00
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/347849
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Robin Kuss <rkuss@instructure.com>
QA-Review: Robin Kuss <rkuss@instructure.com>
Product-Review: Jackson Howe <jackson.howe@instructure.com>