ensure enrollment state presence
Sometimes in the process of ensuring current enrollment state we end up
with a stale object. To avoid caching the stale object we set the
enrollment's enrollment_state association to nil and then reload the
enrollment state object and try again. If the enrollment state is still
stale it's possible we end up with a nil enrollment state. in that case
we can look the record up in the db and set the enrollment state
association to the correct value.

closes FOO-4440
flag = none

test plan:
- specs validate the behavior

Change-Id: I1b621f985ba575606be83c80f2c9c4eb87421354
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/348340
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Michael Hulse <michael.hulse@instructure.com>
QA-Review: August Thornton <august@instructure.com>
Product-Review: August Thornton <august@instructure.com>