base letter grade off scaled score
closes EVAL-4248
flag=none

Test Plan:
- Create and enable a points based grading scheme in the course with
letter grades between 0-10 (A is 10 to 9, B is 9 to 7, C+ is 7 to 5,
C- is 5 to 3, and D is 3 to 0)
- Create two assignments worth 50 points.
- Once the assignments are created, grade the student using two decimal
points towards the upper range of a letter grade. In my example, I
graded the assignments 44.99 / 50.
- Once graded, navigate to the student grades page and notice
that the total score appears to be rounding to 9,
which should be a letter score of A and an A is displayed

Change-Id: Ic9cf5c5a9825803c56518c5d37c759d7fcc6e9e9
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/349434
Reviewed-by: Spencer Olson <solson@instructure.com>
QA-Review: Spencer Olson <solson@instructure.com>
Product-Review: Cameron Ray <cameron.ray@instructure.com>
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>