Allow things that use TopNavPortal to modify breadcrumb
Refs FOO-4473
flag=none

Some feature pages need to alter the trailing breadcrumb text
to reflect things like the current Tab, etc. The current
InstUI TopNavPortal statically reads the breadcrumbs from
ENV so there's no way to make them dynamic.

This commit adds a prop to TopNav (and thus TopNavPortal)
called getBreadCrumbSetter which is a function provided by
the caller of TopNavPortal. This function, if provided,
will be called once with an argument that is object
containing two functions: `getCrumbs` and `setCrumbs`. The
former returns the current breadcrumbs, and the latter
sets the breadcrumbs. If the setter is called with a
Crumb object, that will just replace the last crumb in
the list. If it's called with an array of Crumb objects,
it will replace all the crumbs.

Test plan:
* Try out the new prop and make sure the functions work
* The automated test for this component currently only
* checks that the prop callback returns a properly-formatted
  object and that the getCrumbs function works

Change-Id: If63d9a9cf53037c9890218f574c68f6d88bc9e0f
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/349503
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Martin Ochoa <martin.ochoa@instructure.com>
Reviewed-by: Michael Hulse <michael.hulse@instructure.com>
QA-Review: Martin Ochoa <martin.ochoa@instructure.com>
Product-Review: Martin Ochoa <martin.ochoa@instructure.com>