Paginate the lti_registrations endpoint
We've decided to show 15 registrations per page. If a "page=<number>" query
parameter is not given, we show the first 15 registrations.

This is using the "Link" HTTP header standard for telling the caller which page they are on
and what the URL for the next and previous pages are.

The Link header contains a string that looks like the following:

```
<http://www.example.com/api/foo?page=1&per_page=15>; rel=\"current\",
<http://www.example.com/api/foo?page=1&per_page=15>; rel=\"first\",
<http://www.example.com/api/foo?page=2&per_page=15>; rel=\"next\""
<http://www.example.com/api/foo?page=3&per_page=15>; rel=\"last\""
```

which would tell the caller that:

1. There are three pages of results
2. They are being shown the first page
3. The next page is at example.com/api/foo?page=2
4. The last page is at example.com/api/foo?page=3

fixes INTEROP-8619

flag = lti_registrations_page

test plan:
- with the lti_registrations_page feature flag enabled,
- create more than 15 lti registrations (rails console is probably your best bet here)
- go to /api/v1/accounts/<account_id>/lti_registrations
- see that only 15 are on the page
- add ?page=2 to the end of the URL
- see that the remaining registrations are there (if you created fewer than 31, that is)

Change-Id: If93257c66d13b0c8a558b9c745ea1a0430014f34
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/349594
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Xander Moffatt <xmoffatt@instructure.com>
QA-Review: Xander Moffatt <xmoffatt@instructure.com>
Product-Review: Alexis Nast <alexis.nast@instructure.com>