Don't enumerate page ids in WikiPage#visible_to_user
The WikiPageVisibilityService needs both a user id and either course
ids or page ids to run performantly. Previously, we were providing
the ids of the pages currently selected by the scope, but the way this
scope is used, the scope was targeting almost all pages on the shard.
Thus, we'd be enumerating quite a few page ids on larger shards.

Fortunately, this scope is only used in 1 place (the planner), so
instead of providing user_id+wiki_page_ids to the service, we'll
have the planner pass the course_ids it knows the user is interested
in and the scope will pass those to the visibility service.

closes LX-1824
flag = selective_release_backend

Test plan:
 - Create some pages with todo dates in the next week
 - Assign some of those pages to a student
 - Visit the planner as a student and expect to see only
   the pages they've been assigned

Change-Id: I500ab52959dbc28764d84026d8732f2706ad8888
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/349987
Reviewed-by: Sarah Gerard <sarah.gerard@instructure.com>
QA-Review: Sarah Gerard <sarah.gerard@instructure.com>
Product-Review: Jackson Howe <jackson.howe@instructure.com>
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>