Limit accepted query params on lti_registrations endpoint
The page parameter should be a number or empty; return a 422 status
if it is not.

The dir parameter should be asc, dsc, or empty; return a 422 status
if it is not.

The sort parameter should be one of the valid options; return a 422
status if it is not.

flag = lti_registrations_page

test plan:
- specs pass

refs INTEROP-8616

Change-Id: I35007d0ca04b2c6cd24a50e909b7b964d1fb4605
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/350963
Reviewed-by: Xander Moffatt <xmoffatt@instructure.com>
QA-Review: Xander Moffatt <xmoffatt@instructure.com>
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Product-Review: Tucker Mcknight <tmcknight@instructure.com>