Take AssignTo content out of tray in edit pages
This commit is to create a new wrapper for the assign to
section of the edit pages so that it is no longer in a tray.
This will be behind a feature flag and will replace the existing
wrapper DifferentiatedModulesSection. This new wrapper is
essentially the same as DifferentiatedModulesSections, but removes
logic specific to the tray, and adjusts the save logic to be a useEffect
This commit only ensures that the correct overrides appear on load
of the edit pages and save functionality. validations on save, groups,
and sync to SIS will all be in a separate ticket. It is also only
available for quiz and assignment edit pages at the moment.

closes LX-1873

flag=selective_release_edit_page

Test plan:
- turn on the new selective_release_edit_page flag
- have assignments and quizzes with various types of overrides
  include module overrides, course overrides, section, adhoc, etc
- open the edit page of the assignments and ensure the correct cards
  appear. Ensure the Assign to section matches the design of the
  ticket.
- You should be able to add/remove assignees, dates, etc and they should
  save as expected
- validations within the cards (such as you must select one assignee,
  date validations, etc) should be functional, but will not cause error
  on save
- Assign to section should be disabled if course pacing is on
- important dates should still be functional
- blueprint locking should still be functional

Change-Id: I5069e341defaebfbe3d7098aa004f9f9baa41805
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/351608
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Product-Review: Sarah Gerard <sarah.gerard@instructure.com>
Reviewed-by: Juan Chavez <juan.chavez@instructure.com>
Reviewed-by: Jonathan Guardado <jonathan.guardado@instructure.com>
QA-Review: Juan Chavez <juan.chavez@instructure.com>