submission_type_selection: don't require a title
We were using the presence of a title in the deep linking respones to
determine whether a resource had been selected. Not sending a title
would cause some weird behavior, such as not showing a resource even and
not allowing a save if require_resource_selection is given in the
placement config) even though a URL and other properties had been set.

In this commit we check the presence of `url`, which unlike `title`, is
required according to the LTI 1.3 spec.

closes INTEROP-8700
flag=none

Test plan:
- have a tool with submission_type_selection placement and
  require_resource_selection: true in the placement config
- create a new assignment and choose the tool in the submission type
  selection dropdown. it should fail with the error.
- choose a resource in the tool and send back either `title: null` or no
  title at all in the response (to test no title at all with the LTI 1.3
  test tool, add `.except('title')` to `Models::ContentItem#as_json`).
  The card should still be shown, but "Unnamed Document" will be shown
  instead of a resource title. Saving should succeed.
- edit the assignment and check that the card is still shown with no
  title, and that you can save without any changes.
- edit the assignment, change the type to `External Tool` and back to
  the specific tool. No resource should be shown and saving should fail.
- add back a resource and make sure saves succeeds
- add a resource with a title and make sure the title shows up in the
  card and saving succeeds

Change-Id: Ie32e906e01b81a21bdd185031b567604fa9c7563
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/351727
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Steven McGee <steve.mcgee@instructure.com>
QA-Review: Steven McGee <steve.mcgee@instructure.com>
Product-Review: Alexis Nast <alexis.nast@instructure.com>