Make visibleTextContent a global property on DOM nodes
Refs FOO-4582
flag=none

Michael wrote a clever little function that works like
Node.textContent to get the inner text of a DOM node,
but excludes anything that is display=none, so as not
to confuse the returned value if it has, say, some
screenreader-only stuff inside it.

I thought it would be generally useful so this commit
arranges to make it globally available on any DOM node
inside Canvas. The addition of the new property on
Node happens during UI boot time so it can just be used
naturally.

Test plan:
* automated tests pass
* you agree this is a nifty and useful addition

Change-Id: I5246c5c545d4bdca3cd344ae5c41a5b5e71998bc
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/353077
QA-Review: Charley Kline <ckline@instructure.com>
Product-Review: Charley Kline <ckline@instructure.com>
Reviewed-by: Michael Hulse <michael.hulse@instructure.com>
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>