Remove default @unauthorized_message
This was causing the if statement to always return at the first check
and the subsequent cases were not being called. This lead to the
send_broken_content! not sending messages. The default was added as part
of MAT-1313.

fixes RCX-1834
flag=none

test plan:
- Have a teacher with Content Link Error notification settings set to
Notify Immediately
- Create a page with links to content that is not accessible by a student.
For example, an unpublished assignment.
- Publish the page
- As a student in the course, click the link in the published page that
should go to the unpublished assignment.
- Verify the unauthorized page includes the text "We've let your
instructor know to review this link as soon as possible."
- Go to /users/TEACHER_ID/messages and verify a message is there calling
out the Content Link Error.

Change-Id: If88c7248e04d5ce1fd22fd264f31b5b85e50d54d
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/353134
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Julio Alcantara <julio.alcantara@instructure.com>
Product-Review: Julio Alcantara <julio.alcantara@instructure.com>
QA-Review: Mysti Lilla <mysti@instructure.com>