Fix N+1 in activity stream summaries for courses
flag=dashboard_graphql_integration

closes LX-1901

This creates a batch loader that calculates
stream summaries for multiple contexts,
with functions based on the existing code
in API stream_item.

A conscious choice was made to only allow this
for courses. If a similar pattern was needed
for Users or Groups, it could be expanded to
allow for batching of these types as well.

Test Plan:
- New specs pass
- Enable dashboard_graphql_integration FF
- Enable N+1 Detection
- Hit the following graphiql query
  and notice that there are no N+1s

query MyQuery {
  legacyNode(_id: "1", type: User) {
    ... on User {
      favoriteCoursesConnection {
        nodes {
          _id
          activityStream {
            summary {
              count
              notificationCategory
              type
              unreadCount
            }
          }
        }
      }
    }
  }
}

Change-Id: Id9a89c473abd77941933f402f8813acf6c6d2672
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/353136
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Drake Harper <drake.harper@instructure.com>
QA-Review: Drake Harper <drake.harper@instructure.com>
Product-Review: Drake Harper <drake.harper@instructure.com>