Prefer innerText to the bespoke visibleTextContent
Refs FOO-4582
flag=none

It was pointed out that while the previous addition of
`visibleTextContent` is a very useful idea, there's
already an existing DOM method `.innerText` which has
equivalent functionality, so we don't really need the
new property.

One problem is that jsdom, used in Jest tests, does not
implement innerText, because jsdom does not implement
a layout engine. So the custom implementation that was
previously merged needs to stay, but now has been re-
imagined as just a polyfill for `innerText` itself.

This is a win because the native DOM classes are no
longer messed with unless necessary.

Test plan:
* automated tests pass
* you like this gentler approach better

Change-Id: I9d1e7a56be26ba843053931ffb0ff56b24930ae4
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/353545
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
QA-Review: Charley Kline <ckline@instructure.com>
Reviewed-by: Michael Hulse <michael.hulse@instructure.com>
Reviewed-by: Aaron Shafovaloff <ashafovaloff@instructure.com>
Product-Review: Aaron Shafovaloff <ashafovaloff@instructure.com>