Temporary fix for dynamic import of locale files
Fixes FOO-4657
flag=none

The rspack upgrade changes how pathnames are mapped
for dynamic imports, which broke the language polyfill
code. I wasn't able to get all the way to the bottom
of what's going on; there is probably a better fix for
this (or maybe not!), but this makes the import more
explicit in a way that rspack clearly understands.

Will probably need to revisit this issue to clean up
the code and research exactly what is going. In the
mean time, this will do.

Test plan:
* Be in Google Chrome
* Set your user's language to Íslenska
* Reload the page with the browser console open
* You should see successful "polyfilled is for locale is"
  and not any "Cannot find module" errors

Change-Id: I6c42d3f80da01576d939f5d3b39f042ca1ef594c
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/355131
Reviewed-by: Dustin Cowles <dustin.cowles@instructure.com>
QA-Review: Charley Kline <ckline@instructure.com>
Product-Review: Charley Kline <ckline@instructure.com>
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>