set default CSP logging directives
This change sets the default CSP logging directives to log violations
to the console and our current logging pipeline through the
Content-Security-Policy-Report-Only header. This is a temporary
change to help with debugging and will be transitioned over to using
the Content-Security-Policy header in the future once we are confident
that the policy is correct.

The proper implementation of this change also requires updating the
global Setting "csp.global_whitelist" to allow the necessary domains
that we use in our application.

closes FOO-4813
flag = default_source_csp_logging

Test Plan:
1. CSP Report-Only Header is Set for Frames
2. CSP Report-Only Header is Set for Files
3. CSP Report-Only Header is Now Set for Non-HTML Content Types
4. CSP Report-Only Header is Not Set if default_source_csp_logging
   Feature is Disabled
5. CSP Report-Only Header is Not Set if javascript_csp Feature is
   Disabled
6. CSP Report-Only Header is Not Set if CSP is Enforced
7. CSP Report-Only Header is Not Overridden if Already Set

Change-Id: I4ea921797043147d2a6bcdc0ae5cf87cf8615da9
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/358598
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Jeremy Stanley <jeremy@instructure.com>
Reviewed-by: Cody Cutrer <cody@instructure.com>
QA-Review: August Thornton <august@instructure.com>
Product-Review: August Thornton <august@instructure.com>