Do not schedule lock based on lock_at
For some reason which remains unknown, there is a scheduled job
to set the DiscussionTopic.locked attribute at the lock_at date.
This seems unnecessary since the UI and the backend prevents
students from answering when the lock_at date is in the past.

However setting the locked to true causes big problems with the
selective_release feature which can be used to define overrides
for different users or sections. If the locked is true it is
preferred to the overrides and it can make that not usable.

fixes: VICE-4695
flag=selective_release_backend
flag=selective_release_ui_api

Test plan:
  - Create a discussion with the above feature flags enabled
  - Set Available until date X for Section A which User A is assigned to
  - Set Available until date Y for Everyone Else which is before X
  - Notice that before Y, both User A and another user can reply to Discussion
  - Notice that between Y and X only User A is able to reply to Discussion
  - Notice that after X, no one is able to reply to Discussion

Change-Id: I3b2f98519fc59b56e366cc5eb2912b075c440536
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/358662
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
QA-Review: Ádám Molnár <adam.molnar@instructure.com>
Reviewed-by: Ádám Molnár <adam.molnar@instructure.com>
Product-Review: Tamás Balogh <tamas.balogh@instructure.com>
Reviewed-by: Daniel Matyas Vincze <daniel.vincze@instructure.com>