fix z-index for rubric overlay
this commit fixes a bug that was introduced when the z-index of the ui
dialog was changed to 10000. this caused unintended issues when the
enhanced_rubrics FF was enabled and the new assignments ff was disabled.

closes EVAL-4688
flag=enhanced_rubrics

test plan:
- run `yarn run build:css` to update styles
- create a new assignment and attach a rubric
- add a new rubric criterion
- click on the edit icon for the existing criterion ratings
- verify that the overlay is displayed correctly and there are no
  z-index issues

Change-Id: Ibd184e8bd11b963f652a89c90e92016fb807b598
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/358706
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Spencer Olson <solson@instructure.com>
QA-Review: Chris Soto <christopher.soto@instructure.com>
Product-Review: Ravi Koll <ravi.koll@instructure.com>