Fix UTID overwrite to NULL when copying course
When copying a course with an external tool,
the unified_tool_id sometimes was updated to NULL, as
the exported course data does not hold the UTID information.
This only happened for imports when the tool was already in the
destination course, because of update_unified_tool_id after_save hook.
The hook prevented the UTID from being updated to NULL on first copies.
It is only triggered when other tool data is changed during update
like (tool_id, url, settings, etc.) but for secondary copy
only the utid was changed so it was updated to null.

fixes INTEROP-8888
flag=none

test plan:
- have course A with an external_tool with UTID
- have a course B
- copy course A to course B including the tool
    - e.g. course B settings Import Course Content
- for initial copy the UTID should be copied to the cet record
- do the copy again
- after this change the UTID should still be there in the cet record

Change-Id: I7584beb63b06c613b22100a756149db4ceda1663
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/360091
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Csaba Csuzdi <csaba.csuzdi@instructure.com>
QA-Review: Csaba Csuzdi <csaba.csuzdi@instructure.com>
Product-Review: Alexis Nast <alexis.nast@instructure.com>