set grading standard id to 0 on conclusion
When an assignment is using the canvas default grading scheme. The
assignment grading_standard_id is set to nil. We do not want this
inherited value to change after conclusion. This commit sets the
grading_standard_id to 0 when the course is concluded in that case, so
in case an account default is set after the course was concluded, the
assignment will not be affected.

flag=default_account_grading_scheme
refs EVAL-4734

Test Plan:
Prerequisites: Have a course with students where have been graded. Make
sure that the course has at least one assignment set to display as a
letter grade using the “Canvas Default Grading Scheme” but the course
settings does not have the “Enable course grading scheme” setting
enabled. The account should not have a default set yet.
- Navigate to the course settings and conclude the course
- Go to the root account settings and create a grading scheme and set
the new scheme as the default
- Navigate back to the course you concluded. Notice that the total
column and the assignment were not affected by the default Grading
scheme change

Change-Id: I236cb66290336b8d3a09f4fea46f7825872fa568
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/360097
QA-Review: Sleyder Zuleta <sleyder.zuleta@instructure.com>
Reviewed-by: Sleyder Zuleta <sleyder.zuleta@instructure.com>
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Product-Review: Cameron Ray <cameron.ray@instructure.com>