find replacement attachment by name beyond update threshold
previously, canvas would update all previous versions of a file
each time it is overwritten, but in g/359636 we capped this at
50 to avoid unbounded writes initiated by pathological user
behavior, which recently caused a database incident. this means
later in the chain of overwrites, a find could return a newer
but still deleted version.

this commit slightly tweaks the logic to allow the fallback that
finds a non-deleted file with the same name as the original to
come into play in this use case, limiting the potential impact
to cases with pathological numbers of overwrites *and* renames.

test plan:
 - overwrite a file like a hundred times
 - look up /courses/X/files/<id of the 60th or so overwrite>
 - it should return the latest version

flag=none
refs FOO-4841

Change-Id: I3a97a6d7de0c9f5874da0b920680400b2dc453f6
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/360474
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Mysti Lilla <mysti@instructure.com>
QA-Review: Jeremy Stanley <jeremy@instructure.com>
Product-Review: Jeremy Stanley <jeremy@instructure.com>