lint commits that touch the permissions registry
This commit simply provies an [INFO] message to the owner / uploader
of the commit that touches the permissions registry file. The intent
is to provide a reminder to the uploader to review the permissions
registry file and follow the instructions via the provided link. Which
should include consideration in adding an additional migration to
backfill those permission(s).

closes FOO-4890
flag = none

Test plan:
- Any commit that touches the permissions registry file should
  receive an [INFO] message to review the permissions registry file
  and follow the instructions via the provided link

Change-Id: Ie161ff3870090e241c1a4460494ea9d1ae64e780
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/362460
Reviewed-by: Maya Tyner <maya.tyner@instructure.com>
Reviewed-by: Jeremy Stanley <jeremy@instructure.com>
QA-Review: Jeremy Stanley <jeremy@instructure.com>
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Product-Review: August Thornton <august@instructure.com>