(B) QTY-6959: Ignore deleted enrollments when updating `last_activity_at``
When we try to update the `last_activity_at` on enrollments, sometimes
we have two and one has `deleted` activity_state. This is causing a
deadlock with transactions that look something like this:

```rb
UPDATE "public"."enrollments"
SET "total_activity_time" = 26978,
    "last_activity_at"   = '2024-11-15 17:19:59.273329'
WHERE "enrollments"."course_id" = 33484
  AND "enrollments"."user_id" = 240618; /*controller:submissions_api,action:for_students,hostname:75a8ffb08c13,pid:25784,context_id:5bef41bb-6174-4b00-80a7-0e03a652d607,line:/config/initializers/active_record.rb:960:in `update_all'*/

UPDATE "public"."enrollments"
SET "total_activity_time" = 26978,
    "last_activity_at"    = '2024-11-15 17:19:59.363693'
WHERE "enrollments"."course_id" = 33484
  AND "enrollments"."user_id" = 240618; /*controller:context_modules,action:content_tag_assignment_data,hostname:dc763854c4a3,pid:26708,context_id:e0110d0c-d0f7-4ad6-ae28-bccd5949023b,line:/config/initializers/active_record.rb:960:in `update_all'*/
```

We are assuming for the case of a student enrollment, there should only
be one, per course, per student. But we have two in these examples, all
including `deleted` ones.

We may need to interogate the specs in the virtual box before we
continue with this change.

Co-authored-by: Mark Ng <Mark.ng@Strongmind.com>