Fix module index tray menu flashes on click
The module index menu tray uses the jquery-kyle-menu
there is an edge case with this popup when it is larger than the page,
in this exact case when too many external tools installed to that
placement the tray does not stay open but flashes
and closes immediately when clicked.
This fix has two parts.
1. Change the setting of kyle menu using the
jqueri-ui position component's `collision` parameter to `flip`
instead of `flipfit`. so that the tray does not cover
the menu button just open below.
2. Modify the already modified jquery popup lib's close mechanism
with a check and don't close if it is just opened.

fixes INTEROP-8948
flag=none

test plan:
- Go to a course's modules page
    - Add many (30) external tools to `module_index_menu` placement
    - Or just edit the html with adding multiple (at least 30) `li`
    elements to the `ul` container of the tool list popup (id=toolbar-1)
- Click on the module index menu (three dots) button
- Before this change you should see the tray menu closes immediately
- After this change the tray stays open when clicked

Change-Id: I1b31d7ea93ab7423dd0152063d9d0b5e14b488f5
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/363238
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Csaba Csuzdi <csaba.csuzdi@instructure.com>
QA-Review: Csaba Csuzdi <csaba.csuzdi@instructure.com>
Product-Review: Alexis Nast <alexis.nast@instructure.com>