STK-11345 Gather permissions conditions for logging
This bug has a permissions issue that when we attempt to debug later, no longer exists.

It's probable that this permission issues is caused by something like the state of the course or
the enrollment, and that it's changed later. We've added logging (which will appear in the cloudwatch logs
for the instance), which will tell us which rights were attempted to be gained, and which ones succeeded.

Co-authored-by: Tyler Hablutzel <tyler.hablutzel@strongmind.com>